Pos.Binary

Description

Reexports of Pos.Binary.* classes

Documentation

class Bi t where

Simplified definition of serializable object, Data.Binary.Class-alike.

Write instance Bi SomeType where without any method definitions if you want to use the Binary instance for your type.

Methods

put :: t -> Put

put :: Binary t => t -> Put

get :: Get t

get :: Binary t => Get t

encode :: Bi a => a -> LByteString

Encode a value to a lazy bytestring

encodeStrict :: Bi a => a -> ByteString

Encode a value to a strict bytestring. Use with caution, because converting to strict ByteString is expensive.

decode :: Bi a => LByteString -> a

Decode a value from a lazy ByteString, reconstructing the original structure.

decodeOrFail :: Bi a => LByteString -> Either (LByteString, ByteOffset, String) (LByteString, ByteOffset, a)

decodeFull :: Bi a => LByteString -> Either String a

Like decode, but ensures that the whole input has been consumed.

Different sizes for ints

newtype UnsignedVarInt a

Constructors

UnsignedVarInt

Fields

getUnsignedVarInt :: a

newtype SignedVarInt a

Constructors

SignedVarInt

Fields

getSignedVarInt :: a

newtype FixedSizeInt a

Constructors

FixedSizeInt

Fields

getFixedSizeInt :: a

getAsciiString1b :: String -> Word8 -> Get String

putAsciiString1b :: String -> Put