Pos.Block.Arbitrary

Documentation

newtype HeaderAndParams ssc

This type is used to generate a valid blockheader and associated header verification params. With regards to the block header function verifyHeader, the blockheaders that may be part of the verification parameters are guaranteed to be valid, as are the slot leaders and the current slot.

Constructors

HAndP

Fields

getHAndP :: (VerifyHeaderParams ssc, BlockHeader ssc)

newtype BlockHeaderList ssc

Constructors

BHL

Fields

getHeaderList :: ([BlockHeader ssc], [PublicKey])

newtype SmallTxPayload

Constructors

SmallTxPayload TxPayload

Orphan instances

Arbitrary TxPayload

Methods

arbitrary :: Gen TxPayload

shrink :: TxPayload -> [TxPayload]

Arbitrary GenesisExtraBodyData

Methods

arbitrary :: Gen GenesisExtraBodyData

shrink :: GenesisExtraBodyData -> [GenesisExtraBodyData]

Arbitrary GenesisExtraHeaderData

Methods

arbitrary :: Gen GenesisExtraHeaderData

shrink :: GenesisExtraHeaderData -> [GenesisExtraHeaderData]

Arbitrary MainExtraBodyData

Methods

arbitrary :: Gen MainExtraBodyData

shrink :: MainExtraBodyData -> [MainExtraBodyData]

Arbitrary MainExtraHeaderData

Methods

arbitrary :: Gen MainExtraHeaderData

shrink :: MainExtraHeaderData -> [MainExtraHeaderData]

Arbitrary MsgGetBlocks

Methods

arbitrary :: Gen MsgGetBlocks

shrink :: MsgGetBlocks -> [MsgGetBlocks]

Arbitrary MsgGetHeaders

Methods

arbitrary :: Gen MsgGetHeaders

shrink :: MsgGetHeaders -> [MsgGetHeaders]

Arbitrary (Body (GenesisBlockchain ssc))

Methods

arbitrary :: Gen (Body (GenesisBlockchain ssc))

shrink :: Body (GenesisBlockchain ssc) -> [Body (GenesisBlockchain ssc)]

Arbitrary (SscPayload ssc) => Arbitrary (Body (MainBlockchain ssc))

Methods

arbitrary :: Gen (Body (MainBlockchain ssc))

shrink :: Body (MainBlockchain ssc) -> [Body (MainBlockchain ssc)]

Arbitrary (ConsensusData (GenesisBlockchain ssc))

Methods

arbitrary :: Gen (ConsensusData (GenesisBlockchain ssc))

shrink :: ConsensusData (GenesisBlockchain ssc) -> [ConsensusData (GenesisBlockchain ssc)]

(Arbitrary (SscProof ssc), Bi Raw, Ssc ssc) => Arbitrary (ConsensusData (MainBlockchain ssc))

Methods

arbitrary :: Gen (ConsensusData (MainBlockchain ssc))

shrink :: ConsensusData (MainBlockchain ssc) -> [ConsensusData (MainBlockchain ssc)]

Arbitrary (BodyProof (GenesisBlockchain ssc))

Methods

arbitrary :: Gen (BodyProof (GenesisBlockchain ssc))

shrink :: BodyProof (GenesisBlockchain ssc) -> [BodyProof (GenesisBlockchain ssc)]

(Arbitrary (SscProof ssc), Bi Raw) => Arbitrary (BodyProof (MainBlockchain ssc))

Methods

arbitrary :: Gen (BodyProof (MainBlockchain ssc))

shrink :: BodyProof (MainBlockchain ssc) -> [BodyProof (MainBlockchain ssc)]

Arbitrary (GenericBlock (GenesisBlockchain ssc))

Methods

arbitrary :: Gen (GenericBlock (GenesisBlockchain ssc))

shrink :: GenericBlock (GenesisBlockchain ssc) -> [GenericBlock (GenesisBlockchain ssc)]

(Arbitrary (SscProof ssc), Arbitrary (SscPayloadDependsOnSlot ssc), SscHelpersClass ssc) => Arbitrary (GenericBlock (MainBlockchain ssc))

Methods

arbitrary :: Gen (GenericBlock (MainBlockchain ssc))

shrink :: GenericBlock (MainBlockchain ssc) -> [GenericBlock (MainBlockchain ssc)]

Arbitrary h => Arbitrary (Attributes h)

Methods

arbitrary :: Gen (Attributes h)

shrink :: Attributes h -> [Attributes h]

Arbitrary (GenesisBlockHeader ssc)

Methods

arbitrary :: Gen (GenesisBlockHeader ssc)

shrink :: GenesisBlockHeader ssc -> [GenesisBlockHeader ssc]

(Arbitrary (SscProof ssc), Bi Raw, Ssc ssc) => Arbitrary (MainBlockHeader ssc)

Methods

arbitrary :: Gen (MainBlockHeader ssc)

shrink :: MainBlockHeader ssc -> [MainBlockHeader ssc]

(Arbitrary (SscProof ssc), Bi Raw, Ssc ssc) => Arbitrary (BlockSignature ssc)

Methods

arbitrary :: Gen (BlockSignature ssc)

shrink :: BlockSignature ssc -> [BlockSignature ssc]

(Ssc ssc, Arbitrary (SscProof ssc)) => Arbitrary (MainToSign ssc)

Methods

arbitrary :: Gen (MainToSign ssc)

shrink :: MainToSign ssc -> [MainToSign ssc]

(Arbitrary (SscProof ssc), Arbitrary (SscPayloadDependsOnSlot ssc), SscHelpersClass ssc) => Arbitrary (MsgBlock ssc)

Methods

arbitrary :: Gen (MsgBlock ssc)

shrink :: MsgBlock ssc -> [MsgBlock ssc]

(Arbitrary (SscProof ssc), Bi Raw, Ssc ssc) => Arbitrary (MsgHeaders ssc)

Methods

arbitrary :: Gen (MsgHeaders ssc)

shrink :: MsgHeaders ssc -> [MsgHeaders ssc]

BiSsc ssc => Buildable (BlockHeader ssc, PublicKey)

Methods

build :: (BlockHeader ssc, PublicKey) -> Builder