Pos.Block.Network

Documentation

announceBlock :: WorkMode ssc m => m (Set NodeId) -> SendActions m -> MainBlockHeader ssc -> m ()

announceBlockOuts :: OutSpecs

handleHeadersCommunication :: forall ssc m s. (WorkMode ssc m, Reifies s (Limit MsgGetHeaders)) => ConversationActions (MsgHeaders ssc) (LimitedLength s MsgGetHeaders) m -> Proxy s -> m ()

blockListeners :: (SscWorkersClass ssc, WorkMode ssc m) => m ([ListenerSpec m], OutSpecs)

blockStubListeners :: (SscHelpersClass ssc, WithLogger m) => Tagged ssc ([ListenerSpec m], OutSpecs)

needRecovery :: forall ssc m. (SscWorkersClass ssc, WorkMode ssc m) => Proxy ssc -> m Bool

triggerRecovery :: forall ssc m. (SscWorkersClass ssc, WorkMode ssc m) => m (Set NodeId) -> SendActions m -> m ()

Triggers recovery based on established communication.

requestTipOuts :: OutSpecs

requestTip :: forall ssc s m. (SscWorkersClass ssc, WorkMode ssc m) => Proxy s -> NodeId -> ConversationActions MsgGetHeaders (LimitedLength s (MsgHeaders ssc)) m -> m ()

Is used if we're recovering after offline and want to know what's current blockchain state. Sends "what's your current tip" request to everybody we know.

handleUnsolicitedHeaders :: forall ssc s m. (SscWorkersClass ssc, WorkMode ssc m) => NonEmpty (BlockHeader ssc) -> NodeId -> ConversationActions MsgGetHeaders (LimitedLength s (MsgHeaders ssc)) m -> m ()

mkHeadersRequest :: forall ssc m. WorkMode ssc m => Maybe HeaderHash -> m (Maybe MsgGetHeaders)

Make GetHeaders message using our main chain. This function chooses appropriate from hashes and puts them into GetHeaders message.

requestHeaders :: forall ssc s m. (SscWorkersClass ssc, WorkMode ssc m) => MsgGetHeaders -> NodeId -> Maybe (BlockHeader ssc) -> Proxy s -> ConversationActions MsgGetHeaders (LimitedLength s (MsgHeaders ssc)) m -> m ()

mkBlocksRequest :: HeaderHash -> HeaderHash -> MsgGetBlocks

Make message which requests chain of blocks which is based on our tip. LcaChild is the first block after LCA we don't know. WantedBlock is the newest one we want to get.

handleBlocks :: forall ssc m. (MonadDBCore m, SscWorkersClass ssc, WorkMode ssc m) => m (Set NodeId) -> NodeId -> OldestFirst NE (Block ssc) -> SendActions m -> m ()

retrievalWorker :: forall ssc m. (MonadDBCore m, SscWorkersClass ssc, WorkMode ssc m) => m (Set NodeId) -> (WorkerSpec m, OutSpecs)

data MsgGetHeaders

GetHeaders message. Behaviour of the response depends on particular combination of mghFrom and mghTo.

mghTo resolves to some header (let's call it top for convenience) -- node's tip if it's Nothing, header with hash in Just if it's Just.

If mghFrom is empty, then semantics is "request to return header of block top".

Otherwise (if mghFrom isn't empty) it represents the set of checkpoints. Responding node will try to iterate headers from top to older until it reaches any checkpoint. If it finds checkpoint c, it returns all headers in range [c.next..top]. If it doesn't find any checkpoint or depth of searching exceeds recoveryHeadersMessage, it will try to find the newest checkpoint cc from mghFrom that's in main chain of responding node and then return at most recoveryHeadersMessage headers starting with cc as the oldest one, returning headers in range l2 = [cc.next..x] where x is either top (in case length l2 < recoveryHeadersMessage) or some arbitrary header (and length is precisely recoveryHeadersMessage).

Constructors

MsgGetHeaders

Fields

mghFrom :: ![HeaderHash]

mghTo :: !(Maybe HeaderHash)

data MsgGetBlocks

GetBlocks message (see protocol specification).

Constructors

MsgGetBlocks

Fields

mgbFrom :: !HeaderHash

mgbTo :: !HeaderHash

newtype MsgHeaders ssc

Headers message (see protocol specification).

Constructors

MsgHeaders (NewestFirst NE (BlockHeader ssc))

newtype MsgBlock ssc

Block message (see protocol specification).

Constructors

MsgBlock (Block ssc)