Pos.Block.Pure

Description

Pure functions related to blocks and headers.

Documentation

blockDifficultyIncrement :: Block ssc -> ChainDifficulty

Difficulty of the Block, which is determined from header.

headerDifficultyIncrement :: BlockHeader ssc -> ChainDifficulty

Difficulty of the BlockHeader. 0 for genesis block, 1 for main block.

mkGenericBlock :: forall b. (HasHeaderHash (BBlockHeader b), Blockchain b, BHeaderHash b ~ HeaderHash) => Maybe (BBlockHeader b) -> Body b -> (BHeaderHash b -> BodyProof b -> ConsensusData b) -> ExtraHeaderData b -> ExtraBodyData b -> GenericBlock b

Smart constructor for GenericBlock. Uses mkGenericBlockHeader.

mkGenericHeader :: forall b. (HasHeaderHash (BBlockHeader b), Blockchain b, BHeaderHash b ~ HeaderHash) => Maybe (BBlockHeader b) -> Body b -> (BHeaderHash b -> BodyProof b -> ConsensusData b) -> ExtraHeaderData b -> GenericBlockHeader b

Smart constructor for GenericBlockHeader.

mkMainBlock :: (BiSsc ssc, SscHelpersClass ssc, MonadFail m) => Maybe (BlockHeader ssc) -> SlotId -> SecretKey -> Maybe ProxySKEither -> Body (MainBlockchain ssc) -> MainExtraHeaderData -> MainExtraBodyData -> m (MainBlock ssc)

Smart constructor for MainBlock. Uses mkMainHeader.

recreateMainBlock :: (BiSsc ssc, SscHelpersClass ssc, MonadFail m) => MainBlockHeader ssc -> Body (MainBlockchain ssc) -> MainExtraBodyData -> m (MainBlock ssc)

mkMainHeader :: (BiSsc ssc, SscHelpersClass ssc) => Maybe (BlockHeader ssc) -> SlotId -> SecretKey -> Maybe ProxySKEither -> Body (MainBlockchain ssc) -> MainExtraHeaderData -> MainBlockHeader ssc

Smart constructor for MainBlockHeader.

mkGenesisHeader :: BiSsc ssc => Maybe (BlockHeader ssc) -> EpochIndex -> Body (GenesisBlockchain ssc) -> GenesisBlockHeader ssc

Smart constructor for GenesisBlockHeader. Uses mkGenericHeader.

mkGenesisBlock :: BiSsc ssc => Maybe (BlockHeader ssc) -> EpochIndex -> SlotLeaders -> GenesisBlock ssc

Smart constructor for GenesisBlock. Uses mkGenesisHeader.

genesisHash :: Hash a

Predefined Hash of GenesisBlock.

data VerifyBlockParams ssc

Parameters of Block static verification. Note: to check that block references previous block and/or is referenced by next block, use header verification (via vbpVerifyHeader).

Constructors

VerifyBlockParams

Fields

vbpVerifyHeader :: !(Maybe (VerifyHeaderParams ssc))

Verifies header accordingly to params (verifyHeader)

vbpVerifyGeneric :: !Bool

Checks verifyGenesisBlock property.

vbpVerifySsc :: !Bool

Verifies ssc payload with sscVerifyPayload.

vbpVerifyProxySKs :: !Bool

Check that's number of sks is limited (1000 for now).

vbpMaxSize :: !(Maybe Byte)

Maximal block size.

vbpVerifyNoUnknown :: !Bool

Check that block has no unknown attributes.

data VerifyHeaderParams ssc

Extra data which may be used by verifyHeader function to do more checks.

Constructors

VerifyHeaderParams

Fields

vhpVerifyConsensus :: !Bool

vhpPrevHeader :: !(Maybe (BlockHeader ssc))

Nothing means that block is unknown, not genesis.

vhpNextHeader :: !(Maybe (BlockHeader ssc))

vhpCurrentSlot :: !(Maybe SlotId)

vhpLeaders :: !(Maybe SlotLeaders)

vhpMaxSize :: !(Maybe Byte)

vhpVerifyNoUnknown :: !Bool

Check that header has no unknown attributes.

verifyBlock :: (SscHelpersClass ssc, BiSsc ssc) => VerifyBlockParams ssc -> Block ssc -> VerificationRes

Check predicates defined by VerifyBlockParams. #verifyHeader #verifyGenericBlock

verifyBlocks :: forall ssc f t. (SscHelpersClass ssc, BiSsc ssc, t ~ OldestFirst f (Block ssc), NontrivialContainer t) => Maybe SlotId -> Bool -> BlockVersionData -> Maybe SlotLeaders -> OldestFirst f (Block ssc) -> VerificationRes

Verify a sequence of blocks.

foldl' is used here which eliminates laziness of triple. It doesn't affect laziness of VerificationRes which is good because laziness for this data type is crucial.

verifyGenericBlock :: forall b. Blockchain b => GenericBlock b -> VerificationRes

Perform cheap checks of GenericBlock, which can be done using only block itself. Checks which can be done using only header are ignored here. It is assumed that they will be done separately.

verifyHeader :: forall ssc. BiSsc ssc => VerifyHeaderParams ssc -> BlockHeader ssc -> VerificationRes

Check some predicates (determined by VerifyHeaderParams) about BlockHeader. #verifyConsensusLocal

verifyHeaders :: BiSsc ssc => Bool -> NewestFirst [] (BlockHeader ssc) -> VerificationRes

Verifies a set of block headers.