Pos.Client.Txp.History

Documentation

data TxHistoryEntry

Datatype for returning info about tx history

Constructors

THEntry

Fields

_thTxId :: !TxId

_thTx :: !Tx

_thIsOutput :: !Bool

_thInputs :: ![TxOut]

_thDifficulty :: !(Maybe ChainDifficulty)

thTxId :: Lens' TxHistoryEntry TxId

thTx :: Lens' TxHistoryEntry Tx

thIsOutput :: Lens' TxHistoryEntry Bool

thInputs :: Lens' TxHistoryEntry [TxOut]

thDifficulty :: Lens' TxHistoryEntry (Maybe ChainDifficulty)

data TxHistoryAnswer

Constructors

TxHistoryAnswer

Fields

taLastCachedHash :: HeaderHash

taCachedNum :: Int

taCachedUtxo :: Utxo

taHistory :: [TxHistoryEntry]

class Monad m => MonadTxHistory m where

A class which have methods to get transaction history

Methods

getTxHistory :: SscHelpersClass ssc => Tagged ssc (Address -> Maybe (HeaderHash, Utxo) -> m TxHistoryAnswer)

saveTx :: (TxId, TxAux) -> m ()

getTxHistory :: (SscHelpersClass ssc, MonadTrans t, MonadTxHistory m', t m' ~ m) => Tagged ssc (Address -> Maybe (HeaderHash, Utxo) -> m TxHistoryAnswer)

saveTx :: (MonadTrans t, MonadTxHistory m', t m' ~ m) => (TxId, TxAux) -> m ()

History derivation

getRelatedTxs :: Monad m => Address -> [(WithHash Tx, TxWitness, TxDistribution)] -> TxSelectorT m [TxHistoryEntry]

Select transactions related to given address. Bool indicates whether the transaction is outgoing (i. e. is sent from given address)

deriveAddrHistory :: Monad m => Address -> [Block ssc] -> TxSelectorT m [TxHistoryEntry]

Given a full blockchain, derive address history and Utxo TODO: Such functionality will still be useful for merging blockchains when wallet state is ready, but some metadata for Tx will be required.

deriveAddrHistoryPartial :: Monad m => [TxHistoryEntry] -> Address -> [Block ssc] -> TxSelectorT m [TxHistoryEntry]