Pos.Communication

Description

Re-exports of Pos.Communication.*

Documentation

data BiP

Constructors

BiP

sendTx :: MinWorkMode m => SendActions BiP m -> NetworkAddress -> TxAux -> m ()

Send Tx to given address.

sendVote :: MinWorkMode m => SendActions BiP m -> NetworkAddress -> UpdateVote -> m ()

sendUpdateProposal :: MinWorkMode m => SendActions BiP m -> NetworkAddress -> UpId -> UpdateProposal -> [UpdateVote] -> m ()

allListeners :: (SscListenersClass ssc, WorkMode ssc m) => [Listener BiP m]

All listeners running on one node.

allStubListeners :: (SscListenersClass ssc, WithLogger m) => Proxy ssc -> [Listener BiP m]

All listeners running on one node.

serverLoggerName :: LoggerName

Logger name for server.

sysStartReqListener :: MinWorkMode m => Timestamp -> Listener BiP m

Listener for SysStartRequest message.

sysStartRespListener :: MinWorkMode m => MVar Timestamp -> Listener BiP m

Listener for SysStartResponce message.

handleSysStartResp :: MinWorkMode m => MVar Timestamp -> NodeId -> SendActions BiP m -> SysStartResponse -> m ()

Messages and socket state

data VersionReq

Version request message. VersionResp is expected as response.

Constructors

VersionReq

data VersionResp

Version response (on VersionReq response).

Constructors

VersionResp

Fields

vRespMagic :: Int32

vRespBlockVersion :: BlockVersion

data PeerState ssc

PeerState type aggregates socket states needed for different parts of system.

peerVersion :: HasPeerState c ssc => Lens' c (Maybe BlockVersion)

data SysStartRequest

Communication request for system start.

Constructors

SysStartRequest

data SysStartResponse

Response to SysStartRequest.

Constructors

SysStartResponse !Timestamp

data TxMsgTag

Constructors

TxMsgTag

data TxMsgContents

Data message. Can be used to send one transaction per message.

Constructors

TxMsgContents

Fields

dmTx :: !Tx

dmWitness :: !TxWitness

dmDistribution :: !TxDistribution