Pos.Communication.Relay

Description

Framework for InvReqDat message handling

Documentation

class (Buildable tag, Buildable contents, Buildable key, Typeable tag, Typeable contents, Typeable key, Message (InvMsg key tag), Message (ReqMsg key tag), Message (DataMsg contents)) => Relay m tag key contents | tag -> contents, contents -> tag, contents -> key, tag -> key where

Typeclass for general InvReqDat framework. It describes monads, that store data described by tag, where "key" stands for node identifier.

Minimal complete definition

contentsToTag, contentsToKey, verifyInvTag, verifyReqTag, verifyDataContents, handleInv, handleReq, handleData

Methods

contentsToTag :: contents -> m tag

Converts data to tag. Tag returned in monad m for only type matching reason (multiparam type classes are tricky)

contentsToKey :: contents -> m key

Same for key. Sometime contents has key inside already, so it's redundant to double-pass it everywhere.

verifyInvTag :: tag -> m VerificationRes

verifyReqTag :: tag -> m VerificationRes

verifyDataContents :: contents -> m VerificationRes

handleInv :: tag -> key -> m Bool

Handle inv msg and return whether it's useful or not

handleReq :: tag -> key -> m (Maybe contents)

Handle req msg and return (Just data) in case requested data can be provided

handleData :: contents -> m Bool

Handle data msg and return True if message is to be propagated

data InvMsg key tag

Inventory message. Can be used to announce the fact that you have some data.

Constructors

InvMsg

Fields

imTag :: !tag

imKeys :: !(NonEmpty key)

data ReqMsg key tag

Request message. Can be used to request data (ideally data which was previously announced by inventory message).

Constructors

ReqMsg

Fields

rmTag :: !tag

rmKeys :: !(NonEmpty key)

data DataMsg contents

Data message. Can be used to send actual data.

Constructors

DataMsg

Fields

dmContents :: !contents

relayListeners :: (MonadDHT m, Bi (InvMsg key tag), Bi (ReqMsg key tag), Bi (DataMsg contents), Relay m tag key contents, WithLogger m, WorkMode ssc m, Bi NOP, Message NOP) => RelayProxy key tag contents -> ([ListenerSpec m], OutSpecs)

relayStubListeners :: (WithLogger m, Bi (InvMsg key tag), Bi (ReqMsg key tag), Bi (DataMsg contents), Message (InvMsg key tag), Message (ReqMsg key tag), Message (DataMsg contents)) => RelayProxy key tag contents -> ([ListenerSpec m], OutSpecs)

data RelayProxy key tag contents

Constructors

RelayProxy