Pos.Context

Description

Re-exports of Pos.Context.* functionality.

Documentation

class WithNodeContext ssc m | m -> ssc where

Class for something that has NodeContext inside.

Minimal complete definition

getNodeContext

Methods

getNodeContext :: m (NodeContext ssc)

type LrcSyncData = (Bool, EpochIndex)

Data used for LRC syncronization. First value is False iff LRC is running now. Second value is last epoch for which we have already computed LRC.

data NodeContext ssc

NodeContext contains runtime context of node.

Constructors

NodeContext

Fields

ncSystemStart :: !Timestamp

Time when system started working.

ncSecretKey :: !SecretKey

Secret key used for blocks creation.

ncGenesisUtxo :: !Utxo

Genesis utxo

ncGenesisLeaders :: !SlotLeaders

Leaders for 0-th epoch

ncTimeLord :: !Bool

Is time lord

ncJLFile :: !(Maybe (MVar FilePath))

ncDbPath :: !FilePath

Path to the database

ncSscContext :: !(SscNodeContext ssc)

ncAttackTypes :: ![AttackType]

Attack types used by malicious emulation

ncAttackTargets :: ![AttackTarget]

Attack targets used by malicious emulation

ncPropagation :: !Bool

Whether to propagate txs, ssc data, blocks to neighbors

ncBlkSemaphore :: !(MVar HeaderHash)

Semaphore which manages access to block application. Stored hash is a hash of last applied block.

ncLrcSync :: !(TVar LrcSyncData)

Primitive for synchronization with LRC.

ncUserSecret :: !(TVar UserSecret)

Secret keys (and path to file) which are used to send transactions

ncKademliaDump :: !FilePath

Path to kademlia dump file

ncNtpData :: !(TVar (Microsecond, Microsecond))

Data for NTP Worker. First element is margin (difference between global time and local time) which we got from NTP server in last tme. Second element is time (local time) for which we got margin in last time.

ncNtpLastSlot :: !(TVar SlotId)

Slot which was returned from getCurrentSlot in last time

ncBlockRetrievalQueue :: !(TBQueue (NodeId, NewestFirst NE (BlockHeader ssc)))

Concurrent queue that holds block headers that are to be downloaded.

ncRecoveryHeader :: !(TMVar (NodeId, BlockHeader ssc))

In case of recovery mode this variable holds the latest header hash we know about so we can do chained block requests. Invariant: this mvar is full iff we're more than recoveryHeadersMessage blocks deep relatively to some valid header and we're downloading blocks. Every time we get block that's more difficult than this one, we overwrite. Every time we process some blocks and fail or see that we've downloaded this header, we clean mvar.

ncPublicKey :: NodeContext ssc -> PublicKey

Generate PublicKey from SecretKey of NodeContext.

ncPubKeyAddress :: NodeContext ssc -> Address

Generate Address from SecretKey of NodeContext

Genesis

genesisUtxoM :: (Functor m, WithNodeContext ssc m) => m Utxo

genesisLeadersM :: (Functor m, WithNodeContext ssc m) => m SlotLeaders

Block semaphore.

putBlkSemaphore :: (MonadIO m, WithNodeContext ssc m) => HeaderHash -> m ()

readBlkSemaphore :: (MonadIO m, WithNodeContext ssc m) => m HeaderHash

takeBlkSemaphore :: (MonadIO m, WithNodeContext ssc m) => m HeaderHash

LRC synchronization

waitLrc :: (MonadIO m, WithNodeContext ssc m) => EpochIndex -> m ()

Block until LRC data is available for given epoch.

lrcActionOnEpoch :: (MonadIO m, WithNodeContext ssc m, MonadThrow m) => EpochIndex -> (EpochIndex -> m (Maybe a)) -> m a

lrcActionOnEpochReason :: (MonadIO m, WithNodeContext ssc m, MonadThrow m) => EpochIndex -> Text -> (EpochIndex -> m (Maybe a)) -> m a

NTP

setNtpLastSlot :: (MonadIO m, WithNodeContext ssc m) => SlotId -> m ()

readNtpLastSlot :: (MonadIO m, WithNodeContext ssc m) => m SlotId

readNtpMargin :: (MonadIO m, WithNodeContext ssc m) => m Microsecond

readNtpData :: (MonadIO m, WithNodeContext ssc m) => m (Microsecond, Microsecond)

newtype ContextHolder ssc m a

Wrapper for monadic action which brings NodeContext.

Constructors

ContextHolder

Fields

getContextHolder :: ReaderT (NodeContext ssc) m a

runContextHolder :: NodeContext ssc -> ContextHolder ssc m a -> m a

Run ContextHolder action.