Pos.Crypto

Description

Re-export of Crypto modules.

Documentation

data KeyPair

PublicKey with corresponding SecretKey.

Constructors

KeyPair

Fields

getPub :: PublicKey

getSec :: SecretKey

WithHash

data WithHash a

Constructors

WithHash

Fields

whData :: a

whHash :: Hash a

withHash :: Bi a => a -> WithHash a

_whData :: forall a. Lens' (WithHash a) a

_whHash :: forall a. Lens' (WithHash a) (Hash a)

Lenses for WithHash

AbstractHash

newtype AbstractHash algo a

Hash wrapper with phantom type for more type-safety. Made abstract in order to support different algorithms in different situations

Constructors

AbstractHash (Digest algo)

encodeAbstractHash :: forall algo a. Bi (AbstractHash algo a) => AbstractHash algo a -> Text

Encode hash from base64 form.

decodeAbstractHash :: forall algo a. Bi (AbstractHash algo a) => Text -> AbstractHash algo a

Parses given hash in base64 form.

encodeHash :: Bi (Hash a) => Hash a -> Text

Encode hash from base64 form.

decodeHash :: Bi (Hash a) => Text -> Hash a

Parses given hash in base64 form.

abstractHash :: (HashAlgorithm algo, Bi a) => a -> AbstractHash algo a

Encode thing as Binary data and then wrap into constructor.

unsafeAbstractHash :: (HashAlgorithm algo, Bi a) => a -> AbstractHash algo b

Unsafe version of abstractHash.

Common Hash

type Hash = AbstractHash Blake2s_224

Type alias for commonly used hash

hashHexF :: Format r (AbstractHash algo a -> r)

Specialized formatter for Hash.

shortHashF :: Format r (AbstractHash algo a -> r)

Smart formatter for Hash to show only first 8 characters of Hash.

hash :: Bi a => a -> Hash a

Short version of unsafeHash.

hashRaw :: ByteString -> Hash Raw

Raw constructor application.

unsafeHash :: Bi a => a -> Hash b

Encode thing as Bi data and then wrap into constructor.

Utility

class CastHash a b where

Type class for unsafe cast between hashes. You must ensure that types have identical Bi instances.

Methods

castHash :: AbstractHash algo a -> AbstractHash algo b

class HashAlgorithm a

Minimal complete definition

hashBlockSize, hashDigestSize, hashInternalContextSize, hashInternalInit, hashInternalUpdate, hashInternalFinalize

newtype SecureRandom a

You can use runSecureRandom on any MonadRandom computation to make it use a Really Secure™ randomness source (that is, OpenSSL).

Constructors

SecureRandom

Fields

runSecureRandom :: IO a

secureRandomBS :: MonadIO m => Int -> m ByteString

Generate a cryptographically random ByteString of specific length.

deterministic :: ByteString -> MonadPseudoRandom ChaChaDRG a -> a

You can use deterministic on any MonadRandom computation to make it use a seed (hopefully produced by a Really Secure™ randomness source). The seed has to have enough entropy to make this function secure.

randomNumber :: MonadRandom m => Integer -> m Integer

Generate a random number in range [0, n).

We want to avoid modulo bias, so we use the arc4random_uniform implementation (http:/stackoverflow.coma20051580615030). Specifically, we repeatedly generate a random number in range [0, 2^x) until we hit on something outside of [0, 2^x mod n), which means that it'll be in range [2^x mod n, 2^x). The amount of numbers in this interval is guaranteed to be divisible by n, and thus applying mod to it will be safe.

Keys and related.

newtype VssPublicKey

This key is used as public key in VSS.

Constructors

VssPublicKey

Fields

getVssPublicKey :: PublicKey

newtype VssKeyPair

This key pair is used to decrypt share generated by VSS.

Constructors

VssKeyPair KeyPair

toVssPublicKey :: VssKeyPair -> VssPublicKey

Extract VssPublicKey from VssKeyPair.

vssKeyGen :: MonadIO m => m VssKeyPair

Generate VssKeyPair using Really Secure™ randomness.

deterministicVssKeyGen :: ByteString -> VssKeyPair

Generate VssKeyPair using given seed.

Sharing

newtype DhSecret :: *

Constructors

DhSecret ByteString

type ShareId = Integer

newtype EncShare

Encrypted share which needs to be decrypted using VssKeyPair first.

Constructors

EncShare

Fields

getEncShare :: EncryptedShare

newtype Secret

Secret can be generated by genSharedSecret function along with shares.

Constructors

Secret

Fields

getSecret :: Secret

newtype SecretProof

SecretProof may be used to commit Secret without revealing it.

Constructors

SecretProof Proof

data SecretSharingExtra

This extra data may be used to verify encrypted share.

Constructors

SecretSharingExtra !ExtraGen ![Commitment]

newtype Share

Shares can be used to reconstruct Secret.

Constructors

Share

Fields

getShare :: DecryptedShare

type Threshold = Integer

shareId :: Share -> ShareId

Returns id of Share.

encShareId :: EncShare -> ShareId

Returns id of encrypted share.

decryptShare :: MonadRandom m => VssKeyPair -> EncShare -> m Share

Decrypt share using secret key. Doesn't verify if an encrypted share is valid, for this you need to use verifyEncShare.

getDhSecret :: DhSecret -> ByteString

Extract ByteString from DhSecret.

genSharedSecret :: MonadRandom m => Threshold -> NonEmpty VssPublicKey -> m (SecretSharingExtra, Secret, SecretProof, [EncShare])

Generate random secret using MonadRandom and share it between given public keys.

getValidShares :: Threshold -> [(EncShare, VssPublicKey, Share)] -> [Share]

Get #Threshold decrypted shares from given list, or less if there is not enough.

recoverSecret :: Threshold -> [(EncShare, VssPublicKey, Share)] -> Maybe Secret

Recover secret if there are enough correct shares.

secretToDhSecret :: Secret -> DhSecret

Transform a Secret into a usable random value.

unsafeRecoverSecret :: [Share] -> Secret

Recover Secret from shares. Assumes that number of shares is enough to do it. Consider using getValidShares first or use recoverSecret

verifyEncShare :: SecretSharingExtra -> VssPublicKey -> EncShare -> Bool

Verify an encrypted share using SecretSharingExtra.

verifySecretProof :: SecretSharingExtra -> Secret -> SecretProof -> Bool

Verify that SecretProof corresponds to Secret.

verifyShare :: EncShare -> VssPublicKey -> Share -> Bool

Verify that Share has been decrypted correctly.

Keys

newtype PublicKey

Wrapper around PublicKey.

Constructors

PublicKey PublicKey

newtype SecretKey

Wrapper around SecretKey.

Constructors

SecretKey SecretKey

keyGen :: MonadIO m => m (PublicKey, SecretKey)

Generate a key pair.

deterministicKeyGen :: ByteString -> Maybe (PublicKey, SecretKey)

Create key pair deterministically from 32 bytes.

toPublic :: SecretKey -> PublicKey

Generate a public key from a secret key. Fast (it just drops some bytes off the secret key).

formatFullPublicKey :: PublicKey -> Builder

Builder for PublicKey to show it in base64 encoded form.

fullPublicKeyF :: Format r (PublicKey -> r)

Formatter for PublicKey to show it in base64.

fullPublicKeyHexF :: Format r (PublicKey -> r)

Formatter for PublicKey to show it in hex.

parseFullPublicKey :: Bi PublicKey => Text -> Maybe PublicKey

Parse PublicKey from base64 encoded string.

Signing and verification

newtype Signature a

Wrapper around Signature.

Constructors

Signature Signature

sign :: Bi a => SecretKey -> a -> Signature a

Encode something with Binary and sign it.

checkSig :: Bi a => PublicKey -> a -> Signature a -> Bool

Verify a signature. #verifyRaw

fullSignatureHexF :: Format r (Signature a -> r)

Formatter for Signature to show it in hex.

data Signed a

Value and signature for this value.

Constructors

Signed

Fields

signedValue :: !a

Value to be signed

signedSig :: !(Signature a)

Signature of signedValue

mkSigned :: Bi a => SecretKey -> a -> Signed a

Smart constructor for Signed data type with proper signing.

Versions for raw bytestrings

signRaw :: SecretKey -> ByteString -> Signature Raw

Alias for constructor.

verifyRaw :: PublicKey -> ByteString -> Signature Raw -> Bool

Verify raw ByteString.

Proxy signature scheme

newtype ProxyCert w

Proxy certificate, made of ω + public key of delegate.

Constructors

ProxyCert

Fields

unProxyCert :: Signature

createProxyCert :: Bi w => SecretKey -> PublicKey -> w -> ProxyCert w

Proxy certificate creation from secret key of issuer, public key of delegate and the message space ω.

verifyProxyCert :: Bi w => PublicKey -> PublicKey -> w -> ProxyCert w -> Bool

Checks if certificate is valid, given issuer pk, delegate pk and ω.

data ProxySecretKey w

Convenient wrapper for secret key, that's basically ω plus certificate.

Constructors

ProxySecretKey

Fields

pskOmega :: w

pskIssuerPk :: PublicKey

pskDelegatePk :: PublicKey

pskCert :: ProxyCert w

createProxySecretKey :: Bi w => SecretKey -> PublicKey -> w -> ProxySecretKey w

Creates proxy secret key

verifyProxySecretKey :: Bi w => ProxySecretKey w -> Bool

Checks if proxy secret key is valid (the signature/cert inside is correct).

data ProxySignature w a

Delegate signature made with certificate-based permission. a stays for message type used in proxy (ω in the implementation notes), b for type of message signed.

Constructors

ProxySignature

Fields

pdOmega :: w

pdDelegatePk :: PublicKey

pdCert :: ProxyCert w

pdSig :: Signature

proxySign :: Bi a => SecretKey -> ProxySecretKey w -> a -> ProxySignature w a

Make a proxy delegate signature with help of certificate. If the delegate secret key passed doesn't pair with delegate public key in certificate inside, we panic. Please check this condition outside of this function.

proxyVerify :: (Bi w, Bi a) => PublicKey -> ProxySignature w a -> (w -> Bool) -> a -> Bool

Verify delegated signature given issuer's pk, signature, message space predicate and message itself.