Pos.DB

Description

Re-exports of Pos.DB functionality.

Documentation

getBlock :: (SscHelpersClass ssc, MonadDB ssc m) => HeaderHash -> m (Maybe (Block ssc))

Get block with given hash from Block DB.

getBlockHeader :: (SscHelpersClass ssc, MonadDB ssc m) => HeaderHash -> m (Maybe (BlockHeader ssc))

Returns header of block that was requested from Block DB.

getStoredBlock :: (SscHelpersClass ssc, MonadDB ssc m) => HeaderHash -> m (Maybe (StoredBlock ssc))

Get StoredBlock by hash from Block DB.

getUndo :: MonadDB ssc m => HeaderHash -> m (Maybe Undo)

Get undo data for block with given hash from Block DB.

getBlockWithUndo :: (SscHelpersClass ssc, MonadDB ssc m) => HeaderHash -> m (Maybe (Block ssc, Undo))

Retrieves block and undo together.

deleteBlock :: MonadDB ssc m => HeaderHash -> m ()

putBlock :: (SscHelpersClass ssc, MonadDB ssc m) => Undo -> Block ssc -> m ()

Put given block, its metadata and Undo data into Block DB.

prepareBlockDB :: forall ssc m. (SscHelpersClass ssc, MonadDB ssc m) => GenesisBlock ssc -> m ()

Load data

loadBlundsWhile :: (SscHelpersClass ssc, MonadDB ssc m) => (Block ssc -> Bool) -> HeaderHash -> m (NewestFirst [] (Blund ssc))

Load blunds starting from block with header hash equal to given hash and while predicate is true.

loadBlundsByDepth :: (SscHelpersClass ssc, MonadDB ssc m) => Word -> HeaderHash -> m (NewestFirst [] (Blund ssc))

Load blunds which have depth less than given.

loadBlocksWhile :: (SscHelpersClass ssc, MonadDB ssc m) => (Block ssc -> Bool) -> HeaderHash -> m (NewestFirst [] (Block ssc))

Load blocks starting from block with header hash equal to given hash and while predicate is true.

loadHeadersWhile :: (SscHelpersClass ssc, MonadDB ssc m) => (BlockHeader ssc -> Bool) -> HeaderHash -> m (NewestFirst [] (BlockHeader ssc))

Load headers starting from block with header hash equal to given hash and while predicate is true.

loadHeadersByDepth :: (SscHelpersClass ssc, MonadDB ssc m) => Word -> HeaderHash -> m (NewestFirst [] (BlockHeader ssc))

Load headers which have depth less than given.

loadHeadersByDepthWhile :: (SscHelpersClass ssc, MonadDB ssc m) => (BlockHeader ssc -> Bool) -> Word -> HeaderHash -> m (NewestFirst [] (BlockHeader ssc))

Load headers which have depth less than given and match some criterion.

class (MonadIO m, MonadThrow m) => MonadDB ssc m | m -> ssc where

Minimal complete definition

getNodeDBs, usingReadOptions, usingWriteOptions

Methods

getNodeDBs :: m (NodeDBs ssc)

usingReadOptions :: ReadOptions -> ASetter' (NodeDBs ssc) (DB ssc) -> m a -> m a

usingWriteOptions :: WriteOptions -> ASetter' (NodeDBs ssc) (DB ssc) -> m a -> m a

getBlockDB :: MonadDB ssc m => m (DB ssc)

getUtxoDB :: MonadDB ssc m => m (DB ssc)

getLrcDB :: MonadDB ssc m => m (DB ssc)

getMiscDB :: MonadDB ssc m => m (DB ssc)

openNodeDBs :: MonadIO m => Bool -> FilePath -> m (NodeDBs ssc)

Open all DBs stored on disk.

initNodeDBs :: forall ssc m. (SscHelpersClass ssc, WithNodeContext ssc m, MonadDB ssc m) => m ()

Initialize DBs if necessary.

getTip :: MonadDB ssc m => m HeaderHash

Get current tip from GState DB.

getTipBlock :: (SscHelpersClass ssc, MonadDB ssc m) => m (Block ssc)

Get block corresponding to tip.

getTipBlockHeader :: (SscHelpersClass ssc, MonadDB ssc m) => m (BlockHeader ssc)

Get BlockHeader corresponding to tip.

loadBlundsFromTipWhile :: (SscHelpersClass ssc, MonadDB ssc m) => (Block ssc -> Bool) -> m (NewestFirst [] (Blund ssc))

Load blunds from BlockDB starting from tip and while the condition is true.

loadBlundsFromTipByDepth :: (SscHelpersClass ssc, MonadDB ssc m) => Word -> m (NewestFirst [] (Blund ssc))

Load blunds from BlockDB starting from tip which have depth less than given.

sanityCheckDB :: (MonadMask m, MonadDB ssc m, WithLogger m) => m ()

class DBIteratorClass i where

Minimal complete definition

iterKeyPrefix

Associated Types

type IterKey i :: *

type IterValue i :: *

Methods

iterKeyPrefix :: Proxy i -> ByteString

type IterType i = (IterKey i, IterValue i)

newtype DBIterator i m a

Constructors

DBIterator

Fields

getDBIterator :: ReaderT Iterator m a

newtype DBMapIterator i v m a

Encapsulate `map f elements`, where elements - collection elements of type a. Holds `DBIterator m a` and apply f for every nextItem and curItem call. If f :: a -> b then we iterate by collection elements of type b.

Constructors

DBMapIterator

Fields

getDBMapIterator :: ReaderT (IterType i -> v) (DBIterator i m) a

type DBKeyIterator i = DBMapIterator i (IterKey i)

type DBValueIterator i = DBMapIterator i (IterValue i)

type DBnIterator ssc i = DBHolder ssc (DBIterator i IO)

type DBnMapIterator ssc i v = DBHolder ssc (DBMapIterator i v IO)

runDBnIterator :: forall i m ssc a. (MonadDB ssc m, DBIteratorClass i) => (NodeDBs ssc -> DB ssc) -> DBnIterator ssc i a -> m a

runDBnMapIterator :: forall i v m ssc a. (MonadDB ssc m, DBIteratorClass i) => (NodeDBs ssc -> DB ssc) -> DBnMapIterator ssc i v a -> (IterType i -> v) -> m a

runIterator :: forall i a m ssc. (MonadIO m, MonadMask m, DBIteratorClass i) => DBIterator i m a -> DB ssc -> m a

Run DBIterator by `DB ssc`.

runMapIterator :: forall i v m ssc a. (MonadIO m, MonadMask m, DBIteratorClass i) => DBMapIterator i v m a -> (IterType i -> v) -> DB ssc -> m a

Run DBMapIterator by `DB ssc`.

data DBError

Constructors

DBMalformed !Text
Structure of DB is malformed (e. g. data is inconsistent, something is missing, etc.)

openDB :: MonadIO m => FilePath -> m (DB ssc)

Key/Value helpers

encodeWithKeyPrefix :: forall i. (DBIteratorClass i, Bi (IterKey i)) => IterKey i -> ByteString

rocksDelete :: MonadIO m => ByteString -> DB ssc -> m ()

rocksGetBi :: forall v m ssc. (Bi v, MonadIO m, MonadThrow m) => ByteString -> DB ssc -> m (Maybe v)

Read serialized value from RocksDB using given key.

rocksGetBytes :: MonadIO m => ByteString -> DB ssc -> m (Maybe ByteString)

Read ByteString from RocksDb using given key.

rocksPutBi :: (Bi v, MonadIO m) => ByteString -> v -> DB ssc -> m ()

Write serializable value to RocksDb for given key.

rocksPutBytes :: MonadIO m => ByteString -> ByteString -> DB ssc -> m ()

Write ByteString to RocksDB for given key.

rocksDecodeWP :: forall i m. (MonadThrow m, DBIteratorClass i, Bi (IterKey i)) => ByteString -> m (IterKey i)

rocksDecodeMaybe :: Bi v => ByteString -> Maybe v

rocksDecodeMaybeWP :: forall i. (DBIteratorClass i, Bi (IterKey i)) => ByteString -> Maybe (IterKey i)

rocksDecodeKeyValMaybe :: (Bi k, Bi v) => (ByteString, ByteString) -> Maybe (k, v)

Batch

class RocksBatchOp a where

Minimal complete definition

toBatchOp

Methods

toBatchOp :: a -> [BatchOp]

data SomeBatchOp

Constructors

RocksBatchOp a => SomeBatchOp a

data SomePrettyBatchOp

Constructors

(RocksBatchOp a, Buildable a) => SomePrettyBatchOp a

rocksWriteBatch :: (RocksBatchOp a, MonadIO m) => [a] -> DB ssc -> m ()

Write Batch encapsulation

newtype DBHolder ssc m a

Constructors

DBHolder

Fields

getDBHolder :: ReaderT (NodeDBs ssc) m a

runDBHolder :: NodeDBs ssc -> DBHolder ssc m a -> m a

Execute DBHolder action with given NodeState.

prepareMiscDB :: forall ssc m. MonadDB ssc m => m ()

getProxySecretKeys :: MonadDB ssc m => m [ProxySKLight]

Gets proxy secret keys stored by node

addProxySecretKey :: MonadDB ssc m => ProxySKLight -> m ()

Adds proxy secret key if not present. Nothing if present.

removeProxySecretKey :: MonadDB ssc m => PublicKey -> m ()

Removes proxy secret key if present by issuer pk.

dropOldProxySecretKeys :: MonadDB ssc m => EpochIndex -> m ()

Given epochindex, throws away all outdated PSKs. Remark: it doesn't remove keys that can be used in future.

Secret storage

getSecretStorage :: MonadDB SscGodTossing m => m (Maybe GtSecretStorage)

putSecretStorage :: MonadDB SscGodTossing m => GtSecretStorage -> m ()

putSecretKeyHash :: MonadDB ssc m => Hash SecretKey -> m Bool

Puts or overwrites secret key of the node. Returns if it was overwritten.

checkSecretKeyHash :: MonadDB ssc m => Hash SecretKey -> m Bool

Checks if given secret key hash matches the hash in the database. Puts it into the database and return True if nothing was stored there.

General types.

data DB ssc

Constructors

DB

Fields

rocksReadOpts :: !ReadOptions

rocksWriteOpts :: !WriteOptions

rocksOptions :: !Options

rocksDB :: !DB

data NodeDBs ssc

Constructors

NodeDBs

Fields

_blockDB :: !(DB ssc)

Blocks, block index, undo data.

_gStateDB :: !(DB ssc)

Global state corresponding to some tip.

_lrcDB :: !(DB ssc)

Data computed by LRC.

_miscDB :: !(DB ssc)

Everything small and insignificant

_miscLock :: !RWLock

Lock on misc db

blockDB :: forall ssc. Lens' (NodeDBs ssc) (DB ssc)

gStateDB :: forall ssc. Lens' (NodeDBs ssc) (DB ssc)

lrcDB :: forall ssc. Lens' (NodeDBs ssc) (DB ssc)

miscDB :: forall ssc. Lens' (NodeDBs ssc) (DB ssc)

miscLock :: forall ssc. Lens' (NodeDBs ssc) RWLock

Snapshot

newtype Snapshot

Constructors

Snapshot Snapshot

usingSnapshot :: (MonadIO m, MonadMask m) => DB ssc -> (Snapshot -> m a) -> m a

Block DB related types.

data StoredBlock ssc

Constructors

StoredBlock

Fields

sbBlock :: !(Block ssc)

Block itself.

LRC related types.

data LeadersStorage ssc

Constructors

LeadersStorage

Fields

lrcEpoch :: !EpochIndex

lrcLeaders :: !SlotLeaders

data GtRichmenStorage ssc

Constructors

GtRichmenStorage

Fields

gtRichmenEpoch :: !EpochIndex

gtRichmen :: !RichmenStake

Update System related types.