Pos.DB.DB

Description

Higher-level DB functionality.

Documentation

openNodeDBs :: MonadIO m => Bool -> FilePath -> m NodeDBs

Open all DBs stored on disk.

initNodeDBs :: forall ssc m. (SscHelpersClass ssc, WithNodeContext ssc m, MonadDB m) => m ()

Initialize DBs if necessary.

getTip :: MonadDB m => m HeaderHash

Get current tip from GState DB.

getTipBlock :: (SscHelpersClass ssc, MonadDB m) => m (Block ssc)

Get block corresponding to tip.

getTipBlockHeader :: (SscHelpersClass ssc, MonadDB m) => m (BlockHeader ssc)

Get BlockHeader corresponding to tip.

loadBlundsFromTipWhile :: (SscHelpersClass ssc, MonadDB m) => (Block ssc -> Bool) -> m (NewestFirst [] (Blund ssc))

Load blunds from BlockDB starting from tip and while the condition is true.

loadBlundsFromTipByDepth :: (SscHelpersClass ssc, MonadDB m) => Word -> m (NewestFirst [] (Blund ssc))

Load blunds from BlockDB starting from tip which have depth less than given.

sanityCheckDB :: (MonadMask m, MonadDB m, WithLogger m) => m ()

Orphan instances

(MonadIO m, MonadThrow m, MonadCatch m) => MonadDBCore (DBHolder m)

Methods

dbAdoptedBVData :: DBHolder m BlockVersionData