Pos.DB.GState

Description

GState DB stores whole global state, i. e. result of application of all blocks up to tip.

Documentation

resolveForwardLink :: (HasHeaderHash a, MonadDB m) => a -> m (Maybe HeaderHash)

Tries to retrieve next block using current one (given a block/header).

isBlockInMainChain :: (HasHeaderHash a, MonadDB m) => a -> m Bool

Check if given hash representing block is in main chain.

data BlockExtraOp

Constructors

AddForwardLink HeaderHash HeaderHash
Adds or overwrites forward link

RemoveForwardLink HeaderHash
Removes forward link

SetInMainChain Bool HeaderHash
Enables or disables "in main chain" status of the block

loadHeadersUpWhile :: (SscHelpersClass ssc, MonadDB m, HasHeaderHash a) => a -> (BlockHeader ssc -> Int -> Bool) -> m (OldestFirst [] (BlockHeader ssc))

Returns headers loaded up.

loadBlocksUpWhile :: (SscHelpersClass ssc, MonadDB m, HasHeaderHash a) => a -> (Block ssc -> Int -> Bool) -> m (OldestFirst [] (Block ssc))

Returns blocks loaded up.

prepareGStateBlockExtra :: MonadDB m => HeaderHash -> m ()

module Pos.DB.GState.Common

prepareGStateDB :: forall ssc m. (WithNodeContext ssc m, MonadDB m) => HeaderHash -> m ()

Put missing initial data into GState DB.

sanityCheckGStateDB :: forall m. (MonadDB m, MonadMask m, WithLogger m) => m ()

Check that GState DB is consistent.

usingGStateSnapshot :: (MonadDB m, MonadMask m) => m a -> m a

getPSKByIssuerAddressHash :: MonadDB m => StakeholderId -> m (Maybe ProxySKHeavy)

Retrieves certificate by issuer address (hash of public key) if present.

getPSKByIssuer :: MonadDB m => PublicKey -> m (Maybe ProxySKHeavy)

Retrieves certificate by issuer public key if present.

isIssuerByAddressHash :: MonadDB m => StakeholderId -> m Bool

data DelegationOp

Constructors

AddPSK !ProxySKHeavy
Adds PSK. Overwrites if present.

DelPSK !PublicKey
Removes PSK by issuer PK.

runPskIterator :: forall m a. MonadDB m => DBnIterator PskIter a -> m a

runPskMapIterator :: forall v m a. MonadDB m => DBnMapIterator PskIter v a -> (IterType PskIter -> v) -> m a

Operations

data BalancesOp

Constructors

PutFtsSum !Coin

PutFtsStake !StakeholderId !Coin

Getters

isBootstrapEra :: MonadDB m => m Bool

getEffectiveTotalStake :: MonadDB m => m Coin

getEffectiveStake :: MonadDB m => StakeholderId -> m (Maybe Coin)

genesisFakeTotalStake :: Coin

Initialization

prepareGStateBalances :: forall m. MonadDB m => Utxo -> m ()

Iteration

data BalanceIter :: *

runBalanceIterator :: forall m a. MonadDB m => (forall iter. (MonadIterator (IterType BalanceIter) iter, MonadDB iter) => iter a) -> m a

Run iterator over effective balances.

runBalanceMapIterator :: forall v m a. MonadDB m => (forall iter. (MonadIterator v iter, MonadDB iter) => iter a) -> (IterType BalanceIter -> v) -> m a

Run map iterator over effective balances.

runBalanceIterBootstrap :: forall m a. Monad m => ListHolderT (IterType BalanceIter) m a -> m a

Sanity checks

sanityCheckBalances :: (MonadMask m, MonadDB m, WithLogger m) => m ()

Getters

getTxOutFromDB :: (MonadIO m, MonadThrow m) => TxIn -> DB -> m (Maybe TxOutAux)

getTxOut :: MonadDB m => TxIn -> m (Maybe TxOutAux)

Operations

data UtxoOp

Constructors

DelTxIn !TxIn

AddTxOut !TxIn !TxOutAux

Initialization

prepareGStateUtxo :: MonadDB m => Utxo -> m ()

Iteration

data UtxoIter

runUtxoIterator :: forall i m a. (MonadDB m, DBIteratorClass i, IterKey i ~ TxIn, IterValue i ~ TxOutAux) => DBnIterator i a -> m a

runUtxoMapIterator :: forall i v m a. (MonadDB m, DBIteratorClass i, IterKey i ~ TxIn, IterValue i ~ TxOutAux) => DBnMapIterator i v a -> (IterType i -> v) -> m a

getFilteredUtxo :: MonadDB m => Address -> m Utxo

Get utxo

getAllPotentiallyHugeUtxo :: MonadDB m => m Utxo

Get full utxo. Use with care – the utxo can be very big (hundreds of megabytes).

Sanity checks

sanityCheckUtxo :: (MonadDB m, WithLogger m) => Coin -> m ()

module Pos.Update.DB