Pos.DHT.Model.Class

Documentation

data DHTException

Data type for DHT exceptions.

Constructors

NodeDown

AllPeersUnavailable

class Monad m => MonadDHT m where

Monad for Distributed Hash Table operations.

Minimal complete definition

joinNetwork, discoverPeers, getKnownPeers, currentNodeKey, dhtLoggerName

Methods

joinNetwork :: [DHTNode] -> m ()

discoverPeers :: m [DHTNode]

Peer discovery: query DHT for random key Processing request, node will discover few other nodes We return these newly discovered nodes among with already known (List of known nodes is updated as well)

getKnownPeers :: m [DHTNode]

currentNodeKey :: m DHTKey

dhtLoggerName :: Proxy m -> LoggerName

withDhtLogger :: (HasLoggerName m, MonadDHT m) => m a -> m a

Perform some action using dhtLoggerName.