Pos.DHT.Model.Neighbors

Documentation

sendToNeighbors :: (MonadDHT m, MonadMockable m, Serializable packing body, WithLogger m, Message body) => SendActions packing m -> body -> m ()

Send default message to neighbours in parallel. It's a broadcasting to the neighbours without sessions (i.e. we don't have to wait for reply from the listeners).

sendToNode :: (MonadMockable m, Serializable packing body, Message body) => SendActions packing m -> NetworkAddress -> body -> m ()

converseToNeighbors :: (MonadDHT m, MonadMockable m, WithLogger m, Unpackable packing rcv, Packable packing snd, Message snd) => SendActions packing m -> (NodeId -> ConversationActions snd rcv m -> m ()) -> m ()

converseToNode :: (MonadMockable m, Unpackable packing rcv, Packable packing snd, Message snd) => SendActions packing m -> NetworkAddress -> (NodeId -> ConversationActions snd rcv m -> m t) -> m t