Pos.DHT.Real.Real

Documentation

runKademliaDHT :: (MonadMockable m, Eq (Promise m (Maybe ())), MonadIO m, WithLogger m) => KademliaDHTInstance -> KademliaDHT m a -> m a

Run KademliaDHT

startDHTInstance :: (MonadIO m, Mockable Catch m, Mockable Throw m, WithLogger m, Bi DHTData, Bi DHTKey) => KademliaDHTInstanceConfig -> m KademliaDHTInstance

Start KademliaDHTInstance with KademliaDHTInstanceConfig.

stopDHTInstance :: MonadIO m => KademliaDHTInstance -> m ()

Stop chosen KademliaDHTInstance.

Orphan instances

(MonadIO m, Mockable Async m, Mockable Catch m, Mockable Throw m, Eq (Promise m (Maybe ())), WithLogger m, Bi DHTData, Bi DHTKey) => MonadDHT (KademliaDHT m)

Methods

joinNetwork :: [DHTNode] -> KademliaDHT m ()

discoverPeers :: KademliaDHT m [DHTNode]

getKnownPeers :: KademliaDHT m [DHTNode]

currentNodeKey :: KademliaDHT m DHTKey

dhtLoggerName :: Proxy (* -> *) (KademliaDHT m) -> LoggerName