Pos.Data.Attributes

Description

Helper data type for block, tx attributes.

Map with integer 1-byte keys, arbitrary-type polymorph values. Needed primarily for partial serialization. Values are either parsed and put to some constructor or left as unparsed.

Documentation

data Attributes h

Convenient wrapper for the datatype to represent it (in binary format) as k-v map.

Constructors

Attributes

Fields

attrData :: h

Data, containing known keys (deserialized)

attrRemain :: ByteString

Unparsed ByteString

getAttributes :: (Word8 -> h -> Maybe (Get h)) -> Maybe Word32 -> h -> Get (Attributes h)

Generate Attributes reader given mapper from keys to Get, maximum input length and the attribute value h itself.

The mapper will be applied until it returns Nothing.

putAttributes :: (h -> [(Word8, Put)]) -> Attributes h -> Put

Generate Put given the way to serialize inner attribute value into set of keys and values.

mkAttributes :: h -> Attributes h