Pos.Exception

Description

Exceptions hierarchy in cardano-sl.

Documentation

data CardanoException

Root of exceptions in cardano-sl.

Constructors

(Buildable e, Exception e) => CardanoException e

cardanoExceptionToException :: (Buildable e, Exception e) => e -> SomeException

Helper to define sub-exception of CardanoException.

cardanoExceptionFromException :: Exception e => SomeException -> Maybe e

Helper to define sub-exception of CardanoException.

data CardanoFatalError

Error indicating that something really bad happened. Should be used when serious assertions fail (local equivalent of panic). panic is still alright to use, but preferably in pure environment.

Constructors

CardanoFatalError !Text

reportFatalError :: (WithLogger m, MonadThrow m) => Text -> m a

Print red message about fatal error and throw exception.

assertionFailed :: (WithLogger m, MonadThrow m) => Text -> m a

Report CardanoFatalError for failed assertions.