Pos.Genesis

Description

Blockchain genesis. Not to be confused with genesis block in epoch. Blockchain genesis means genesis values which are hardcoded in advance (before system starts doing anything). Genesis block in epoch exists in every epoch and it's not known in advance.

Static state

data StakeDistribution

Stake distribution in genesis block. FlatStakes is a flat distribution, i. e. each node has the same amount of coins. BitcoinStakes is a Bitcoin mining pool-style ditribution.

Constructors

FlatStakes !Word !Coin

BitcoinStakes !Word !Coin

RichPoorStakes

Fields

sdRichmen :: !Word

sdRichStake :: !Coin

sdPoor :: !Word

sdPoorStake :: !Coin

ExponentialStakes

ExplicitStakes !(HashMap Address (Coin, TxOutDistribution))

CombinedStakes StakeDistribution StakeDistribution

data GenesisData

Hardcoded genesis data

Constructors

GenesisData

Fields

gdAddresses :: [Address]

gdDistribution :: StakeDistribution

gdVssCertificates :: VssCertificatesMap

gdBootstrapBalances :: !(HashMap StakeholderId Coin)

getTotalStake :: StakeDistribution -> Coin

compileGenData :: GenesisData

Fetch pre-generated genesis data from genesis.bin in compile time Doesn't use TH with lift because it's difficult to provide Lift instance to GenesisData

genesisStakeDistribution :: StakeDistribution

genesisUtxo :: StakeDistribution -> Utxo

Genesis Utxo.

genesisDelegation :: HashMap StakeholderId [StakeholderId]

genesisAddresses :: [Address]

List of addresses in genesis. See genesisPublicKeys.

genesisSeed :: SharedSeed

Seed that will be used for the 0th epoch. We must hardcode a seed because we need to somehow determine leaders for the first ever epoch (stakes are hardcoded as well so we can run FTS on those stakes using this seed).

genesisBalances :: HashMap StakeholderId Coin

Genesis data used in development mode

genesisDevKeyPairs :: [(PublicKey, SecretKey)]

List of pairs from SecretKey with corresponding PublicKey.

genesisDevPublicKeys :: [PublicKey]

List of PublicKeys in genesis.

genesisDevSecretKeys :: [SecretKey]

List of SecretKeys in genesis.

Ssc

genesisLeaders :: Utxo -> SlotLeaders

Leaders of genesis. See followTheSatoshi.

Orphan instances

Default StakeDistribution

Methods

def :: StakeDistribution