Pos.Launcher

Node launchers.

runNodeProduction :: forall ssc. SscConstraint ssc => PeerId -> RealModeResources (ProductionMode ssc) -> ([WorkerSpec (ProductionMode ssc)], OutSpecs) -> NodeParams -> SscParams ssc -> Production ()

Run full node in real mode.

runNodeStats :: forall ssc. SscConstraint ssc => PeerId -> RealModeResources (StatsMode ssc) -> ([WorkerSpec (StatsMode ssc)], OutSpecs) -> NodeParams -> SscParams ssc -> Production ()

Run full node in benchmarking node

stakesDistr :: DistrOption -> DistrOption -> Maybe (Int, Int, Integer, Double) -> Bool -> StakeDistribution

data LoggingParams

Contains all parameters required for hierarchical logger initialization.

Constructors

LoggingParams

Fields

lpRunnerTag :: !LoggerName

Prefix for logger, like "time-slave"

lpHandlerPrefix :: !(Maybe FilePath)

Prefix of path for all logs

lpConfigPath :: !(Maybe FilePath)

Path to logger configuration

lpEkgPort :: !(Maybe Int)

data BaseParams

Contains basic & networking parameters for running node.

Constructors

BaseParams

Fields

bpLoggingParams :: !LoggingParams

Logger parameters

data NodeParams

Contains algorithm specific & storage parameters for Node.

Constructors

NodeParams

Fields

npDbPathM :: !FilePath

Path to node's database.

npRebuildDb :: !Bool

True if data-base should be rebuilt

npSystemStart :: !Timestamp

System start

npSecretKey :: !SecretKey

Primary secret key of node

npUserSecret :: !UserSecret

All node secret keys

npBaseParams :: !BaseParams

See BaseParams

npCustomUtxo :: !Utxo

predefined custom utxo

npJLFile :: !(Maybe FilePath)

npAttackTypes :: ![AttackType]

List of attack types used by malicious emulation

npAttackTargets :: ![AttackTarget]

List of targets to attack by malicious emulation

npPropagation :: !Bool

Whether to propagate txs, ssc data, blocks to neighbors

npReportServers :: ![Text]

List of report server URLs

npUpdateParams :: !UpdateParams

Params for update system

data RealModeResources m

Constructors

RealModeResources

Fields

rmTransport :: Transport m

rmGetPeers :: m (Set NodeId)

rmFindPeers :: m (Set NodeId)

hoistResources :: Functor n => (forall t. m t -> n t) -> RealModeResources m -> RealModeResources n

High level runners

runRawRealMode :: forall ssc a. SscConstraint ssc => PeerId -> RealModeResources (RawRealMode ssc) -> NodeParams -> SscParams ssc -> RawRealMode ssc (ListenersWithOut (RawRealMode ssc)) -> OutSpecs -> ActionSpec (RawRealMode ssc) a -> Production a

RawRealMode runner.

runProductionMode :: forall ssc a. SscConstraint ssc => PeerId -> RealModeResources (ProductionMode ssc) -> NodeParams -> SscParams ssc -> (ActionSpec (ProductionMode ssc) a, OutSpecs) -> Production a

ProductionMode runner.

runStatsMode :: forall ssc a. SscConstraint ssc => PeerId -> RealModeResources (StatsMode ssc) -> NodeParams -> SscParams ssc -> (ActionSpec (StatsMode ssc) a, OutSpecs) -> Production a

StatsMode runner. [CSL-169]: spawn here additional listener, which would accept stat queries can be done as part of refactoring (or someone who will refactor will create new issue).

runServiceMode :: PeerId -> RealModeResources ServiceMode -> BaseParams -> ListenersWithOut ServiceMode -> OutSpecs -> ActionSpec ServiceMode a -> Production a

ServiceMode runner.

Exported for custom usage in CLI utils

setupLoggers :: MonadIO m => LoggingParams -> m ()

bracketDHTInstance :: BaseParams -> KademliaParams -> (KademliaDHTInstance -> Production a) -> Production a

runServer :: (MonadIO m, MonadMockable m, MonadFix m, WithLogger m) => PeerId -> Transport m -> m (ListenersWithOut m) -> OutSpecs -> (Node m -> m t) -> (t -> m ()) -> ActionSpec m b -> m b

runServer_ :: (MonadIO m, MonadMockable m, MonadFix m, WithLogger m) => PeerId -> Transport m -> ListenersWithOut m -> OutSpecs -> ActionSpec m b -> m b

loggerBracket :: LoggingParams -> IO a -> IO a

RAII for node starter.

createTransportTCP :: (MonadIO m, WithLogger m, Mockable Throw m) => TCPAddr -> m (Transport m)

bracketTransport :: TCPAddr -> (Transport Production -> Production a) -> Production a

bracketResources :: BaseParams -> TCPAddr -> Set NodeId -> (RealModeResources Production -> Production a) -> IO a

Bracket a transport and use a static set of peers for discovery.

bracketResourcesKademlia :: BaseParams -> TCPAddr -> KademliaParams -> (KademliaDHTInstance -> RealModeResources Production -> Production a) -> IO a

Bracket a transport and a Kademlia node, using the latter for discovery.

data RealModeResources m

Constructors

RealModeResources

Fields

rmTransport :: Transport m

rmGetPeers :: m (Set NodeId)

rmFindPeers :: m (Set NodeId)

runNode :: (SscConstraint ssc, WorkMode ssc m, MonadDBCore m) => RealModeResources m -> ([WorkerSpec m], OutSpecs) -> (WorkerSpec m, OutSpecs)

Run full node in any WorkMode.

initSemaphore :: WorkMode ssc m => m ()

initLrc :: WorkMode ssc m => m ()

runNode' :: forall ssc m. (SscConstraint ssc, WorkMode ssc m, MonadDBCore m) => RealModeResources m -> [WorkerSpec m] -> WorkerSpec m

Run full node in any WorkMode.