Pos.Lrc

Documentation

newtype InvalidRichmenStake thd

Wrapper over RichmenStake. Its Arbitrary instance enforces that the stake distribution inside must be invalid, i.e. one of the stakeholders does not have sufficient coins to participate.

Constructors

Invalid

Fields

getInvalid :: RichmenStake

newtype ValidRichmenStake thd

Wrapper over RichmenStake. Its Arbitrary instance enforces that the stake distribution inside must be valid with respect to the threshold thd, i.e. all of the coins are non-zero and every stakeholder has sufficient coins (above 'coinPortionToDouble thd' %) to participate. Using reflection, the thd phantom type is used to get the threshold desired in each case.

Constructors

Valid

Fields

getValid :: RichmenStake

allLrcConsumers :: (SscWorkersClass ssc, WorkMode ssc m, MonadDBCore m) => [LrcConsumer m]

module Pos.Lrc.DB.Common

module Pos.Lrc.DB.Issuers

Getters

getLeaders :: MonadDB m => EpochIndex -> m (Maybe SlotLeaders)

Operations

putLeaders :: MonadDB m => EpochIndex -> SlotLeaders -> m ()

Initialization

prepareLrcLeaders :: (WithNodeContext ssc m, MonadDB m) => m ()

prepareLrcDB :: (WithNodeContext ssc m, MonadDB m) => m ()

Put missing initial data into LRC DB.

Initialization

prepareLrcRichmen :: (WithNodeContext ssc m, MonadDB m) => m ()

Concrete instances

Ssc

data RCSsc

getRichmenSsc :: MonadDB m => EpochIndex -> m (Maybe RichmenStake)

putRichmenSsc :: MonadDB m => EpochIndex -> FullRichmenData -> m ()

US

data RCUs

getRichmenUS :: MonadDB m => EpochIndex -> m (Maybe FullRichmenData)

putRichmenUS :: MonadDB m => EpochIndex -> FullRichmenData -> m ()

Delegation

data RCDlg

getRichmenDlg :: MonadDB m => EpochIndex -> m (Maybe Richmen)

putRichmenDlg :: MonadDB m => EpochIndex -> FullRichmenData -> m ()

module Pos.Lrc.DB.Seed

module Pos.Lrc.FtsPure

findRichmenStake

Arguments

:: MonadIterator (StakeholderId, Coin) m

=> Coin
Eligibility threshold

-> m RichmenStake

Find nodes which have at least 'eligibility threshold' coins.

findRichmenPure :: [(StakeholderId, Coin)] -> (Coin -> Coin) -> RichmenType -> FullRichmenData

Pure version of findRichmen which uses in-memory Utxo.

findAllRichmenMaybe

Arguments

:: (MonadDB m, MonadIterator (StakeholderId, Coin) m)

=> Maybe Coin
Eligibility threshold (optional)

-> Maybe Coin
Delegation threshold (optional)

-> m (RichmenStake, RichmenStake)

Function considers all variants of computation and compute using one pass by stake DB and one pass by delegation DB.

findDelegatedRichmen :: (MonadDB m, MonadIterator (StakeholderId, Coin) m) => Coin -> m RichmenStake

Find delegated richmen.

data RichmenType

Constructors

RTUsual

RTDelegation (HashMap StakeholderId [StakeholderId])

lrcOnNewSlotWorker :: (SscWorkersClass ssc, WorkMode ssc m, MonadDBCore m) => m (Set NodeId) -> (WorkerSpec m, OutSpecs)

lrcSingleShot :: (SscWorkersClass ssc, WorkMode ssc m, MonadDBCore m) => m (Set NodeId) -> EpochIndex -> m ()

Run leaders and richmen computation for given epoch. If stable block for this epoch is not known, LrcError will be thrown.

lrcSingleShotNoLock :: (SscWorkersClass ssc, WorkMode ssc m, MonadDBCore m) => m (Set NodeId) -> EpochIndex -> m ()

Same, but doesn't take lock on the semaphore.