Pos.Lrc.FtsPure

Description

Everything related to follow-the-satoshi procedure.

Documentation

followTheSatoshi :: SharedSeed -> Utxo -> NonEmpty StakeholderId

Choose several random stakeholders (specifically, their amount is currently hardcoded in epochSlots).

The probability that a stakeholder will be chosen is proportional to the number of coins this stakeholder holds. The same stakeholder can be picked more than once.

How the algorithm works: we sort all unspent outputs in a deterministic way (lexicographically) and have an ordered sequence of pairs (StakeholderId, Coin). Then we choose several random is between 1 and amount of satoshi in the system; to find owner of ith coin we find the lowest x such that sum of all coins in this list up to ith is not less than i (and then xth address is the owner).

With P2SH addresses, we don't know who is going to end up with funds sent to them. Therefore, P2SH addresses can contain addrDestination which specifies which addresses should count as “owning” funds for the purposes of follow-the-satoshi.

followTheSatoshiM :: MonadIterator (StakeholderId, Coin) m => SharedSeed -> Coin -> m (NonEmpty StakeholderId)

A version of followTheSatoshi that uses an iterator over TxOuts instead of Utxo.