Pos.Reporting.Methods

Description

Methods of reporting different unhealthy behaviour to server.

Documentation

sendReportNode :: (MonadIO m, MonadMask m, WithNodeContext її m) => ReportType -> m ()

Sends node's logs, taking LoggerConfig from NodeContext, retrieving all logger files from it. List of servers is also taken from node's configuration.

getNodeInfo :: (MonadDHT m, MonadIO m) => m Text

Retrieves node info that we would like to know when analyzing malicious behavior of node.

reportMisbehaviour :: forall m її. ReportingContext її m => Text -> m ()

Reports misbehaviour given reason string. Effectively designed for WorkMode context.

reportMisbehaviourMasked :: forall m її. ReportingContext її m => Text -> m ()

Report misbehaveour, but catch all errors inside

reportingFatal :: forall m a ё. ReportingContext ё m => m a -> m a

Execute action, report CardanoFatalError and FatalError if it happens and rethrow. Errors related to reporting itself are caught, logged and ignored.

sendReport :: (MonadIO m, MonadMask m) => [FilePath] -> [Text] -> ReportType -> Text -> Version -> String -> m ()

Given logs files list and report type, sends reports to URI asked. All files _must_ exist. Report server URI should be in form like "http(s):/host:port" without specified endpoint.

Important notice: if given paths are logs that we're currently writing to using this executable (or forked thread), you'll get an error, because there's no possibility to have two handles on the same file, see IO documentation on handles. Use second parameter for that.

retrieveLogFiles :: LoggerConfig -> [([Text], FilePath)]

Given logger config, retrieves all (logger name, filepath) for every logger that has file handle.

chooseLogFiles :: MonadIO m => FilePath -> m [FilePath]

Retrieves real filepathes of logs given filepathes from log description. Example: there's component.log in config, but this function will return [component.log.122, component.log.123].