Pos.Script

Description

A wrapper over Plutus (the scripting language used in transactions).

Documentation

data Script :: *

A script for inclusion into a transaction.

Constructors

Script

Fields

scrVersion :: ScriptVersion

Version

scrScript :: LByteString

Serialized script

type TxScriptError = String

txScriptCheck

Arguments

:: Bi Script_v0

=> TxSigData

-> Script
Validator

-> Script
Redeemer

-> Either TxScriptError ()

Validate a transaction, given a validator and a redeemer.

parseValidator :: Bi Script_v0 => Text -> Either String Script

Parse a script intended to serve as a validator (or “lock”) in a transaction output.

parseRedeemer :: Bi Script_v0 => Maybe Script -> Text -> Either String Script

Parse a script intended to serve as a redeemer (or “proof”) in a transaction input.

Can be given an optional validator (e.g. if the redeemer uses functions or types defined by the validator).

stdlib :: Program

isKnownScriptVersion :: ScriptVersion -> Bool