Pos.Slotting

Description

Slotting functionality.

Documentation

class Monad m => MonadSlots m where

Type class providing information about time when system started functioning.

Methods

getSystemStartTime :: m Timestamp

getCurrentTime :: m Timestamp

getCurrentSlot :: m SlotId

getSlotDuration :: m Millisecond

getSystemStartTime :: (MonadTrans t, MonadSlots m', t m' ~ m) => m Timestamp

getCurrentTime :: (MonadTrans t, MonadSlots m', t m' ~ m) => m Timestamp

getCurrentSlot :: (MonadTrans t, MonadSlots m', t m' ~ m) => m SlotId

getSlotDuration :: (MonadTrans t, MonadSlots m', t m' ~ m) => m Millisecond

getCurrentSlotFlat :: MonadSlots m => m FlatSlotId

Get flat id of current slot based on MonadSlots.

getSlotStart :: MonadSlots m => SlotId -> m Timestamp

Get timestamp when given slot starts.

getCurrentSlotUsingNtp :: (MonadSlots m, Mockable CurrentTime m) => SlotId -> (Microsecond, Microsecond) -> m SlotId

onNewSlot :: OnNewSlot ssc m => Bool -> (SlotId -> m ()) -> m ()

Run given action as soon as new slot starts, passing SlotId to it. This function uses Mockable and assumes consistency between MonadSlots and Mockable implementations.

onNewSlotImpl :: OnNewSlot ssc m => Bool -> Bool -> (SlotId -> m ()) -> m ()

data SlottingState

Data needed for the slotting algorithm to work.

Constructors

SlottingState

Fields

_ssNtpLastSlot :: !SlotId

Slot which was returned from getCurrentSlot in last time

_ssNtpData :: !(Microsecond, Microsecond)

Data for the NTP Worker. First element: margin (difference between global time and local time) which we got from NTP server in last time. Second element: time (local) for which we got margin in last time.

ssNtpLastSlot :: Lens' SlottingState SlotId

ssNtpData :: Lens' SlottingState (Microsecond, Microsecond)

slottingWorker :: WorkMode ssc m => (WorkerSpec m, OutSpecs)

This worker is necessary for Slotting to work.