Pos.Ssc.Class

Description

Re-exports of Pos.Ssc.Class.* modules.

Documentation

class Ssc ssc => SscHelpersClass ssc where

Minimal complete definition

sscVerifyPayload

Methods

sscVerifyPayload :: Tagged ssc (Either EpochIndex (MainBlockHeader ssc) -> SscPayload ssc -> Either (SscVerifyError ssc) ())

class Ssc ssc => SscListenersClass ssc where

Class for defining listeners in DHT SSC implementation.

Minimal complete definition

sscListeners, sscStubListeners

Methods

sscListeners :: WorkMode ssc m => m (Tagged ssc ([ListenerSpec m], OutSpecs))

sscStubListeners :: WithLogger m => Tagged ssc ([ListenerSpec m], OutSpecs)

type LocalQuery ssc a = forall m. (MonadReader (SscLocalData ssc) m, WithLogger m) => m a

type LocalUpdate ssc a = forall m. (MonadState (SscLocalData ssc) m, WithLogger m) => m a

class Ssc ssc => SscLocalDataClass ssc where

This type class abstracts local data used for SSC. Local means that it is not stored in blocks.

Minimal complete definition

sscGetLocalPayloadQ, sscNormalizeU, sscNewLocalData

Methods

sscGetLocalPayloadQ :: SlotId -> LocalQuery ssc (SscPayload ssc)

Get local payload to be put into main block and for given SlotId. If payload for given SlotId can't be constructed, empty payload can be returned.

sscNormalizeU :: EpochIndex -> RichmenStake -> SscGlobalState ssc -> LocalUpdate ssc ()

Make SscLocalData valid for given epoch, richmen and global state. of best known chain).

sscNewLocalData :: (MonadSlots m, MonadDB m) => m (SscLocalData ssc)

Create new (empty) local data. We are using this function instead of Default class, because it gives more flexibility. For instance, one can read something from DB or get current slot.

Modern

class Ssc ssc => SscGStateClass ssc where

Minimal complete definition

sscLoadGlobalState, sscGlobalStateToBatch, sscRollbackU, sscVerifyAndApplyBlocks, sscCalculateSeedQ

Methods

sscLoadGlobalState :: (MonadDB m, WithLogger m) => m (SscGlobalState ssc)

Load global state from DB by recreating it from recent blocks.

sscGlobalStateToBatch :: SscGlobalState ssc -> Tagged ssc [SomeBatchOp]

Dump global state to DB.

sscRollbackU :: NewestFirst NE (Block ssc) -> SscGlobalUpdate ssc ()

Rollback application of blocks.

sscVerifyAndApplyBlocks :: RichmenStake -> OldestFirst NE (Block ssc) -> SscVerifier ssc ()

Verify SSC-related part of given blocks with respect to current GState and apply them on success. Blocks must be from the same epoch.

sscCalculateSeedQ :: EpochIndex -> RichmenStake -> SscGlobalQuery ssc (Either (SscSeedError ssc) SharedSeed)

Calculate SharedSeed for given epoch using SscGlobalState.

type SscGlobalQuery ssc a = forall m. (MonadReader (SscGlobalState ssc) m, WithLogger m) => m a

type SscGlobalUpdate ssc a = forall m. (MonadState (SscGlobalState ssc) m, WithLogger m) => m a

type SscVerifier ssc a = forall m. SscVerifyMode ssc m => m a

class (Typeable ssc, Typeable (SscPayload ssc), Typeable (SscProof ssc), Typeable (SscSeedError ssc), Eq (SscProof ssc), Eq (SscGlobalState ssc), Show (SscProof ssc), Show (SscPayload ssc), Buildable (SscPayload ssc), Buildable (SscSeedError ssc), Buildable (SscVerifyError ssc), Buildable (SscGlobalState ssc), Bi (SscProof ssc), Bi (SscPayload ssc), NFData (SscPayload ssc), NFData (SscProof ssc)) => Ssc ssc where

Main Shared Seed Calculation type class. Stores all needed type parameters for general implementation of SSC.

Minimal complete definition

mkSscProof, sscCreateNodeContext

Associated Types

type SscLocalData ssc

Internal SSC state stored in memory

type SscPayload ssc

Payload which will be stored in main blocks

type SscGlobalState ssc

Global state, which is formed from all known blocks

type SscProof ssc

Proof that SSC payload is correct (it'll be included into block header)

type SscSeedError ssc

Error that can happen when calculating the seed

type SscNodeContext ssc

SSC specific context in NodeContext

type SscParams ssc

Needed options for creating SscNodeContext

type SscVerifyError ssc

Type for verification error

Methods

mkSscProof :: Tagged ssc (SscPayload ssc -> SscProof ssc)

Create proof (for inclusion into block header) from payload

sscCreateNodeContext :: MonadIO m => Tagged ssc (SscParams ssc -> m (SscNodeContext ssc))

Create SscNodeContext

class Ssc ssc => SscWorkersClass ssc where

Class for SSC workers.

Minimal complete definition

sscWorkers, sscLrcConsumers

Methods

sscWorkers :: WorkMode ssc m => m (Set NodeId) -> Tagged ssc ([WorkerSpec m], OutSpecs)

All workers specific to SSC.

sscLrcConsumers :: (WorkMode ssc m, MonadDBCore m) => Tagged ssc [LrcConsumer m]

type SscConstraint ssc = (Ssc ssc, SscListenersClass ssc, SscLocalDataClass ssc, SscHelpersClass ssc, SscGStateClass ssc, SscWorkersClass ssc, SecurityWorkersClass ssc)

type WorkModeSsc ssc = (SscLocalDataClass ssc, SscHelpersClass ssc)