Pos.Ssc.Class.LocalData

Description

This module defines type class for local data storage.

Documentation

type LocalQuery ssc a = forall m. MonadReader (SscLocalData ssc) m => m a

type LocalUpdate ssc a = forall m. MonadState (SscLocalData ssc) m => m a

class Ssc ssc => SscLocalDataClass ssc where

This type class abstracts local data used for SSC. Local means that it is not stored in blocks.

Minimal complete definition

sscGetLocalPayloadQ, sscApplyGlobalStateU

Methods

sscGetLocalPayloadQ :: LocalQuery ssc (SlotId, SscPayload ssc)

Get local payload to be put into main block and SlotId for which it's valid (wrt current tip).

sscApplyGlobalStateU :: Richmen -> SscGlobalState ssc -> LocalUpdate ssc ()

Update LocalData using global data from blocks (last version of best known chain).