Pos.Ssc.Class.Storage

Description

Storage for generic Shared Seed calculation implementation.

Modern

class Ssc ssc => SscStorageClass ssc where

Minimal complete definition

sscLoadGlobalState, sscApplyBlocksM, sscRollbackM, sscVerifyBlocksM, sscCalculateSeedM

Methods

sscLoadGlobalState :: MonadDB ssc m => HeaderHash -> m (SscGlobalState ssc)

sscApplyBlocksM :: OldestFirst NE (Block ssc) -> SscGlobalUpdate ssc ()

sscRollbackM :: NewestFirst NE (Block ssc) -> SscGlobalUpdate ssc ()

Rollback application of blocks.

sscVerifyBlocksM :: Bool -> Richmen -> OldestFirst NE (Block ssc) -> SscGlobalQuery ssc VerificationRes

Verify Ssc-related predicates of block sequence which is about to be applied. It should check that SSC payload will be consistent if this blocks are applied (after possible rollback if first argument isn't zero).

sscCalculateSeedM :: EpochIndex -> SscGlobalQuery ssc (Either (SscSeedError ssc) SharedSeed)

type SscGlobalQuery ssc a = forall m. MonadReader (SscGlobalState ssc) m => m a

type SscImpureQuery ssc a = forall m. (MonadReader (SscGlobalState ssc) m, WithNodeContext ssc m, MonadIO m) => m a

type SscGlobalUpdate ssc a = forall m. MonadState (SscGlobalState ssc) m => m a