Pos.Ssc.Class.Types

Description

Types for Shared Seed calculation.

Documentation

class (Typeable ssc, Typeable (SscPayload ssc), Typeable (SscProof ssc), Typeable (SscSeedError ssc), Eq (SscProof ssc), Show (SscProof ssc), Show (SscPayload ssc), Buildable (SscPayload ssc), Buildable (SscSeedError ssc), Buildable (SscVerifyError ssc), Buildable (SscGlobalState ssc), Bi (SscProof ssc), Bi (SscPayload ssc), Default (SscLocalData ssc)) => Ssc ssc where

Main Shared Seed Calculation type class. Stores all needed type parameters for general implementation of SSC.

Minimal complete definition

mkSscProof, sscCreateNodeContext

Associated Types

type SscLocalData ssc

Internal SSC state stored in memory

type SscPayload ssc

Payload which will be stored in main blocks

type SscGlobalState ssc

Global state, which is formed from all known blocks

type SscProof ssc

Proof that SSC payload is correct (it'll be included into block header)

type SscSeedError ssc

Error that can happen when calculating the seed

type SscNodeContext ssc

SSC specific context in NodeContext

type SscParams ssc

Needed options for creating SscNodeContext

type SscVerifyError ssc

Type for verification error

Methods

mkSscProof :: Tagged ssc (SscPayload ssc -> SscProof ssc)

Create proof (for inclusion into block header) from payload

sscCreateNodeContext :: MonadIO m => SscParams ssc -> m (SscNodeContext ssc)

Create SscNodeContext