Pos.Ssc.GodTossing

Description

Re-exports of GodTossing modules.

GodTossing is a coin tossing with guaranteed output delivery. Nodes exchange commitments, openings, and shares, and in the end arrive at a shared seed.

See https://eprint.iacr.org/2015/889.pdf (“A Provably Secure Proof-of-Stake Blockchain Protocol”), section 4 for more details.

Documentation

data CommitmentOpening

Pair of Commitment and Opening.

Constructors

CommitmentOpening

Fields

coCommitment :: !Commitment

coOpening :: !Opening

data SeedError

Data type for error during seed calculation.

Constructors

ExtraneousOpenings (HashSet StakeholderId)
Some nodes in the OpeningsMap aren't in the set of participants

ExtraneousShares (HashSet StakeholderId)
Some nodes in the SharesMap aren't in the set of participants

NoParticipants
There were no participants so a random string couldn't be generated

BrokenCommitment StakeholderId
Commitment can't be deserialized or didn't match secret (either recovered or in openings)

NoSecretFound StakeholderId
Secret couldn't be recovered, or wasn't found in either OpeningsMap or SharesMap

BrokenSecret StakeholderId
Secret can't be deserialized

BrokenShare StakeholderId
Share can't be deserialized

Helpers

genCommitmentAndOpening :: (MonadFail m, MonadIO m) => Threshold -> NonEmpty (AsBinary VssPublicKey) -> m (Commitment, Opening)

Generate securely random SharedSeed.

isCommitmentId :: SlotId -> Bool

isCommitmentIdx :: LocalSlotIndex -> Bool

isOpeningId :: SlotId -> Bool

isOpeningIdx :: LocalSlotIndex -> Bool

isSharesId :: SlotId -> Bool

isSharesIdx :: LocalSlotIndex -> Bool

hasCommitment :: Bi Commitment => EpochIndex -> StakeholderId -> GtGlobalState -> Bool

hasOpening :: StakeholderId -> GtGlobalState -> Bool

hasShares :: StakeholderId -> GtGlobalState -> Bool

hasVssCertificate :: StakeholderId -> GtGlobalState -> Bool

mkSignedCommitment :: Bi Commitment => SecretKey -> EpochIndex -> Commitment -> SignedCommitment

Make signed commitment from commitment and epoch index using secret key.

secretToSharedSeed :: Secret -> SharedSeed

Convert Secret to SharedSeed.

Verification and Checks

checkCertSign :: (StakeholderId, VssCertificate) -> Bool

Check that the VSS certificate is signed properly #checkPubKeyAddress #checkSig

checkCertTTL :: EpochIndex -> VssCertificate -> Bool

Check that the VSS certificate has valid TTL: more than 0 and less than vssMaxTTL

verifyCommitment :: Commitment -> Bool

Verify that Commitment is correct.

#verifyEncShare

verifyCommitmentSignature :: Bi Commitment => EpochIndex -> SignedCommitment -> Bool

Verify signature in SignedCommitment using epoch index.

#checkSig

verifySignedCommitment :: Bi Commitment => StakeholderId -> EpochIndex -> SignedCommitment -> VerificationRes

Verify SignedCommitment using public key and epoch index.

#verifyCommitmentPK #verifyCommitmentSignature #verifyCommitment

verifyOpening :: Commitment -> Opening -> Bool

Verify that Secret provided with Opening corresponds to given commitment.

#verifySecretProof

checkShare :: (SetContainer set, ContainerKey set ~ StakeholderId) => CommitmentsMap -> set -> VssCertificatesMap -> (StakeholderId, StakeholderId, AsBinary Share) -> Bool

Check that the decrypted share matches the encrypted share in the commitment

#verifyShare

checkShares :: (SetContainer set, ContainerKey set ~ StakeholderId) => CommitmentsMap -> set -> VssCertificatesMap -> StakeholderId -> InnerSharesMap -> Bool

checkOpeningMatchesCommitment :: CommitmentsMap -> (StakeholderId, Opening) -> Bool

Check that the secret revealed in the opening matches the secret proof in the commitment

checkCommShares :: [AsBinary VssPublicKey] -> SignedCommitment -> Bool

GtPayload

verifyGtPayload :: (SscPayload ssc ~ GtPayload, Bi Commitment) => MainBlockHeader ssc -> SscPayload ssc -> Either TossVerFailure ()

Verification helper

verifyEntriesGuard :: MonadError TossVerFailure m => (entry -> key) -> (entry -> verificationVal) -> (NonEmpty key -> TossVerFailure) -> (verificationVal -> Bool) -> [entry] -> m ()

VSS

vssThreshold :: Integral a => a -> Threshold

Figure out the threshold (i.e. how many secret shares would be required to recover each node's secret) using number of participants.

computeParticipants :: Richmen -> VssCertificatesMap -> VssCertificatesMap

getStableCertsPure :: EpochIndex -> VssCertData -> VssCertificatesMap

genesisCertificates :: VssCertificatesMap

Certificates in genesis represented as VssCertificatesMap.

genesisVssKeyPairs :: [VssKeyPair]

List of VssKeyPair in genesis.

localOnNewSlot :: MonadSscLD SscGodTossing m => RichmenSet -> SlotId -> m ()

Clean-up some data when new slot starts.

sscIsDataUseful :: MonadSscLD SscGodTossing m => GtMsgTag -> StakeholderId -> m Bool

Check whether SSC data with given tag and public key can be added to local data.

sscProcessMessage :: (MonadSscLD SscGodTossing m, SscBi) => Richmen -> GtMsgContents -> StakeholderId -> m (Either TossVerFailure ())

Process message and save it if needed. Result is whether message has been actually added.

getLocalPayload :: LocalQuery SscGodTossing (SlotId, GtPayload)

Instances

instance SscLocalDataClass SscGodTossing

data GtLocalData

Constructors

GtLocalData

Fields

_ldCommitments :: !CommitmentsMap

Local set of Commitments. These are valid commitments which are known to the node and not stored in blockchain. It is useful only for the first k slots, after that it should be discarded.

_ldOpenings :: !OpeningsMap

Local set of openings

_ldShares :: !SharesMap

Local set of decrypted shares (encrypted shares are stored in commitments).

_ldCertificates :: !VssCertData

Local set of VSS certificates

_ldLastProcessedSlot :: !SlotId

Last slot we are aware of.

ldCommitments :: Lens' GtLocalData CommitmentsMap

ldOpenings :: Lens' GtLocalData OpeningsMap

ldShares :: Lens' GtLocalData SharesMap

ldCertificates :: Lens' GtLocalData VssCertData

ldLastProcessedSlot :: Lens' GtLocalData SlotId

getOurCommitment :: MonadDB SscGodTossing m => EpochIndex -> m (Maybe SignedCommitment)

Get our commitment for given epoch if it's known.

getOurOpening :: MonadDB SscGodTossing m => EpochIndex -> m (Maybe Opening)

Get our opening corresponding for given epoch if it's known.

putOurSecret :: MonadDB SscGodTossing m => SignedCommitment -> Opening -> EpochIndex -> m ()

Put our secret for given epoch.

calculateSeed

Arguments

:: CommitmentsMap
All participating nodes

-> OpeningsMap

-> SharesMap
Decrypted shares

-> Either SeedError SharedSeed

Calculate SharedSeed. SharedSeed is a random bytestring that all nodes generate together and agree on.

TODO: do we need to check secrets' lengths? Probably not.

Instances

instance SscStorageClass SscGodTossing

getGlobalCerts :: MonadSscGS SscGodTossing m => SlotId -> m VssCertificatesMap

gtGetGlobalState :: MonadSscGS SscGodTossing m => m GtGlobalState

getStableCerts :: MonadSscGS SscGodTossing m => EpochIndex -> m VssCertificatesMap

Get stable VSS certificates for given epoch.

Types

data Commitment

Commitment is a message generated during the first stage of GodTossing. It contains encrypted shares and proof of secret. Invariant which must be ensured: commShares is not empty.

Constructors

Commitment

Fields

commExtra :: !(AsBinary SecretSharingExtra)

commProof :: !(AsBinary SecretProof)

commShares :: !(HashMap (AsBinary VssPublicKey) (AsBinary EncShare))

type CommitmentSignature = Signature (EpochIndex, Commitment)

Signature which ensures that commitment was generated by node with given public key for given epoch.

type SignedCommitment = (PublicKey, Commitment, CommitmentSignature)

type CommitmentsMap = HashMap StakeholderId SignedCommitment

type InnerSharesMap = HashMap StakeholderId (AsBinary Share)

Each node generates a SharedSeed, breaks it into Shares, and sends those encrypted shares to other nodes. In a SharesMap, for each node we collect shares which said node has received and decrypted.

Specifically, if node identified by Address X has received a share from node identified by key Y, this share will be at sharesMap ! X ! Y.

newtype Opening

Opening reveals secret.

Constructors

Opening

Fields

getOpening :: AsBinary Secret

type OpeningsMap = HashMap StakeholderId Opening

type SharesMap = HashMap StakeholderId InnerSharesMap

data VssCertificate

VssCertificate allows VssPublicKey to participate in MPC. Each stakeholder should create a Vss keypair, sign public key with signing key and send it into blockchain.

A public key of node is included in certificate in order to enable validation of it using only node's P2PKH address. Expiry epoch is last epoch when certificate is valid, expiry epoch is included in certificate and signature.

Other nodes accept this certificate if it is valid and if node really has some stake.

Constructors

VssCertificate

Fields

vcVssKey :: !(AsBinary VssPublicKey)

vcExpiryEpoch :: !EpochIndex

vcSignature :: !(Signature (AsBinary VssPublicKey, EpochIndex))

vcSigningKey :: !PublicKey

mkVssCertificate :: SecretKey -> AsBinary VssPublicKey -> EpochIndex -> VssCertificate

type VssCertificatesMap = HashMap StakeholderId VssCertificate

VssCertificatesMap contains all valid certificates collected during some period of time.

type NodeSet = HashSet StakeholderId

data GtMsgTag

Tag associated with message.

Constructors

CommitmentMsg

OpeningMsg

SharesMsg

VssCertificateMsg

isGoodSlotForTag :: GtMsgTag -> LocalSlotIndex -> Bool

isGoodSlotIdForTag :: GtMsgTag -> SlotId -> Bool

data GtMsgContents

Data message. Can be used to send actual data.

Constructors

MCCommitment !SignedCommitment

MCOpening !Opening

MCShares !InnerSharesMap

MCVssCertificate !VssCertificate

msgContentsTag :: GtMsgContents -> GtMsgTag

GtMsgTag appropriate for given DataMsg.

data SscGodTossing

Data type which represents shared seed calculation tag in -XTypeApplication hacks with type families.

Instance types

data GtPayload

Block payload

Constructors

CommitmentsPayload !CommitmentsMap !VssCertificatesMap

OpeningsPayload !OpeningsMap !VssCertificatesMap

SharesPayload !SharesMap !VssCertificatesMap

CertificatesPayload !VssCertificatesMap

data GtProof

Proof of MpcData. We can use ADS for commitments, openings, shares as well, if we find it necessary.

Constructors

CommitmentsProof !(Hash CommitmentsMap) !(Hash VssCertificatesMap)

OpeningsProof !(Hash OpeningsMap) !(Hash VssCertificatesMap)

SharesProof !(Hash SharesMap) !(Hash VssCertificatesMap)

CertificatesProof !(Hash VssCertificatesMap)

data GtGlobalState

Global state of GodTossing, contains relevant SSC data from blocks.

Constructors

GtGlobalState

Fields

_gsCommitments :: !CommitmentsMap

Commitments are added during the first phase of epoch.

_gsOpenings :: !OpeningsMap

Openings are added during the second phase of epoch.

_gsShares :: !SharesMap

Decrypted shares to be used in the third phase.

_gsVssCertificates :: !VssCertData

Vss certificates are added at any time if they are valid and received from stakeholders.

data GtContext

Constructors

GtContext

Fields

gtcVssKeyPair :: !VssKeyPair

Vss key pair used for MPC.

gtcParticipateSsc :: !(TVar Bool)

gtcVssCertificateVerified :: !(TVar Bool)

data GtParams

Constructors

GtParams

Fields

gtpRebuildDb :: !Bool

gtpSscEnabled :: !Bool

Whether node should participate in SSC in case SSC requires participation.

gtpVssKeyPair :: !VssKeyPair

Key pair used for secret sharing

data GtSecretStorage

Constructors

GtSecretStorage

Fields

gssCommitment :: !SignedCommitment

Our commitment.

gssOpening :: !Opening

Corresponding opening

gssEpoch :: !EpochIndex

Epoch for which this secret were generated

data TossVerFailure

Constructors

TossVerFailure

Fields

tvfErrorTag :: !TossVerErrorTag

tvfStakeholders :: !(NonEmpty StakeholderId)

NotCommitmentPhase !SlotId

NotOpeningPhase !SlotId

NotSharesPhase !SlotId

CertificateInvalidSign !(NonEmpty (StakeholderId, VssCertificate))

CertificateInvalidTTL !(NonEmpty (VssCertificate, EpochIndex))

TossInternallError !Text

data TossVerErrorTag

Constructors

CommitmentInvalid

CommitingNoParticipants

CommitmentAlreadySent

CommSharesOnWrongParticipants

OpeningAlreadySent

OpeningWithoutCommitment

OpeningNotMatchCommitment

SharesNotRichmen

InternalShareWithoutCommitment

SharesAlreadySent

DecrSharesNotMatchCommitment

CertificateAlreadySent

CertificateNotRichmen

Lenses

GtPayload

gsCommitments :: Lens' GtGlobalState CommitmentsMap

gsOpenings :: Lens' GtGlobalState OpeningsMap

gsShares :: Lens' GtGlobalState SharesMap

gsVssCertificates :: Lens' GtGlobalState VssCertData

mkGtProof :: (Bi VssCertificate, Bi Commitment, Bi Opening) => GtPayload -> GtProof

Smart constructor for GtProof from GtPayload.

createGtContext :: MonadIO m => GtParams -> m GtContext

_gpCertificates :: GtPayload -> VssCertificatesMap

emptyPayload :: GtPayload

type SscBi = (Bi GtProof, Bi GtPayload, Bi Opening, Bi VssCertificate, Bi Commitment)

data VssCertData

Wrapper around VssCertificate with TTL. Every VssCertificate has own TTL. Wrapper supports simple HashMap operations. Wrapper holds VssCertificatesMap and Set of certificates sorted by expiry epoch.

Constructors

VssCertData

Fields

lastKnownEoS :: !EpochOrSlot

Last known slot, every element of expirySlotSet > lastKnownEoS

certs :: !VssCertificatesMap

Not expired certificates

whenInsMap :: !(HashMap StakeholderId EpochOrSlot)

Slot when certs was inserted. It is needed for deletion from insSlotSet (by StakeholderId).

whenInsSet :: !(Set (EpochOrSlot, StakeholderId))

Set of pairs (insertion slot, address hash) Every element of insSlotSet <= lastKnownEoS

whenExpire :: !(Set (EpochOrSlot, StakeholderId))

Set of pairs (expiry slot, address hash). Expiry slot is first slot when certificate expires. Pairs are sorted by expiry slot (in increasing order, so the oldest certificate is first element).

expiredCerts :: !(Set (EpochOrSlot, (StakeholderId, EpochOrSlot, VssCertificate)))

Set of expired certs for current lastKnownEoS. We store only certificates which expried no earlier than in previous epoch. Set (full expired slot, (id, insertion slot, cert))

empty :: VssCertData

Create empty VssCertData.

insert :: StakeholderId -> VssCertificate -> VssCertData -> VssCertData

Remove old certificate corresponding to the specified StakeholderId and insert new certificate.

lookup :: StakeholderId -> VssCertData -> Maybe VssCertificate

Lookup certificate corresponding to the specified StakeholderId.

lookupExpiryEpoch :: StakeholderId -> VssCertData -> Maybe EpochIndex

Lookup expiry epoch of certificate corresponding to the specified StakeholderId.

setLastKnownSlot :: SlotId -> VssCertData -> VssCertData

setLastKnownEoS :: EpochOrSlot -> VssCertData -> VssCertData

Set last known slot (lks). 1. If new lks is bigger than lastKnownEoS then some expired certificates will be removed. 2. If new lks is less than lastKnownEoS then some inserted after nlks certificates will be removed (and whenExpire) also will be updated.

keys :: VssCertData -> [StakeholderId]

Ids of stakeholders issued certificates.

member :: StakeholderId -> VssCertData -> Bool

Return True if the specified address hash is present in the map, False otherwise.

expiryEpoch :: VssCertificate -> EpochIndex

Convert expiry epoch of certificate to FlatSlotId.

expiryEoS :: VssCertificate -> EpochOrSlot

Functions which delete certificates. Be careful

delete :: StakeholderId -> VssCertData -> VssCertData

Delete certificate corresponding to the specified StakeholderId. This function is dangerous, because after using it you can't rollback deleted certificates. Use carefully.

difference :: VssCertData -> HashMap StakeholderId a -> VssCertData

This function is dangerous, because after you using it you can't rollback deleted certificates. Use carefully.

filter :: (StakeholderId -> Bool) -> VssCertData -> VssCertData

Filtering the certificates. This function is dangerous, because after you using it you can't rollback deleted certificates. Use carefully.