Pos.Ssc.Type

Description

Global export of the SSC type. Exported in order to be able to switch to different type-level implementations.

Documentation

type SscType = SscGodTossing

We can also use CPP flags that I proposed that would enable switching between implementations without changing the code.