Pos.Statistics

Documentation

statlogCountEvent :: (MonadStats m, StatLabel l, EntryType l ~ CountStat) => l -> Word64 -> m ()

Log count statistic event.

statlogValueEvent :: (MonadStats m, StatLabel l, EntryType l ~ ValueStat) => l -> Double -> m ()

Log value statistic event.

class Monad m => MonadStats m where

MonadStats is a monad which has methods for stats collecting

Methods

statLog :: StatLabel l => l -> EntryType l -> m ()

resetStat :: StatLabel l => l -> m ()

statLog :: (MonadTrans t, MonadStats m', t m' ~ m, StatLabel l) => l -> EntryType l -> m ()

resetStat :: (MonadTrans t, MonadStats m', t m' ~ m, StatLabel l) => l -> m ()

logStatM :: StatLabel l => l -> m (EntryType l) -> m ()

Default convenience method, which we can override (to truly do nothing in NoStatsT, for example)

type NoStatsT = TaggedTrans NoStatsTag IdentityT

Stats wrapper for collecting statistics without collecting it.

getNoStatsT :: NoStatsT m a -> m a

type StatsT = ReaderT StatsMap

Statistics wrapper around some monadic action to collect statistics during execution of this action. Used in benchmarks.

runStatsT :: MonadIO m => StatsT m a -> m a

runStatsT' :: StatsMap -> StatsT m a -> m a

getStatsMap :: Monad m => StatsT m StatsMap

type StatEntry e = (Monoid e, FullySerializable e)

Stat entry is a simple counter or a structure for aggregating statistical data about real value

class (FullySerializable l, Buildable l, Hashable l, StatEntry (EntryType l)) => StatLabel l where

StatLabel is some datatype which determines a single stat

Minimal complete definition

labelName, toJLEvent

Associated Types

type EntryType l :: *

Methods

labelName :: Proxy l -> Text

toJLEvent :: l -> EntryType l -> JLEvent

newtype CountStat

Counter for specified statistics.

Constructors

CountStat

Fields

getCounter :: Word64

data ValueStat

Value for specified collected statistic.

Constructors

ValueStat

Fields

valueCount :: !Word64

valueMin :: !Double

valueMax :: !Double

valueSum :: !Double

data StatProcessTx

Singleton data type to represent collected statistics about processes Tx's.

Constructors

StatProcessTx