Pos.Txp.Core

Documentation

addrBelongsTo :: TxOutAux -> Address -> Bool

A predicate for TxOutAux which checks whether given address belongs to it.

mkTxProof :: TxPayload -> TxProof

Construct TxProof which proves given TxPayload.

txInToPair :: TxIn -> (TxId, Word32)

Make a pair from TxIn.

txOutStake :: TxOutAux -> TxOutDistribution

Use this function if you need to know how a TxOut distributes stake (e.g. for the purpose of running follow-the-satoshi).

flattenTxPayload :: TxPayload -> [TxAux]

Convert TxPayload into a flat list of TxAuxs.

topsortTxs :: forall a. Eq a => (a -> WithHash Tx) -> [a] -> Maybe [a]

Does topological sort on things that contain transactions – e.g. can be used both for sorting [Tx] and [(Tx, TxWitness)].

(Backwards dfs from every node with reverse visiting order recording. Returns nothing on loop encountered. Return order is head-first.)

type TxId = Hash Tx

Represents transaction identifier as Hash of Tx.

Witness

data TxInWitness

A witness for a single input.

Constructors

PkWitness

Fields

twKey :: !PublicKey

twSig :: !TxSig

ScriptWitness

Fields

twValidator :: !Script

twRedeemer :: !Script

RedeemWitness

Fields

twRedeemKey :: !RedeemPublicKey

twRedeemSig :: !(RedeemSignature TxSigData)

UnknownWitnessType !Word8 !ByteString

type TxWitness = Vector TxInWitness

A witness is a proof that a transaction is allowed to spend the funds it spends (by providing signatures, redeeming scripts, etc). A separate proof is provided for each input.

type TxOutDistribution = [(StakeholderId, Coin)]

Distribution of "fake" stake that follow-the-satoshi would use for a particular transaction output. Sum of coins in the list should be the same as txOutValue of corresponding output.

newtype TxDistribution

Distribution of “fake” stake that follow-the-satoshi would use for a particular transaction. Length of stored list must be same as length of _txOutputs of corresponding transaction.

Constructors

TxDistribution

Fields

getTxDistribution :: NonEmpty TxOutDistribution

data TxSigData

Data that is being signed when creating a TxSig.

Constructors

TxSigData

Fields

txSigInput :: !TxIn

Input that we're signing (i.e. our signature certifies that we own funds referenced by this input)

txSigOutsHash :: !(Hash (NonEmpty TxOut))

Outputs of the transaction (i.e. our signature certifies that we actually want the funds to go to these particular outputs)

txSigDistrHash :: !(Hash TxDistribution)

Distribution of the transaction

type TxSig = Signature TxSigData

Signature of addrId.

Tx parts

data TxIn

Transaction input.

Constructors

TxIn

Fields

txInHash :: !TxId

Which transaction's output is used

txInIndex :: !Word32

Index of the output in transaction's outputs

data TxOut

Transaction output.

Constructors

TxOut

Fields

txOutAddress :: !Address

txOutValue :: !Coin

data TxOutAux

Transaction output and auxilary data corresponding to it. [CSL-366] Add more data.

Constructors

TxOutAux

Fields

toaOut :: !TxOut

Tx output

toaDistr :: !TxOutDistribution

Stake distribution associated with output

type TxAttributes = Attributes ()

Represents transaction attributes: map from 1-byte integer to arbitrary-type value. To be used for extending transaction with new fields via softfork.

Tx

data Tx

Transaction.

NB: transaction witnesses are stored separately.

Constructors

UnsafeTx

Fields

_txInputs :: !(NonEmpty TxIn)

Inputs of transaction.

_txOutputs :: !(NonEmpty TxOut)

Outputs of transaction.

_txAttributes :: !TxAttributes

Attributes of transaction

type TxAux = (Tx, TxWitness, TxDistribution)

Transaction + auxiliary data

mkTx :: MonadFail m => NonEmpty TxIn -> NonEmpty TxOut -> TxAttributes -> m Tx

Create valid Tx or fail. Verify inputs and outputs are non empty; have enough coins.

txInputs :: Lens' Tx (NonEmpty TxIn)

txOutputs :: Lens' Tx (NonEmpty TxOut)

txAttributes :: Lens' Tx TxAttributes

txF :: Bi Tx => Format r (Tx -> r)

Specialized formatter for Tx.

txaF :: Bi Tx => Format r (TxAux -> r)

Specialized formatter for Tx with auxiliary data

Payload and proof

data TxProof

Constructors

TxProof

Fields

txpNumber :: !Word32

txpRoot :: !(MerkleRoot Tx)

txpWitnessesHash :: !(Hash [TxWitness])

txpDistributionsHash :: !(Hash [TxDistribution])

data TxPayload

Payload of Txp component which is part of main block. Constructor is unsafe, because it lets one create invalid payload, for example with different number of transactions and witnesses.

Constructors

UnsafeTxPayload

Fields

_txpTxs :: !(MerkleTree Tx)

Transactions are the main payload.

_txpWitnesses :: ![TxWitness]

Transaction witnesses. Invariant: there are as many witnesses as there are transactions in the block. This is checked during deserialisation. We can't put them into the same Merkle tree with transactions, as the whole point of segwit is to separate transactions and witnesses.

TODO: should they be put into a separate Merkle tree or left as a list?

_txpDistributions :: ![TxDistribution]

Distributions for P2SH addresses in transaction outputs. * length mbTxAddrDistributions == length mbTxs * i-th element is Just if at least one output of i-th transaction is P2SH * n-th element of i-th element is Just if n-th output of i-th transaction is P2SH Ask @neongreen if you don't understand wtf is going on. Basically, address distributions are needed so that (potential) receivers of P2SH funds would count as stakeholders.

mkTxPayload :: (Bi Tx, MonadFail m) => [(Tx, TxWitness, TxDistribution)] -> m TxPayload

Smart constructor of TxPayload which can fail if some invariants are violated.

Currently there are two invariants: • number of txs must be same as number of witnesses and number of distributions; • each distribution must have the same number of elements as number of outputs in corresponding transaction.

txpTxs :: Lens' TxPayload (MerkleTree Tx)

txpWitnesses :: Lens' TxPayload [TxWitness]

txpDistributions :: Lens' TxPayload [TxDistribution]

Undo

type TxUndo = NonEmpty TxOutAux

Particular undo needed for transactions

type TxpUndo = [TxUndo]