Pos.Types

Description

Re-export of Pos.Types.* + binary instances

Documentation

module Pos.Types.Arbitrary

module Pos.Types.Address

module Pos.Types.Block

module Pos.Types.Coin

Coin

data Coin

Coin is the least possible unit of currency.

data CoinPortion

CoinPortion is some portion of Coin, it must be in [0 .. coinPortionDenominator]. Main usage of it is multiplication with Coin. Usually it's needed to determine some threshold expressed as portion of total stake.

coinF :: Format r (Coin -> r)

Coin formatter which restricts type.

unsafeGetCoin :: Coin -> Word64

Unwraps Coin. It's called “unsafe” so that people wouldn't use it willy-nilly if they want to sum coins or something. It's actually safe.

getCoinPortion :: CoinPortion -> Word64

mkCoin :: Word64 -> Coin

Make Coin from Word64.

coinPortionDenominator :: Word64

Denominator used by CoinPortion.

mkCoinPortion :: MonadFail m => Word64 -> m CoinPortion

Make CoinPortion from Word64 checking whether it is not greater than coinPortionDenominator.

unsafeCoinPortionFromDouble :: Double -> CoinPortion

Make CoinPortion from Double. Caller must ensure that value is in [0 .. 1]. Internally CoinPortion stores Word64 which is divided by coinPortionDenominator to get actual value. So some rounding may take place.

Slotting

newtype EpochIndex

Index of epoch.

Constructors

EpochIndex

Fields

getEpochIndex :: Word64

class HasEpochIndex a where

Class for something that has EpochIndex.

Minimal complete definition

epochIndexL

Methods

epochIndexL :: Lens' a EpochIndex

type FlatSlotId = Word64

FlatSlotId is a flat version of SlotId

newtype LocalSlotIndex

Index of slot inside a concrete epoch.

Constructors

LocalSlotIndex

Fields

getSlotIndex :: Word16

data SlotId

Slot is identified by index of epoch and local index of slot in this epoch. This is a global index

Constructors

SlotId

Fields

siEpoch :: !EpochIndex

siSlot :: !LocalSlotIndex

newtype EpochOrSlot

Represents SlotId or EpochIndex. Useful because genesis blocks have only EpochIndex, while main blocks have SlotId.

Constructors

EpochOrSlot

Fields

unEpochOrSlot :: Either EpochIndex SlotId

class HasEpochOrSlot a where

Methods

_getEpochOrSlot :: a -> Either EpochIndex SlotId

getEpochOrSlot :: a -> EpochOrSlot

epochOrSlotG :: Getter a EpochOrSlot

newtype Timestamp

Timestamp is a number which represents some point in time. It is used in MonadSlots and its meaning is up to implementation of this type class. The only necessary knowledge is that difference between timestamps is microsecond. Hence underlying type is Microsecond.

Constructors

Timestamp

Fields

getTimestamp :: Microsecond

slotIdF :: Format r (SlotId -> r)

Specialized formatter for SlotId.

epochOrSlot :: (EpochIndex -> a) -> (SlotId -> a) -> EpochOrSlot -> a

Apply one of the function depending on content of EpochOrSlot.

Address

data Address

Address is where you can send coins.

Constructors

PubKeyAddress

Fields

addrKeyHash :: !(AddressHash PublicKey)

addrPkAttributes :: !(Attributes AddrPkAttrs)

ScriptAddress

Fields

addrScriptHash :: !(AddressHash Script)

UnknownAddressType !Word8 !ByteString

data AddrPkAttrs

Constructors

AddrPkAttrs

Fields

addrPkDerivationPath :: Maybe [Word32]

type AddressHash = AbstractHash Blake2s_224

type StakeholderId = AddressHash PublicKey

Stakeholder identifier (stakeholders are identified by their public keys)

ChainDifficulty

newtype ChainDifficulty

Chain difficulty represents necessary effort to generate a chain. In the simplest case it can be number of blocks in chain.

Constructors

ChainDifficulty

Fields

getChainDifficulty :: Word64

class HasDifficulty a where

Type class for something that has ChainDifficulty.

Minimal complete definition

difficultyL

Methods

difficultyL :: Lens' a ChainDifficulty

Version

newtype ApplicationName

Constructors

ApplicationName

Fields

getApplicationName :: Text

data BlockVersion

Communication protocol version.

Constructors

BlockVersion

Fields

bvMajor :: !Word16

bvMinor :: !Word16

bvAlt :: !Word8

type NumSoftwareVersion = Word32

Numeric software version associated with ApplicationName.

data SoftwareVersion

Software version.

Constructors

SoftwareVersion

Fields

svAppName :: !ApplicationName

svNumber :: !NumSoftwareVersion

HeaderHash related types and functions

data BlockHeaderStub

type HeaderHash = Hash BlockHeaderStub

Hash of block header. This should be Hash (BlockHeader ssc) but we don't want to have ssc in HeaderHash type.

class HasHeaderHash a where

Class for something that has HeaderHash.

Minimal complete definition

headerHash

Methods

headerHash :: a -> HeaderHash

headerHashG :: Getter a HeaderHash

headerHashF :: Format r (HeaderHash -> r)

Specialized formatter for HeaderHash.

module Pos.Types.Slotting

module Pos.Types.Timestamp

module Pos.Types.Tx

module Pos.Types.Types

module Pos.Types.Utxo

module Pos.Types.Version