Pos.Types.Arbitrary

Description

Arbitrary instances for core types for using in tests and benchmarks

Documentation

newtype BadSigsTx

Ill-formed Tx with bad signatures.

Constructors

BadSigsTx

Fields

getBadSigsTx :: NonEmpty ((Tx, TxDistribution), TxIn, TxOutAux, TxInWitness)

newtype CoinPairOverflowSum

This datatype has two coins that will always overflow when added. It is used in tests to make sure addition raises the appropriate exception when this happens.

Constructors

TwoCoinsSum

Fields

get2CSum :: (Coin, Coin)

newtype CoinPairOverflowSub

This datatype has two coins that will always underflow when subtracted. It is used in tests to make sure subtraction raises the appropriate exception when this happens.

Constructors

TwoCoinsSub

Fields

get2CSub :: (Coin, Coin)

newtype CoinPairOverflowMul

This datatype has a Coin and an Integer that will always overflow when multiplied. It is used in tests to make sure multiplication raises the appropriate exception when this happens.

Constructors

TwoCoinsM

Fields

get2CMul :: (Coin, Integer)

newtype DoubleInZeroToOneRange

Another wrapper over Double. Its Arbitrary instance guarantees the Double inside can always be safely turned into a CoinPortion. Used in tests to ensure converting a valid Double to/from CoinPortion works properly.

Constructors

DoubleInRange

Fields

getDoubleInRange :: Double

newtype IntegerToCoinNoOverflow

This datatype has an Integer that will never overflow when turned into a Coin. Useful for testing that conversion between valid Integers and Coins works properly.

Constructors

Integer

Fields

getInteger :: Integer

newtype IntegerToCoinOverflow

IntegerToCoinOverflow is a wrapped over Integer. Its Arbitrary instance makes it so that these integers will always overflow when converted into a Coin. Used in tests to make sure an exception is raised when there is an attempt to turn an excessively large Integer into a Coin.

Constructors

LargeInteger

Fields

getLargeInteger :: Integer

newtype GoodTx

Well-formed transaction Tx.

Constructors

GoodTx

Fields

getGoodTx :: NonEmpty ((Tx, TxDistribution), TxIn, TxOutAux, TxInWitness)

newtype LessThanZeroOrMoreThanOne

A wrapper over Double. Its Arbitrary instance ensures the Double within can never be converted into a CoinPortion without an exception being raised. Used in tests to safeguard that converting an invalid Double to a CoinPortion always raised an exception.

Constructors

BadCoinPortion

Fields

getDouble :: Double

newtype SafeCoinPairMul

This datatype has a Coin and an Integer that will always overflow when multiplied. It is used to make sure coin multiplication by an integer raises the appropriate exception when this happens.

Constructors

CoinPairMul

Fields

getPairMul :: (Coin, Integer)

newtype SafeCoinPairSum

This datatype has two coins that will never overflow when added. It is therefore safe to add them. Useful in tests to ensure adding two coins whose sum is a valid Coin always works.

Constructors

CoinPairSum

Fields

getPairSum :: (Coin, Coin)

newtype SafeCoinPairSub

This datatype has two coins that will never underflow when subtracted. It is therefore safe to subtract them. Useful in tests to show that two coins whose subtraction does not underflow always works.

Constructors

CoinPairSub

Fields

getPairSub :: (Coin, Coin)

newtype SafeWord

A wrapper over Word64. Its Arbitrary instance guarantees the Word64 inside can always be safely converted into CoinPortion. Used in tests to ensure converting a valid Word64 to/from CoinPortion works properly.

Constructors

SafeWord

Fields

getSafeWord :: Word64

newtype SmallBadSigsTx

Constructors

SmallBadSigsTx BadSigsTx

newtype SmallHashMap

Constructors

SmallHashMap (HashMap PublicKey (HashMap PublicKey (AsBinary Share)))

newtype SmallGoodTx

Constructors

SmallGoodTx GoodTx

Orphan instances

Random EpochIndex

Methods

randomR :: RandomGen g => (EpochIndex, EpochIndex) -> g -> (EpochIndex, g)

random :: RandomGen g => g -> (EpochIndex, g)

randomRs :: RandomGen g => (EpochIndex, EpochIndex) -> g -> [EpochIndex]

randoms :: RandomGen g => g -> [EpochIndex]

randomRIO :: (EpochIndex, EpochIndex) -> IO EpochIndex

randomIO :: IO EpochIndex

Random LocalSlotIndex

Methods

randomR :: RandomGen g => (LocalSlotIndex, LocalSlotIndex) -> g -> (LocalSlotIndex, g)

random :: RandomGen g => g -> (LocalSlotIndex, g)

randomRs :: RandomGen g => (LocalSlotIndex, LocalSlotIndex) -> g -> [LocalSlotIndex]

randoms :: RandomGen g => g -> [LocalSlotIndex]

randomRIO :: (LocalSlotIndex, LocalSlotIndex) -> IO LocalSlotIndex

randomIO :: IO LocalSlotIndex

Arbitrary Address

Methods

arbitrary :: Gen Address

shrink :: Address -> [Address]

Arbitrary ChainDifficulty

Methods

arbitrary :: Gen ChainDifficulty

shrink :: ChainDifficulty -> [ChainDifficulty]

Arbitrary BlockVersion

Methods

arbitrary :: Gen BlockVersion

shrink :: BlockVersion -> [BlockVersion]

Arbitrary ApplicationName

Methods

arbitrary :: Gen ApplicationName

shrink :: ApplicationName -> [ApplicationName]

Arbitrary SoftwareVersion

Methods

arbitrary :: Gen SoftwareVersion

shrink :: SoftwareVersion -> [SoftwareVersion]

Arbitrary SharedSeed

Methods

arbitrary :: Gen SharedSeed

shrink :: SharedSeed -> [SharedSeed]

Arbitrary Coin

Methods

arbitrary :: Gen Coin

shrink :: Coin -> [Coin]

Arbitrary CoinPortion

Methods

arbitrary :: Gen CoinPortion

shrink :: CoinPortion -> [CoinPortion]

Arbitrary EpochIndex

Methods

arbitrary :: Gen EpochIndex

shrink :: EpochIndex -> [EpochIndex]

Arbitrary LocalSlotIndex

Methods

arbitrary :: Gen LocalSlotIndex

shrink :: LocalSlotIndex -> [LocalSlotIndex]

Arbitrary SlotId

Methods

arbitrary :: Gen SlotId

shrink :: SlotId -> [SlotId]

Arbitrary EpochOrSlot

Methods

arbitrary :: Gen EpochOrSlot

shrink :: EpochOrSlot -> [EpochOrSlot]

Arbitrary Script

Methods

arbitrary :: Gen Script

shrink :: Script -> [Script]

Arbitrary Timestamp

Methods

arbitrary :: Gen Timestamp

shrink :: Timestamp -> [Timestamp]

Arbitrary Microsecond

Methods

arbitrary :: Gen Microsecond

shrink :: Microsecond -> [Microsecond]

Arbitrary Millisecond

Methods

arbitrary :: Gen Millisecond

shrink :: Millisecond -> [Millisecond]

Arbitrary Tx
Arbitrary transactions generated from this instance will only be valid with regards to mxTx

Methods

arbitrary :: Gen Tx

shrink :: Tx -> [Tx]

Arbitrary TxOutAux

Methods

arbitrary :: Gen TxOutAux

shrink :: TxOutAux -> [TxOutAux]

Arbitrary TxOut

Methods

arbitrary :: Gen TxOut

shrink :: TxOut -> [TxOut]

Arbitrary TxIn

Methods

arbitrary :: Gen TxIn

shrink :: TxIn -> [TxIn]

Arbitrary TxDistribution

Methods

arbitrary :: Gen TxDistribution

shrink :: TxDistribution -> [TxDistribution]

Arbitrary TxInWitness

Methods

arbitrary :: Gen TxInWitness

shrink :: TxInWitness -> [TxInWitness]

Arbitrary TxSigData

Methods

arbitrary :: Gen TxSigData

shrink :: TxSigData -> [TxSigData]

Arbitrary TxProof

Methods

arbitrary :: Gen TxProof

shrink :: TxProof -> [TxProof]

Arbitrary (MerkleRoot Tx)

Methods

arbitrary :: Gen (MerkleRoot Tx)

shrink :: MerkleRoot Tx -> [MerkleRoot Tx]

Arbitrary (MerkleTree Tx)

Methods

arbitrary :: Gen (MerkleTree Tx)

shrink :: MerkleTree Tx -> [MerkleTree Tx]

Arbitrary a_6989586621679177033 => Arbitrary (UnsignedVarInt a_6989586621679177033)

Methods

arbitrary :: Gen (UnsignedVarInt a_6989586621679177033)

shrink :: UnsignedVarInt a_6989586621679177033 -> [UnsignedVarInt a_6989586621679177033]

Arbitrary a_6989586621679177041 => Arbitrary (SignedVarInt a_6989586621679177041)

Methods

arbitrary :: Gen (SignedVarInt a_6989586621679177041)

shrink :: SignedVarInt a_6989586621679177041 -> [SignedVarInt a_6989586621679177041]

Arbitrary a_6989586621679177049 => Arbitrary (FixedSizeInt a_6989586621679177049)

Methods

arbitrary :: Gen (FixedSizeInt a_6989586621679177049)

shrink :: FixedSizeInt a_6989586621679177049 -> [FixedSizeInt a_6989586621679177049]