Pos.Types.Block.Class

Documentation

class Blockchain p where

Blockchain type class generalizes some functionality common for different blockchains.

Minimal complete definition

mkBodyProof, verifyBBlock

Associated Types

data BodyProof p :: *

Proof of data stored in the body. Ensures immutability.

data ConsensusData p :: *

Consensus data which can be used to check consensus properties.

type ExtraHeaderData p :: *

Whatever extra data.

type BBlockHeader p :: *

Block header used in this blockchain.

type BHeaderHash p :: *

Hash of BBlockHeader. This is something like Hash (BBlockHeader p).

data Body p :: *

Body contains payload and other heavy data.

type ExtraBodyData p :: *

Whatever extra data.

type BBlock p :: *

Block used in this blockchain.

Methods

mkBodyProof :: Body p -> BodyProof p

checkBodyProof :: Body p -> BodyProof p -> Bool

checkBodyProof :: Eq (BodyProof p) => Body p -> BodyProof p -> Bool

verifyBBlock :: GenericBlock p -> Either Text ()

data GenericBlockHeader b

Header of block contains some kind of summary. There are various benefits which people get by separating header from other data.

Constructors

GenericBlockHeader

Fields

_gbhPrevBlock :: !(BHeaderHash b)

Pointer to the header of the previous block.

_gbhBodyProof :: !(BodyProof b)

Proof of body.

_gbhConsensus :: !(ConsensusData b)

Consensus data to verify consensus algorithm.

_gbhExtra :: !(ExtraHeaderData b)

Any extra data.

data GenericBlock b

In general Block consists of header and body. It may contain extra data as well.

Constructors

GenericBlock

Fields

_gbHeader :: !(GenericBlockHeader b)

_gbBody :: !(Body b)

_gbExtra :: !(ExtraBodyData b)

Lenses

class HasPrevBlock s where

Minimal complete definition

prevBlockL

Methods

prevBlockL :: Lens' s HeaderHash

gbBody :: forall b. Lens' (GenericBlock b) (Body b)

gbHeader :: forall b. Lens' (GenericBlock b) (GenericBlockHeader b)

gbExtra :: forall b. Lens' (GenericBlock b) (ExtraBodyData b)

gbBodyProof :: Lens' (GenericBlock b) (BodyProof b)

Lens from GenericBlock to BodyProof.

gbhConsensus :: forall b. Lens' (GenericBlockHeader b) (ConsensusData b)

gbhExtra :: forall b. Lens' (GenericBlockHeader b) (ExtraHeaderData b)

gbhPrevBlock :: forall b. Lens' (GenericBlockHeader b) (BHeaderHash b)

gbhBodyProof :: forall b. Lens' (GenericBlockHeader b) (BodyProof b)