Pos.Types.Block.Types

Documentation

data MainBlockchain ssc

Represents blockchain consisting of main blocks, i. e. blocks with transactions and MPC messages.

type MainBlockHeader ssc = GenericBlockHeader (MainBlockchain ssc)

Header of generic main block.

newtype MainExtraBodyData

Represents main block extra data

Constructors

MainExtraBodyData

Fields

_mebAttributes :: BlockBodyAttributes

data MainExtraHeaderData

Represents main block header extra data

Constructors

MainExtraHeaderData

Fields

_mehBlockVersion :: !BlockVersion

Version of block.

_mehSoftwareVersion :: !SoftwareVersion

Software version.

_mehAttributes :: !BlockHeaderAttributes

Header attributes

type BlockHeaderAttributes = Attributes ()

Represents main block header attributes: map from 1-byte integer to arbitrary-type value. To be used for extending header with new fields via softfork.

type BlockBodyAttributes = Attributes ()

Represents main block body attributes: map from 1-byte integer to arbitrary-type value. To be used for extending block with new fields via softfork.

type BiSsc ssc = (Ssc ssc, Bi (GenericBlockHeader (GenesisBlockchain ssc)), Bi (GenericBlockHeader (MainBlockchain ssc)))

Ssc w/ buildable blockchain

data BlockSignature ssc

Signature of the block. Can be either regular signature from the issuer or delegated signature having a constraint on epoch indices (it means the signature is valid only if block's slot id has epoch inside the constrained interval).

Constructors

BlockSignature (Signature (MainToSign ssc))

BlockPSignatureEpoch (ProxySigLight (MainToSign ssc))

BlockPSignatureSimple (ProxySigHeavy (MainToSign ssc))

data MainToSign ssc

Data to be signed in main block.

Constructors

MainToSign

Fields

_msHeaderHash :: !HeaderHash

_msBodyProof :: !(BodyProof (MainBlockchain ssc))

_msSlot :: !SlotId

_msChainDiff :: !ChainDifficulty

_msExtraHeader :: !MainExtraHeaderData

type MainBlock ssc = GenericBlock (MainBlockchain ssc)

MainBlock is a block with transactions and MPC messages. It's the main part of our consensus algorithm.

data GenesisBlockchain ssc

Represents blockchain consisting of genesis blocks. Genesis block doesn't have any special payload and is not strictly necessary. However, it is good idea to store list of leaders explicitly, because calculating it may be expensive operation. For example, it is useful for SPV-clients.

type GenesisBlockHeader ssc = GenericBlockHeader (GenesisBlockchain ssc)

Header of Genesis block.

type GenesisBlock ssc = GenericBlock (GenesisBlockchain ssc)

Genesis block parametrized by GenesisBlockchain.

data GenesisExtraBodyData

Represents genesis block header extra data

Constructors

GenesisExtraBodyData

Fields

_gebAttributes :: !GenesisBodyAttributes

Header attributes

type GenesisBodyAttributes = Attributes ()

Represents genesis block header attributes.

data GenesisExtraHeaderData

Represents genesis block header extra data

Constructors

GenesisExtraHeaderData

Fields

_gehAttributes :: !GenesisHeaderAttributes

Header attributes

type GenesisHeaderAttributes = Attributes ()

Represents genesis block header attributes.

type BlockHeader ssc = Either (GenesisBlockHeader ssc) (MainBlockHeader ssc)

Either header of ordinary main block or genesis block.

type Block ssc = Either (GenesisBlock ssc) (MainBlock ssc)

Block.

type BiHeader ssc = Bi (BlockHeader ssc)

mehBlockVersion :: Lens' MainExtraHeaderData BlockVersion

mehSoftwareVersion :: Lens' MainExtraHeaderData SoftwareVersion

mehAttributes :: Lens' MainExtraHeaderData BlockHeaderAttributes

mebAttributes :: Iso' MainExtraBodyData BlockBodyAttributes

gehAttributes :: Iso' GenesisExtraHeaderData GenesisHeaderAttributes

gebAttributes :: Iso' GenesisExtraBodyData GenesisBodyAttributes