Pos.Types.Timestamp

Description

Slotting functionality.

Documentation

newtype Timestamp

Timestamp is a number which represents some point in time. It is used in MonadSlots and its meaning is up to implementation of this type class. The only necessary knowledge is that difference between timestamps is microsecond. Hence underlying type is Microsecond.

Constructors

Timestamp

Fields

getTimestamp :: Microsecond

timestampF :: Format r (Timestamp -> r)

Specialized formatter for Timestamp data type.

Orphan instances

Read Timestamp

Methods

readsPrec :: Int -> ReadS Timestamp

readList :: ReadS [Timestamp]

readPrec :: ReadPrec Timestamp

readListPrec :: ReadPrec [Timestamp]

Show Timestamp

Methods

showsPrec :: Int -> Timestamp -> ShowS

show :: Timestamp -> String

showList :: [Timestamp] -> ShowS

Buildable Timestamp

Methods

build :: Timestamp -> Builder