Pos.Types.Tx

Description

Transaction related functions.

Documentation

verifyTxAlone :: Tx -> VerificationRes

Verify that Tx itself is correct. Most likely you will also want to verify that inputs are legal, signed properly and have enough coins; verifyTxAlone doesn't do that.

data VTxGlobalContext

Global context data needed for script execution -- is same for the whole transaction. VT stands for "Verify Tx". Currently empty.

Constructors

VTxGlobalContext

data VTxLocalContext

Local context data for scripts -- differs per input.

Constructors

VTxLocalContext

Fields

vtlTxOut :: TxOutAux

Transaction output

verifyTx

Arguments

:: Monad m

=> Bool
Verify that tx itself is correct

-> Bool
Verify that script & address versions in tx are known

-> VTxGlobalContext

-> (TxIn -> m (Maybe VTxLocalContext))

-> TxAux

-> m (Either [Text] [TxOutAux])

CHECK: Verify Tx correctness using magic function which resolves input into Address and Coin. It optionally does checks from verifyTxAlone and also the following checks:

sum of inputs >= sum of outputs;

every input has a proper witness verifying that input;

script witnesses have matching script versions;

every input is a known unspent output.

Note that verifyTx doesn't attempt to verify scripts with versions higher than maximum script version we can handle. That's because we want blocks with such transactions to be accepted (to avoid hard forks). However, we won't include such transactions into blocks when we're creating a block.

verifyTxPure

Arguments

:: Bool
Verify that tx itself is correct

-> Bool
Verify that script & address versions in tx are known

-> VTxGlobalContext

-> (TxIn -> Maybe VTxLocalContext)

-> TxAux

-> Either [Text] [TxOutAux]

topsortTxs :: forall a. (a -> WithHash Tx) -> [a] -> Maybe [a]

Does topological sort on things that contain transactions – e.g. can be used both for sorting [Tx] and [(Tx, TxWitness)].

(Backwards dfs from every node with reverse visiting order recording. Returns nothing on loop encountered. Return order is head-first.)