Pos.Types.Types

Description

Definitions of the most fundamental types.

Documentation

type TxAttributes = Attributes ()

Represents transaction attributes: map from 1-byte integer to arbitrary-type value. To be used for extending transaction with new fields via softfork.

data TxInWitness

A witness for a single input.

Constructors

PkWitness

Fields

twKey :: PublicKey

twSig :: TxSig

ScriptWitness

Fields

twValidator :: Script

twRedeemer :: Script

type TxWitness = Vector TxInWitness

A witness is a proof that a transaction is allowed to spend the funds it spends (by providing signatures, redeeming scripts, etc). A separate proof is provided for each input.

newtype TxDistribution

Distribution of “fake” stake that follow-the-satoshi would use for a particular transaction.

Constructors

TxDistribution

Fields

getTxDistribution :: [[(StakeholderId, Coin)]]

type TxSigData = (TxId, Word32, Hash [TxOut], Hash TxDistribution)

Data that is being signed when creating a TxSig.

type TxSig = Signature TxSigData

Signature of addrId.

type TxId = Hash Tx

Represents transaction identifier as Hash of Tx.

data TxIn

Transaction input.

Constructors

TxIn

Fields

txInHash :: !TxId

Which transaction's output is used

txInIndex :: !Word32

Index of the output in transaction's outputs

toPair :: TxIn -> (TxId, Word32)

Make pair from TxIn

data TxOut

Transaction output.

Constructors

TxOut

Fields

txOutAddress :: !Address

txOutValue :: !Coin

txOutStake :: TxOutAux -> [(StakeholderId, Coin)]

Use this function if you need to know how a TxOut distributes stake (e.g. for the purpose of running follow-the-satoshi).

data Tx

Transaction.

NB: transaction witnesses are stored separately.

Constructors

Tx

Fields

txInputs :: ![TxIn]

Inputs of transaction.

txOutputs :: ![TxOut]

Outputs of transaction.

txAttributes :: !TxAttributes

Attributes of transaction

_txInputs :: Lens' Tx [TxIn]

_txOutputs :: Lens' Tx [TxOut]

_txAttributes :: Lens' Tx TxAttributes

txF :: Format r (Tx -> r)

Specialized formatter for Tx.

txaF :: Bi Script => Format r (TxAux -> r)

Specialized formatter for Tx with auxiliary data

type TxAux = (Tx, TxWitness, TxDistribution)

Transaction + auxiliary data

type TxOutAux = (TxOut, [(StakeholderId, Coin)])

type Utxo = Map (TxId, Word32) TxOutAux

Unspent transaction outputs.

Transaction inputs are identified by (transaction ID, index in list of output) pairs.

formatUtxo :: Utxo -> Builder

Format Utxo map as json.

utxoF :: Format r (Utxo -> r)

Specialized formatter for Utxo.

type TxUndo = [[TxOutAux]]

Particular undo needed for transactions

newtype SharedSeed

This is a shared seed used for follow-the-satoshi. This seed is randomly generated by each party and eventually they agree on the same value.

Constructors

SharedSeed

Fields

getSharedSeed :: ByteString

type SlotLeaders = NonEmpty StakeholderId

NonEmpty list of slot leaders.

type ProxySigEpoch a = ProxySignature (EpochIndex, EpochIndex) a

Proxy signature used in csl -- holds a pair of epoch indices. Block is valid if it's epoch index is inside this range.

type ProxySKEpoch = ProxySecretKey (EpochIndex, EpochIndex)

Same alias for the proxy secret key (see ProxySigEpoch).

type ProxySigSimple a = ProxySignature () a

Simple proxy signature without ttl/epoch index constraints.

type ProxySKSimple = ProxySecretKey ()

Correspondent SK for no-ttl proxy signature scheme.

type ProxySKEither = Either ProxySKEpoch ProxySKSimple

Some proxy secret key.

class Blockchain p where

Blockchain type class generalizes some functionality common for different blockchains.

Minimal complete definition

mkBodyProof

Associated Types

data BodyProof p :: *

Proof of data stored in the body. Ensures immutability.

data ConsensusData p :: *

Consensus data which can be used to check consensus properties.

type ExtraHeaderData p :: *

Whatever extra data.

type BBlockHeader p :: *

Block header used in this blockchain.

type BHeaderHash p :: *

Hash of BBlockHeader. This is something like Hash (BBlockHeader p).

data Body p :: *

Body contains payload and other heavy data.

type ExtraBodyData p :: *

Whatever extra data.

type BBlock p :: *

Block used in this blockchain.

Methods

mkBodyProof :: Body p -> BodyProof p

checkBodyProof :: Body p -> BodyProof p -> Bool

checkBodyProof :: Eq (BodyProof p) => Body p -> BodyProof p -> Bool

data GenericBlockHeader b

Header of block contains some kind of summary. There are various benefits which people get by separating header from other data.

Constructors

GenericBlockHeader

Fields

_gbhPrevBlock :: !(BHeaderHash b)

Pointer to the header of the previous block.

_gbhBodyProof :: !(BodyProof b)

Proof of body.

_gbhConsensus :: !(ConsensusData b)

Consensus data to verify consensus algorithm.

_gbhExtra :: !(ExtraHeaderData b)

Any extra data.

data GenericBlock b

In general Block consists of header and body. It may contain extra data as well.

Constructors

GenericBlock

Fields

_gbHeader :: !(GenericBlockHeader b)

_gbBody :: !(Body b)

_gbExtra :: !(ExtraBodyData b)

data MainBlockchain ssc

Represents blockchain consisting of main blocks, i. e. blocks with transactions and MPC messages.

type MainBlockHeader ssc = GenericBlockHeader (MainBlockchain ssc)

Header of generic main block.

newtype MainExtraBodyData

Represents main block extra data

Constructors

MainExtraBodyData

Fields

_mebAttributes :: BlockBodyAttributes

data MainExtraHeaderData

Represents main block header extra data

Constructors

MainExtraHeaderData

Fields

_mehBlockVersion :: !BlockVersion

Version of block.

_mehSoftwareVersion :: !SoftwareVersion

Software version.

_mehAttributes :: !BlockHeaderAttributes

Header attributes

type BlockHeaderAttributes = Attributes ()

Represents main block header attributes: map from 1-byte integer to arbitrary-type value. To be used for extending header with new fields via softfork.

type BlockBodyAttributes = Attributes ()

Represents main block body attributes: map from 1-byte integer to arbitrary-type value. To be used for extending block with new fields via softfork.

type BiSsc ssc = (Ssc ssc, Bi (GenericBlockHeader (GenesisBlockchain ssc)), Bi (GenericBlockHeader (MainBlockchain ssc)))

Ssc w/ buildable blockchain

data BlockSignature ssc

Signature of the block. Can be either regular signature from the issuer or delegated signature having a constraint on epoch indices (it means the signature is valid only if block's slot id has epoch inside the constrained interval).

Constructors

BlockSignature (Signature (MainToSign ssc))

BlockPSignatureEpoch (ProxySigEpoch (MainToSign ssc))

BlockPSignatureSimple (ProxySigSimple (MainToSign ssc))

type MainToSign ssc = (HeaderHash, BodyProof (MainBlockchain ssc), SlotId, ChainDifficulty)

Data to be signed in main block.

type MainBlock ssc = GenericBlock (MainBlockchain ssc)

MainBlock is a block with transactions and MPC messages. It's the main part of our consensus algorithm.

data GenesisBlockchain ssc

Represents blockchain consisting of genesis blocks. Genesis block doesn't have any special payload and is not strictly necessary. However, it is good idea to store list of leaders explicitly, because calculating it may be expensive operation. For example, it is useful for SPV-clients.

type GenesisBlockHeader ssc = GenericBlockHeader (GenesisBlockchain ssc)

Header of Genesis block.

type GenesisBlock ssc = GenericBlock (GenesisBlockchain ssc)

Genesis block parametrized by GenesisBlockchain.

type BlockHeader ssc = Either (GenesisBlockHeader ssc) (MainBlockHeader ssc)

Either header of ordinary main block or genesis block.

type Block ssc = Either (GenesisBlock ssc) (MainBlock ssc)

Block.

HeaderHash related functions

blockHeaderHash :: BiHeader ssc => BlockHeader ssc -> HeaderHash

This function is required because type inference fails in attempts to hash only Right or Left.

Lenses

class HasPrevBlock s where

Class for something that has previous block (lens to Hash for this block).

Minimal complete definition

prevBlockL

Methods

prevBlockL :: Lens' s HeaderHash

blockHeader :: Getter (Block ssc) (BlockHeader ssc)

Lens from Block to BlockHeader.

This gives a “redundant constraint” message warning which will be fixed in lens-4.15 (not in LTS yet).

blockLeaderKey :: Lens' (MainBlock ssc) PublicKey

Lens from MainBlock to PublicKey.

blockLeaders :: Lens' (GenesisBlock ssc) SlotLeaders

Lens from GenesisBlock to SlotLeaders.

blockMpc :: Lens' (MainBlock ssc) (SscPayload ssc)

Lens from MainBlock to SscPayload.

blockSignature :: Lens' (MainBlock ssc) (BlockSignature ssc)

Lens from MainBlock to Signature.

blockSlot :: Lens' (MainBlock ssc) SlotId

Lens from MainBlock to SlotId.

blockTxs :: Lens' (MainBlock ssc) (MerkleTree Tx)

Lens from MainBlock to MerkleTree.

blockTxas :: Getter (MainBlock ssc) [TxAux]

Getter from MainBlock to a list of transactions together with auxiliary data.

blockProxySKs :: Lens' (MainBlock ssc) [ProxySKSimple]

Lens from MainBlock to ProxySKSimple list.

gbBody :: forall b. Lens' (GenericBlock b) (Body b)

gbBodyProof :: Lens' (GenericBlock b) (BodyProof b)

Lens from GenericBlock to BodyProof.

gbExtra :: forall b. Lens' (GenericBlock b) (ExtraBodyData b)

gbHeader :: forall b. Lens' (GenericBlock b) (GenericBlockHeader b)

gcdDifficulty :: Lens' (ConsensusData (GenesisBlockchain ssc)) ChainDifficulty

Lens for ChainDifficulty of GenesisBlockchain in ConsensusData.

gcdEpoch :: Lens' (ConsensusData (GenesisBlockchain ssc)) EpochIndex

Lens for EpochIndex of GenesisBlockchain in ConsensusData.

gbhConsensus :: forall b. Lens' (GenericBlockHeader b) (ConsensusData b)

gbhExtra :: forall b. Lens' (GenericBlockHeader b) (ExtraHeaderData b)

gbhPrevBlock :: forall b. Lens' (GenericBlockHeader b) (BHeaderHash b)

gbhBodyProof :: forall b. Lens' (GenericBlockHeader b) (BodyProof b)

getBlockHeader :: Block ssc -> BlockHeader ssc

Take BlockHeader from either GenesisBlock or MainBlock.

headerLeaderKey :: Lens' (MainBlockHeader ssc) PublicKey

Lens from MainBlockHeader to PublicKey.

headerSignature :: Lens' (MainBlockHeader ssc) (BlockSignature ssc)

Lens from MainBlockHeader to Signature.

headerSlot :: Lens' (MainBlockHeader ssc) SlotId

Lens from MainBlockHeader to SlotId.

mbMpc :: Lens' (Body (MainBlockchain ssc)) (SscPayload ssc)

Lens for SscPayload in main block body.

mbTxs :: Lens' (Body (MainBlockchain ssc)) (MerkleTree Tx)

Lens for transaction tree in main block body.

mbWitnesses :: Lens' (Body (MainBlockchain ssc)) [TxWitness]

Lens for witness list in main block body.

mbProxySKs :: Lens' (Body (MainBlockchain ssc)) [ProxySKSimple]

Lens for ProxySKs in main block body.

mbUpdatePayload :: Lens' (Body (MainBlockchain ssc)) UpdatePayload

Lens for UpdatePayload in main block body.

mcdSlot :: Lens' (ConsensusData (MainBlockchain ssc)) SlotId

Lens for SlotId of MainBlockchain in ConsensusData.

mcdLeaderKey :: Lens' (ConsensusData (MainBlockchain ssc)) PublicKey

Lens for PublicKey of MainBlockchain in ConsensusData.

mcdDifficulty :: Lens' (ConsensusData (MainBlockchain ssc)) ChainDifficulty

Lens for ChainDifficulty of MainBlockchain in ConsensusData.

mcdSignature :: Lens' (ConsensusData (MainBlockchain ssc)) (BlockSignature ssc)

Lens for Signature of MainBlockchain in ConsensusData.

mehBlockVersion :: Lens' MainExtraHeaderData BlockVersion

mehSoftwareVersion :: Lens' MainExtraHeaderData SoftwareVersion

mehAttributes :: Lens' MainExtraHeaderData BlockHeaderAttributes

mebAttributes :: Iso' MainExtraBodyData BlockBodyAttributes

Orphan instances

HasHeaderHash HeaderHash

Methods

headerHash :: HeaderHash -> HeaderHash

headerHashG :: Getter HeaderHash HeaderHash

HasEpochIndex SlotId

Methods

epochIndexL :: Lens' SlotId EpochIndex

(HasEpochOrSlot a, HasEpochOrSlot b) => HasEpochOrSlot (Either a b)

Methods

_getEpochOrSlot :: Either a b -> Either EpochIndex SlotId

getEpochOrSlot :: Either a b -> EpochOrSlot

epochOrSlotG :: Getter (Either a b) EpochOrSlot

(HasEpochIndex a, HasEpochIndex b) => HasEpochIndex (Either a b)

Methods

epochIndexL :: Lens' (Either a b) EpochIndex