Pos.Types.Utxo

Description

Utxo related operations.

Documentation

class Monad m => MonadUtxoRead m where

Methods

utxoGet :: TxIn -> m (Maybe TxOutAux)

utxoGet :: (MonadTrans t, MonadUtxoRead m', t m' ~ m) => TxIn -> m (Maybe TxOutAux)

class MonadUtxoRead m => MonadUtxo m where

Methods

utxoPut :: TxIn -> TxOutAux -> m ()

utxoPut :: (MonadTrans t, MonadUtxo m', t m' ~ m) => TxIn -> TxOutAux -> m ()

utxoDel :: TxIn -> m ()

utxoDel :: (MonadTrans t, MonadUtxo m', t m' ~ m) => TxIn -> m ()

applyTxToUtxo :: MonadUtxo m => WithHash Tx -> TxDistribution -> m ()

Remove unspent outputs used in given transaction, add new unspent outputs.

deleteTxIn :: TxIn -> Utxo -> Utxo

Delete given TxIn from Utxo if any.

findTxIn :: TxIn -> Utxo -> Maybe TxOutAux

Find transaction input in Utxo assuming it is valid.

verifyTxUtxo :: MonadUtxoRead m => Bool -> Bool -> TxAux -> m (Either [Text] [TxOutAux])

Verify single Tx using MonadUtxoRead as TxIn resolver.

verifyAndApplyTxs :: forall m. MonadUtxo m => Bool -> Bool -> [(WithHash Tx, TxWitness, TxDistribution)] -> m (Either [Text] TxUndo)

Verify transactions correctness with respect to Utxo applying them one-by-one. Note: transactions must be topsorted to pass check. Warning: this function may apply some transactions and fail eventually. Use it only on temporary data.

applyTxToUtxo' :: MonadUtxo m => (TxId, TxAux) -> m ()

belongsTo :: TxOutAux -> Address -> Bool

A predicate for TxOut which selects outputs for given address

filterUtxoByAddr :: Address -> Utxo -> Utxo

Select only TxOuts for given addresses

utxoToStakes :: Utxo -> HashMap StakeholderId Coin

newtype UtxoReaderT m a

Constructors

UtxoReaderT

Fields

getUtxoReaderT :: ReaderT Utxo m a

runUtxoReaderT :: UtxoReaderT m a -> Utxo -> m a

type UtxoReader = UtxoReaderT Identity

runUtxoReader :: UtxoReader a -> Utxo -> a

newtype UtxoStateT m a

Constructors

UtxoStateT

Fields

getUtxoStateT :: StateT Utxo m a

runUtxoStateT :: UtxoStateT m a -> Utxo -> m (a, Utxo)

evalUtxoStateT :: Monad m => UtxoStateT m a -> Utxo -> m a

execUtxoStateT :: Monad m => UtxoStateT m a -> Utxo -> m Utxo

type UtxoState = UtxoStateT Identity

runUtxoState :: UtxoState a -> Utxo -> (a, Utxo)

evalUtxoState :: UtxoState a -> Utxo -> a

execUtxoState :: UtxoState a -> Utxo -> Utxo

applyTxToUtxoPure :: WithHash Tx -> TxDistribution -> Utxo -> Utxo

Pure version of applyTxToUtxo.

applyTxToUtxoPure' :: (TxId, TxAux) -> Utxo -> Utxo

Pure version of applyTxToUtxo'.

verifyTxUtxoPure :: Bool -> Bool -> Utxo -> TxAux -> VerificationRes

Pure version of verifyTxUtxo.