Pos.Update

Description

Whole Update System in one module :)

Documentation

module Pos.Update.Core

downloadUpdate :: UpdateMode m => ConfirmedProposalState -> m ()

Download and save archive update by given ConfirmedProposalState

downloadHash :: [Text] -> Hash LByteString -> IO (Either Text LByteString)

Download a file by its hash.

Tries all servers in turn, fails if none of them work.

usApplyBlocks :: (MonadThrow m, USGlobalApplyMode m) => OldestFirst NE UpdateBlock -> Maybe PollModifier -> m [SomeBatchOp]

Apply chain of definitely valid blocks to US part of GState DB and to US local data. This function assumes that no other thread applies block in parallel. It also assumes that parent of oldest block is current tip. If verification is done prior to application, one can pass PollModifier obtained from verification to this function.

usCanCreateBlock :: (WithLogger m, MonadDB m, HasContext LrcContext m) => m Bool

Checks whether our software can create block according to current global state.

usRollbackBlocks :: forall m. USGlobalApplyMode m => NewestFirst NE (UpdateBlock, USUndo) -> m [SomeBatchOp]

Revert application of given blocks to US part of GState DB and US local data. The caller must ensure that the tip stored in DB is headerHash of head.

usVerifyBlocks :: USGlobalVerifyMode m => Bool -> OldestFirst NE UpdateBlock -> m (PollModifier, OldestFirst NE USUndo)

Verify whether sequence of blocks can be applied to US part of current GState DB. This function doesn't make pure checks, they are assumed to be done earlier, most likely during objects construction.

If the first argument is True it means that all data must be known. Currently it only means that UpdateProposals must have only known attributes, but I can't guarantee this comment will always be up-to-date.

Proposals

isProposalNeeded :: (HasContext UpdateContext m, MonadIO m) => UpId -> m Bool

This function returns true if update proposal with given identifier should be requested.

getLocalProposalNVotes :: (HasContext UpdateContext m, MonadIO m) => UpId -> m (Maybe (UpdateProposal, [UpdateVote]))

Get update proposal with given id if it is known.

processProposal :: USLocalLogicMode m => UpdateProposal -> m (Either PollVerFailure ())

Process proposal received from network, checking it against current state (global + local) and adding to local state if it's valid with respect to it. If proposal is added to store, 'Right ()' is returned. Otherwise 'Left err' is returned and err lets caller decide whether sender could be sure that error would happen.

Votes

isVoteNeeded :: USLocalLogicMode m => UpId -> PublicKey -> Bool -> m Bool

This function returns true if update vote proposal with given identifier issued by stakeholder with given PublicKey and with given decision should be requested.

getLocalVote :: (HasContext UpdateContext m, MonadIO m) => UpId -> PublicKey -> Bool -> m (Maybe UpdateVote)

Get update vote for proposal with given id from given issuer and with given decision if it is known.

processVote :: USLocalLogicMode m => UpdateVote -> m (Either PollVerFailure ())

Process vote received from network, checking it against current state (global + local) and adding to local state if it's valid with respect to it. If vote is added to store, 'Right ()' is returned. Otherwise 'Left err' is returned and err lets caller decide whether sender could be sure that error would happen.

Normalization

usNormalize :: USLocalLogicMode m => m ()

Remove local data from memory state to make it consistent with current GState. This function assumes that GState is locked. It tries to leave as much data as possible. It assumes that blkSemaphore is taken.

processNewSlot :: USLocalLogicMode m => SlotId -> m ()

Update memory state to make it correct for given slot.

usPreparePayload :: USLocalLogicMode m => SlotId -> m (Maybe UpdatePayload)

Prepare UpdatePayload for inclusion into new block with given SlotId. This function assumes that blkSemaphore is taken and nobody can apply/rollback blocks in parallel. Sometimes payload can't be created. It can happen if we are trying to create block for slot which has already passed, for example.

usLrcConsumer :: MonadDBCore m => LrcConsumer m

Consumer will be called on every Richmen computation.

module Pos.Update.MemState

usListeners :: UpdateMode m => m ([ListenerSpec m], OutSpecs)

Listeners for requests related to update system

usStubListeners :: WithLogger m => ([ListenerSpec m], OutSpecs)

data ProposalMsgTag

Tag for proposal messages

Constructors

ProposalMsgTag

data VoteMsgTag

Tag for vote messages

Constructors

VoteMsgTag

module Pos.Update.Params

usWorkers :: WorkMode ssc m => m (Set NodeId) -> ([WorkerSpec m], OutSpecs)

Update System related workers.