Pos.Update.Core

Description

Core of Update System.

UpdateProposal and related

data UpdateProposal

Proposal for software update

Constructors

UpdateProposal

Fields

upBlockVersion :: !BlockVersion

upBlockVersionData :: !BlockVersionData

upSoftwareVersion :: !SoftwareVersion

upData :: !(HashMap SystemTag UpdateData)

UpdateData for each system which this update affects. It must be non-empty.

upAttributes :: !UpAttributes

Attributes which are currently empty, but provide extensibility.

type UpId = Hash UpdateProposal

ID of softwaree update proposal

type UpAttributes = Attributes ()

data UpdateData

Data which describes update. It is specific for each system.

Constructors

UpdateData

Fields

udAppDiffHash :: !(Hash Raw)

Hash of binary diff between two applications. This diff can be passed to updater to create new application.

udPkgHash :: !(Hash Raw)

Hash of package to install new application. This package can be used to install new application from scratch instead of updating existing application.

udUpdaterHash :: !(Hash Raw)

Hash if update application which can be used to install this update (relevant only when updater is used, not package).

udMetadataHash :: !(Hash Raw)

Hash of metadata relevant to this update. It is raw hash, because metadata can include image or something (maybe). Anyway, we can always use unsafeHash.

data BlockVersionData

Data which is associated with BlockVersion.

Constructors

BlockVersionData

Fields

bvdScriptVersion :: !ScriptVersion

bvdSlotDuration :: !Millisecond

bvdMaxBlockSize :: !Byte

bvdMaxTxSize :: !Byte

bvdMpcThd :: !CoinPortion

bvdHeavyDelThd :: !CoinPortion

bvdUpdateVoteThd :: !CoinPortion

bvdUpdateProposalThd :: !CoinPortion

bvdUpdateImplicit :: !FlatSlotId

bvdUpdateSoftforkThd :: !CoinPortion

data SystemTag

Tag of system for which update data is purposed, e.g. win64, mac32

mkSystemTag :: MonadFail m => Text -> m SystemTag

systemTagMaxLength :: Integral i => i

patakUpdateData :: HashMap SystemTag UpdateData

skovorodaUpdateData :: Hash Raw -> HashMap SystemTag UpdateData

upScriptVersion :: UpdateProposal -> ScriptVersion

upSlotDuration :: UpdateProposal -> Millisecond

upMaxBlockSize :: UpdateProposal -> Byte

formatVoteShort :: UpdateVote -> Builder

Format UpdateVote compactly.

shortVoteF :: Format r (UpdateVote -> r)

Formatter for UpdateVote which displays it compactly.

UpdateVote and related

data UpdateVote

Vote for update proposal

Constructors

UpdateVote

Fields

uvKey :: !PublicKey

Public key of stakeholder, who votes

uvProposalId :: !UpId

Proposal to which this vote applies

uvDecision :: !Bool

Approval/rejection bit

uvSignature :: !(Signature (UpId, Bool))

Signature of (Update proposal, Approval/rejection bit) by stakeholder

type VoteId = (UpId, PublicKey, Bool)

type StakeholderVotes = HashMap PublicKey VoteState

Type alias for set of votes from stakeholders

type UpdateProposals = HashMap UpId UpdateProposal

type LocalVotes = HashMap UpId (HashMap PublicKey UpdateVote)

mkVoteId :: UpdateVote -> VoteId

Payload and proof

data UpdatePayload

Update System payload. BodyProof contains UpdateProof = Hash UpdatePayload.

Constructors

UpdatePayload

Fields

upProposal :: !(Maybe UpdateProposal)

upVotes :: ![UpdateVote]

type UpdateProof = Hash UpdatePayload

Proof that body of update message contains UpdatePayload.

mkUpdateProof :: Bi UpdatePayload => UpdatePayload -> UpdateProof

VoteState

data VoteState

This type represents summary of votes issued by stakeholder.

Constructors

PositiveVote
Stakeholder voted once positively.

NegativeVote
Stakeholder voted once positively.

PositiveRevote
Stakeholder voted negatively, then positively.

NegativeRevote
Stakeholder voted positively, then negatively.

canCombineVotes :: Bool -> Maybe VoteState -> Bool

Check whether given decision is a valid vote if applied to existing vote (which may not exist).

combineVotes :: Bool -> Maybe VoteState -> Maybe VoteState

Apply decision to given vote (or Nothing). This function returns Nothing if decision can't be applied. canCombineVotes can be used to check whether it will be successful.

isPositiveVote :: VoteState -> Bool

newVoteState :: Bool -> VoteState

Create new VoteState from bool, which is simple vote, not revote.