Pos.Util.Modifier

Description

Wrapper for modifier pattern which is used intensively in code

Documentation

data MapModifier k v

MapModifier is a type which collects modifications (insertions and deletions) of something map-like.

lookupM :: (Applicative m, Eq k, Hashable k) => (k -> m (Maybe v)) -> k -> MapModifier k v -> m (Maybe v)

Perform monadic lookup taking MapModifier into account.

lookup :: (Eq k, Hashable k) => (k -> Maybe v) -> k -> MapModifier k v -> Maybe v

Perform lookup taking MapModifier into account.

keysM :: (Functor m, Eq k, Hashable k) => m [k] -> MapModifier k v -> m [k]

Get keys of something map-like in Functor context taking MapModifier into account.

keys :: (Eq k, Hashable k) => [k] -> MapModifier k v -> [k]

Get keys of something map-like taking MapModifier into account.

valuesM :: (Functor m, Eq k, Hashable k) => m [(k, v)] -> MapModifier k v -> m [v]

Get values of something map-like in Functor context taking MapModifier into account.

values :: (Eq k, Hashable k) => [(k, v)] -> MapModifier k v -> [v]

Get values of something map-like taking MapModifier into account.

toListM :: (Functor m, Eq k, Hashable k) => m [(k, v)] -> MapModifier k v -> m [(k, v)]

Get contents of something map-like in Functor context taking MapModifier into account.

toList :: (Eq k, Hashable k) => [(k, v)] -> MapModifier k v -> [(k, v)]

Get contents of something map-like taking MapModifier into account.

insertions :: MapModifier k v -> [(k, v)]

Get all insertions into MapModifier.

deletions :: MapModifier k v -> [k]

Get all deletions from MapModifier.

insert :: (Eq k, Hashable k) => k -> v -> MapModifier k v -> MapModifier k v

Insert something into MapModifier. Effect of it is like inserting something into underlying container (which replaces existing value if it exists).

delete :: (Eq k, Hashable k) => k -> MapModifier k v -> MapModifier k v

Delete something into MapModifier. Nothing is actually deleted from MapModifier itself. Effect is akin to deleting value from underlying container.

mapMaybeM :: (Functor m, Eq k, Hashable k) => m [(k, v2)] -> (v1 -> Maybe v2) -> MapModifier k v1 -> m [(k, v2)]

Transform this modifer by applying a function to every insertion and retaining only some of them. Underlying map should be already transformed.