Pos.Util.UserSecret

Description

Secret key file storage and management functions based on file locking.

Documentation

data UserSecret

User secret data. Includes secret keys only for now (not including auxiliary _usPath).

usKeys :: Lens' UserSecret [EncryptedSecretKey]

usVss :: Lens' UserSecret (Maybe VssKeyPair)

usPrimKey :: Lens' UserSecret (Maybe SecretKey)

getUSPath :: UserSecret -> FilePath

simpleUserSecret :: SecretKey -> FilePath -> UserSecret

Create a simple UserSecret from secret key and file path

initializeUserSecret :: (MonadIO m, WithLogger m) => FilePath -> m ()

Create user secret file at the given path, but only when one doesn't already exist.

readUserSecret :: (MonadIO m, WithLogger m) => FilePath -> m UserSecret

Reads user secret from file, assuming that file exists, and has mode 600, throws exception in other case

peekUserSecret :: (MonadIO m, WithLogger m) => FilePath -> m UserSecret

Reads user secret from the given file. If the file does not exist/is empty, returns empty user secret

takeUserSecret :: (MonadIO m, WithLogger m) => FilePath -> m UserSecret

Read user secret putting an exclusive lock on it. To unlock, use writeUserSecretRelease.

writeUserSecret :: MonadIO m => UserSecret -> m ()

Writes user secret .

writeUserSecretRelease :: (MonadFail m, MonadIO m) => UserSecret -> m ()

Writes user secret and releases the lock. UserSecret can't be used after this function call anymore.