Pos.Wallet

Documentation

class Monad m => MonadKeys m where

Methods

getPrimaryKey :: m (Maybe SecretKey)

getSecretKeys :: m [EncryptedSecretKey]

addSecretKey :: EncryptedSecretKey -> m ()

deleteSecretKey :: Word -> m ()

getPrimaryKey :: (MonadTrans t, MonadKeys m', t m' ~ m) => m (Maybe SecretKey)

getSecretKeys :: (MonadTrans t, MonadKeys m', t m' ~ m) => m [EncryptedSecretKey]

addSecretKey :: (MonadTrans t, MonadKeys m', t m' ~ m) => EncryptedSecretKey -> m ()

deleteSecretKey :: (MonadTrans t, MonadKeys m', t m' ~ m) => Word -> m ()

newSecretKey :: (MonadIO m, MonadKeys m) => PassPhrase -> m EncryptedSecretKey

Helper for generating a new secret key

type KeyStorage = ReaderT KeyData

type KeyData = TVar UserSecret

data KeyError

Constructors

PrimaryKey !Text
Failed attempt to delete primary key

runKeyStorage :: (MonadIO m, WithLogger m) => FilePath -> KeyStorage m a -> m a

runKeyStorageRaw :: KeyStorage m a -> KeyData -> m a

data WalletParams

Constructors

WalletParams

Fields

wpDbPath :: !(Maybe FilePath)

wpRebuildDb :: !Bool

wpKeyFilePath :: !FilePath

wpSystemStart :: Timestamp

wpGenesisKeys :: !Bool

wpBaseParams :: !BaseParams

wpGenesisUtxo :: !Utxo

Genesis utxo

runRawRealWallet :: PeerId -> RealModeResources WalletRealMode -> WalletParams -> ListenersWithOut WalletRealMode -> (ActionSpec WalletRealMode a, OutSpecs) -> Production a

runWalletRealMode :: PeerId -> RealModeResources WalletRealMode -> WalletParams -> (ActionSpec WalletRealMode a, OutSpecs) -> Production a

WalletMode runner

runWalletReal :: PeerId -> RealModeResources WalletRealMode -> WalletParams -> ([WorkerSpec WalletRealMode], OutSpecs) -> Production ()

runWallet :: WalletMode ssc m => RealModeResources m -> ([WorkerSpec m], OutSpecs) -> (WorkerSpec m, OutSpecs)

submitVote :: (MinWorkMode m, MonadDBLimits m) => SendActions m -> [NodeId] -> UpdateVote -> m ()

Send UpdateVote to given addresses

submitUpdateProposal :: (MinWorkMode m, MonadDBLimits m) => SendActions m -> SafeSigner -> [NodeId] -> UpdateProposal -> m ()

Send UpdateProposal with one positive vote to given addresses

sendVoteOuts :: OutSpecs

sendProposalOuts :: OutSpecs

class Monad m => MonadBalances m where

A class which have the methods to get state of address' balance

Methods

getOwnUtxo :: Address -> m Utxo

getBalance :: Address -> m Coin

getOwnUtxo :: (MonadTrans t, MonadBalances m', t m' ~ m) => Address -> m Utxo

getBalance :: (MonadTrans t, MonadBalances m', t m' ~ m) => Address -> m Coin

class Monad m => MonadTxHistory m where

A class which have methods to get transaction history

Methods

getTxHistory :: SscHelpersClass ssc => Tagged ssc (Address -> Maybe (HeaderHash, Utxo) -> m TxHistoryAnswer)

saveTx :: (TxId, TxAux) -> m ()

getTxHistory :: (SscHelpersClass ssc, MonadTrans t, MonadTxHistory m', t m' ~ m) => Tagged ssc (Address -> Maybe (HeaderHash, Utxo) -> m TxHistoryAnswer)

saveTx :: (MonadTrans t, MonadTxHistory m', t m' ~ m) => (TxId, TxAux) -> m ()

class Monad m => MonadBlockchainInfo m where

Methods

networkChainDifficulty :: m (Maybe ChainDifficulty)

localChainDifficulty :: m ChainDifficulty

blockchainSlotDuration :: m Millisecond

connectedPeers :: m Word

networkChainDifficulty :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m (Maybe ChainDifficulty)

localChainDifficulty :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m ChainDifficulty

blockchainSlotDuration :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m Millisecond

connectedPeers :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m Word

class Monad m => MonadUpdates m where

Abstraction over getting update proposals

Methods

waitForUpdate :: m ConfirmedProposalState

applyLastUpdate :: m ()

waitForUpdate :: (MonadTrans t, MonadUpdates m', t m' ~ m) => m ConfirmedProposalState

applyLastUpdate :: (MonadTrans t, MonadUpdates m', t m' ~ m) => m ()

type WalletMode ssc m = (TxMode ssc m, MonadKeys m, MonadBlockchainInfo m, MonadUpdates m, WithPeerState m)

type WalletRealMode = PeerStateHolder (KeyStorage (WalletDB (ReportingContextT (LoggerNameBox Production))))