Pos.Wallet.Web

Documentation

type WalletApi = TestReset :<|> (GetWallet :<|> (GetWallets :<|> (UpdateWallet :<|> (DeleteWallet :<|> (ImportKey :<|> (WalletRestore :<|> (NewWallet :<|> (IsValidAddress :<|> (GetProfile :<|> (UpdateProfile :<|> (NewPayment :<|> (NewPaymentExt :<|> (UpdateTx :<|> (GetHistory :<|> (SearchHistory :<|> (NextUpdate :<|> (ApplyUpdate :<|> (RedeemADA :<|> (RedeemADAPaperVend :<|> (ReportingInitialized :<|> (ReportingElectroncrash :<|> (GetSlotsDuration :<|> (GetVersion :<|> GetSyncProgress)))))))))))))))))))))))

Servant API which provides access to wallet. TODO: Should be composed depending on the resource - wallets, txs, ... http://haskell-servant.github.io/tutorial/0.4/server.html#nested-apis

walletApi :: Proxy WalletApi

Helper Proxy.

type TestReset = API :> ("test" :> ("reset" :> Post '[JSON] (Either WalletError ())))

All endpoints are defined as a separate types, for description in Swagger-based HTML-documentation.

type GetWallet = API :> ("wallets" :> (Capture "walletId" CAddress :> Get '[JSON] (Either WalletError CWallet)))

type GetWallets = API :> ("wallets" :> Get '[JSON] (Either WalletError [CWallet]))

type UpdateWallet = API :> ("wallets" :> (Capture "walletId" CAddress :> (ReqBody '[JSON] CWalletMeta :> Put '[JSON] (Either WalletError CWallet))))

type DeleteWallet = API :> ("wallets" :> (Capture "walletId" CAddress :> Delete '[JSON] (Either WalletError ())))

type ImportKey = API :> ("wallets" :> ("keys" :> (ReqBody '[JSON] Text :> Post '[JSON] (Either WalletError CWallet))))

type WalletRestore = API :> ("wallets" :> ("restore" :> (Capture "passphrase" CPassPhrase :> (ReqBody '[JSON] CWalletInit :> Post '[JSON] (Either WalletError CWallet)))))

type NewWallet = API :> ("wallets" :> (Capture "passphrase" CPassPhrase :> (ReqBody '[JSON] CWalletInit :> Post '[JSON] (Either WalletError CWallet))))

type IsValidAddress = API :> ("addresses" :> (Capture "address" Text :> ("currencies" :> (Capture "currency" CCurrency :> Get '[JSON] (Either WalletError Bool)))))

type GetProfile = API :> ("profile" :> Get '[JSON] (Either WalletError CProfile))

type UpdateProfile = API :> ("profile" :> (ReqBody '[JSON] CProfile :> Post '[JSON] (Either WalletError CProfile)))

type NewPayment = API :> ("txs" :> ("payments" :> (Capture "passphrase" CPassPhrase :> (Capture "from" CAddress :> (Capture "to" CAddress :> (Capture "amount" Coin :> Post '[JSON] (Either WalletError CTx)))))))

type NewPaymentExt = API :> ("txs" :> ("payments" :> (Capture "passphrase" CPassPhrase :> (Capture "from" CAddress :> (Capture "to" CAddress :> (Capture "amount" Coin :> (Capture "currency" CCurrency :> (Capture "title" Text :> (Capture "description" Text :> Post '[JSON] (Either WalletError CTx))))))))))

type UpdateTx = API :> ("txs" :> ("payments" :> (Capture "address" CAddress :> (Capture "transaction" CTxId :> (ReqBody '[JSON] CTxMeta :> Post '[JSON] (Either WalletError ()))))))

type GetHistory = API :> ("txs" :> ("histories" :> (Capture "address" CAddress :> (QueryParam "skip" Word :> (QueryParam "limit" Word :> Get '[JSON] (Either WalletError ([CTx], Word)))))))

type SearchHistory = API :> ("txs" :> ("histories" :> (Capture "address" CAddress :> (Capture "search" Text :> (QueryParam "skip" Word :> (QueryParam "limit" Word :> Get '[JSON] (Either WalletError ([CTx], Word))))))))

type NextUpdate = API :> ("update" :> Get '[JSON] (Either WalletError CUpdateInfo))

type ApplyUpdate = API :> ("update" :> Post '[JSON] (Either WalletError ()))

type RedeemADA = API :> ("redemptions" :> ("ada" :> (ReqBody '[JSON] CWalletRedeem :> Post '[JSON] (Either WalletError CTx))))

type RedeemADAPaperVend = API :> ("papervend" :> ("redemptions" :> ("ada" :> (ReqBody '[JSON] CPaperVendWalletRedeem :> Post '[JSON] (Either WalletError CTx)))))

type ReportingInitialized = API :> ("reporting" :> ("initialized" :> (ReqBody '[JSON] CInitialized :> Post '[JSON] (Either WalletError ()))))

type ReportingElectroncrash = API :> ("reporting" :> ("electroncrash" :> (MultipartForm CElectronCrashReport :> Post '[JSON] (Either WalletError ()))))

type GetSlotsDuration = API :> ("settings" :> ("slots" :> ("duration" :> Get '[JSON] (Either WalletError Word))))

type GetVersion = API :> ("settings" :> ("version" :> Get '[JSON] (Either WalletError SoftwareVersion)))

type GetSyncProgress = API :> ("settings" :> ("sync" :> ("progress" :> Get '[JSON] (Either WalletError SyncProgress))))

data SyncProgress

Constructors

SyncProgress

Fields

_spLocalCD :: ChainDifficulty

_spNetworkCD :: Maybe ChainDifficulty

_spPeers :: Word

newtype CAddress

Client address

Constructors

CAddress CHash

data CCurrency

Currencies handled by client. Note: Cardano does not deal with other currency than ADA yet.

Constructors

ADA

BTC

ETH

newtype CHash

Client hash

Constructors

CHash Text

data CTType

type of transactions It can be an input output exchange transaction CTInOut CTExMeta -- Ex == exchange

Constructors

CTIn CTxMeta

CTOut CTxMeta

newtype CPassPhrase

Constructors

CPassPhrase Text

data CProfile

Client profile (CP) all data of client are "meta data" - that is not provided by Cardano (Flow type: accountType)

Constructors

CProfile

Fields

cpLocale :: Text

type CPwHash = Text

Password hash of client profile

data CTx

Client transaction (CTx) Provides all Data about a transaction needed by client. It includes meta data which are not part of Cardano, too (Flow type: transactionType)

Constructors

CTx

Fields

ctId :: !CTxId

ctAmount :: !CCoin

ctFrom :: ![(CAddress, CCoin)]

ctTo :: ![(CAddress, CCoin)]

ctConfirmations :: !Word

ctType :: !CTType

data CTxId

Client transaction id

data CTxMeta

meta data of transactions

Constructors

CTxMeta

Fields

ctmCurrency :: CCurrency

ctmTitle :: Text

ctmDescription :: Text

ctmDate :: POSIXTime

data CTExMeta

meta data of exchanges

Constructors

CTExMeta

Fields

cexCurrency :: CCurrency

cexTitle :: Text

cexDescription :: Text

cexDate :: POSIXTime

cexRate :: Text

cexLabel :: Text

cexAddress :: CAddress

data CInitialized

Represents a knowledge about how much time did it take for client (wallet) to initialize. All numbers are milliseconds.

Constructors

CInitialized

Fields

cTotalTime :: Word

Total time from very start to main post-sync screen.

cPreInit :: Word

Time passed from beginning to network connection with peers established.

data CWallet

Client Wallet (CW) (Flow type: walletType)

Constructors

CWallet

Fields

cwAddress :: !CAddress

cwAmount :: !CCoin

cwMeta :: !CWalletMeta

data CWalletType

A wallet can be used as personal or shared wallet

Constructors

CWTPersonal

CWTShared

data CWalletAssurance

A level of assurance for the wallet "meta type"

Constructors

CWAStrict

CWANormal

data CWalletMeta

Meta data of CWallet Includes data which are not provided by Cardano

Constructors

CWalletMeta

Fields

cwType :: !CWalletType

cwCurrency :: !CCurrency

cwName :: !Text

cwAssurance :: !CWalletAssurance

cwUnit :: !Int

https://issues.serokell.io/issue/CSM-163#comment=96-2480

data CWalletInit

Query data for wallet creation (wallet meta + backup phrase)

Constructors

CWalletInit

Fields

cwBackupPhrase :: !BackupPhrase

cwInitMeta :: !CWalletMeta

data CUpdateInfo

Update system data

Constructors

CUpdateInfo

Fields

cuiSoftwareVersion :: !SoftwareVersion

cuiBlockVesion :: !BlockVersion

cuiScriptVersion :: !ScriptVersion

cuiImplicit :: !Bool

cuiVotesFor :: !Int

cuiVotesAgainst :: !Int

cuiPositiveStake :: !CCoin

cuiNegativeStake :: !CCoin

data CWalletRedeem

Query data for redeem

Constructors

CWalletRedeem

Fields

crWalletId :: !CAddress

crSeed :: !Text

data CPaperVendWalletRedeem

Query data for redeem

Constructors

CPaperVendWalletRedeem

Fields

pvWalletId :: !CAddress

pvSeed :: !Text

pvBackupPhrase :: !BackupPhrase

data CCoin

mkCCoin :: Coin -> CCoin

data CElectronCrashReport

Constructors

CElectronCrashReport

Fields

cecVersion :: Text

cecPlatform :: Text

cecProcessType :: Text

cecGuid :: Text

cecVersionJson :: Text

cecProductName :: Text

cecProd :: Text

cecCompanyName :: Text

cecUploadDump :: FileData

data NotifyEvent

Constructors

ConnectionOpened

NetworkDifficultyChanged ChainDifficulty

LocalDifficultyChanged ChainDifficulty

ConnectedPeersChanged Word

UpdateAvailable

ConnectionClosed

addressToCAddress :: Address -> CAddress

Transform Address into CAddress

cAddressToAddress :: CAddress -> Either Text Address

passPhraseToCPassPhrase :: PassPhrase -> CPassPhrase

cPassPhraseToPassPhrase :: CPassPhrase -> Either Text PassPhrase

mkCTx

Arguments

:: Address
An address for which transaction info is forming

-> ChainDifficulty
Current chain difficulty (to get confirmations)

-> TxHistoryEntry
Tx history entry

-> CTxMeta
Transaction metadata

-> CTx

mkCTxId :: Text -> CTxId

txIdToCTxId :: TxId -> CTxId

transform TxId into CTxId

txContainsTitle :: Text -> CTx -> Bool

toCUpdateInfo :: ConfirmedProposalState -> CUpdateInfo

Creates CTUpdateInfo from ConfirmedProposalState

walletDocsText :: Text

walletTableDocsText :: Text

data WalletError

Constructors

Internal !Text
Some internal error.

walletServeWebFull :: forall ssc. SscConstraint ssc => RawRealMode ssc (Set NodeId) -> SendActions (RawRealMode ssc) -> Bool -> FilePath -> Bool -> Word16 -> RawRealMode ssc ()

walletServerOuts :: OutSpecs

walletServeWebLite :: forall ssc. SscHelpersClass ssc => Proxy ssc -> WalletRealMode (Set NodeId) -> SendActions WalletRealMode -> FilePath -> Bool -> Word16 -> WalletRealMode ()

walletServerOuts :: OutSpecs

type WalletWebHandler m = WalletWebSockets (WalletWebDB m)

walletApplication :: WalletWebMode m => m (Server WalletApi) -> m Application

walletServer :: (Monad m, MonadIO m, WalletWebMode (WalletWebHandler m)) => m (Set NodeId) -> SendActions m -> WalletWebHandler m (WalletWebHandler m :~> Handler) -> WalletWebHandler m (Server WalletApi)

walletServeImpl

Arguments

:: (MonadIO m, MonadMask m, WalletWebMode (WalletWebHandler m))

=> WalletWebHandler m Application
Application getter

-> FilePath
Path to wallet acid-state

-> Bool
Rebuild flag for acid-state

-> Word16
Port to listen

-> m ()

walletServerOuts :: OutSpecs

type WalletWebSockets = ReaderT ConnectionsVar

Holder for web wallet data

type WebWalletSockets m = (MonadWalletWebSockets m, MonadIO m)

type MonadWalletWebSockets = MonadReader ConnectionsVar

MonadWalletWebSockets stands for monad which is able to get web wallet sockets

getWalletWebSockets :: MonadWalletWebSockets m => m ConnectionsVar

type ConnectionsVar = TVar (Maybe Connection)

initWSConnection :: MonadIO m => m ConnectionsVar

closeWSConnection :: MonadIO m => ConnectionsVar -> m ()

upgradeApplicationWS :: ConnectionsVar -> Application -> Application

notify :: WebWalletSockets m => NotifyEvent -> m ()

runWalletWS :: ConnectionsVar -> WalletWebSockets m a -> m a

type WalletWebDB = ReaderT WalletState

Holder for web wallet data

runWalletWebDB :: WalletState -> WalletWebDB m a -> m a

Execute WalletWebDB action with given WalletState

getWalletState :: Monad m => WalletWebDB m WalletState

type WalletState = ExtendedState WalletStorage

type MonadWalletWebDB = MonadReader WalletState

MonadWalletWebDB stands for monad which is able to get web wallet state

getWalletWebState :: MonadWalletWebDB m => m WalletState

type WebWalletModeDB m = (MonadWalletWebDB m, MonadIO m, MonadMockable m)

Constraint for working with web wallet DB

openState :: MonadIO m => Bool -> FilePath -> m WalletState

openMemState :: MonadIO m => m WalletState

closeState :: MonadIO m => WalletState -> m ()

Getters

getProfile :: WebWalletModeDB m => m CProfile

getWalletMetas :: WebWalletModeDB m => m [CWalletMeta]

getWalletMeta :: WebWalletModeDB m => CAddress -> m (Maybe CWalletMeta)

getTxMeta :: WebWalletModeDB m => CAddress -> CTxId -> m (Maybe CTxMeta)

getWalletHistory :: WebWalletModeDB m => CAddress -> m (Maybe [CTxMeta])

getUpdates :: WebWalletModeDB m => m [CUpdateInfo]

getNextUpdate :: WebWalletModeDB m => m (Maybe CUpdateInfo)

getHistoryCache :: WebWalletModeDB m => CAddress -> m (Maybe (HeaderHash, Utxo, [TxHistoryEntry]))

Setters

testReset :: WebWalletModeDB m => m ()

createWallet :: WebWalletModeDB m => CAddress -> CWalletMeta -> m ()

setProfile :: WebWalletModeDB m => CProfile -> m ()

setWalletMeta :: WebWalletModeDB m => CAddress -> CWalletMeta -> m ()

setWalletTransactionMeta :: WebWalletModeDB m => CAddress -> CTxId -> CTxMeta -> m ()

setWalletHistory :: WebWalletModeDB m => CAddress -> [(CTxId, CTxMeta)] -> m ()

addOnlyNewTxMeta :: WebWalletModeDB m => CAddress -> CTxId -> CTxMeta -> m ()

removeWallet :: WebWalletModeDB m => CAddress -> m ()

addUpdate :: WebWalletModeDB m => CUpdateInfo -> m ()

removeNextUpdate :: WebWalletModeDB m => m ()

updateHistoryCache :: WebWalletModeDB m => CAddress -> HeaderHash -> Utxo -> [TxHistoryEntry] -> m ()