Pos.Worker

Description

High level workers.

Documentation

runWorkers :: (SscWorkersClass ssc, SecurityWorkersClass ssc, WorkMode ssc m) => SendActions BiP m -> m ()

Run all necessary workers in separate threads. This call doesn't block.

A note about order: currently all onNewSlot updates can be run in parallel and we try to maintain this rule. If at some point order becomes important, update this comment! I don't think you will read it, but who knows…

statsWorkers :: WorkMode ssc m => [SendActions BiP m -> m ()]

Workers for collecting statistics about transactions in background.