Pos.Binary

Description

Reexports of Pos.Binary.* classes

Documentation

class Bi t where

Simplified definition of serializable object, Data.Binary.Class-alike.

Write instance Bi SomeType where without any method definitions if you want to use the Binary instance for your type.

Methods

put :: t -> Put

put :: Binary t => t -> Put

get :: Get t

get :: Binary t => Get t

encode :: Bi a => a -> LByteString

Encode a value to a lazy bytestring

encodeStrict :: Bi a => a -> ByteString

Encode a value to a strict bytestring. Use with caution, because converting to strict ByteString is expensive.

decode :: Bi a => LByteString -> a

Decode a value from a lazy ByteString, reconstructing the original structure.

decodeOrFail :: Bi a => LByteString -> Either (LByteString, ByteOffset, String) (LByteString, ByteOffset, a)

decodeFull :: Bi a => LByteString -> Either String a

Like decode, but ensures that the whole input has been consumed.

Different sizes for ints

newtype UnsignedVarInt a

A newtype wrapper for non-negative varints. During serialization its contents will be encoded as a variable-sized integer.

Despite its name, e.g. UnsignedVarInt (-50 :: Int) will be serialized and deserialized correctly; however, UnsignedVarInt is optimized for non-negative numbers, and will always take maximum space (e.g. 10 bytes in case of Int64). Specifically, Int is simply coerced into its Word representation before being serialized.

Constructors

UnsignedVarInt

Fields

getUnsignedVarInt :: a

newtype SignedVarInt a

A newtype wrapper for varints. Uses zig-zag encoding to serialize negative integers – e.g. -3 is turned into 5, -4 is turned into 7, etc; thus it's fair but less optimal for positive integers.

Constructors

SignedVarInt

Fields

getSignedVarInt :: a

newtype TinyVarInt

A newtype wrapper for non-negative integers less than 2^14. Use it if you want to be extra careful. Compared to SignedVarInt and UnsignedVarInt, it provides two benefits:

It is guaranteed to take either 1 or 2 bytes (the standard decoder for varints can consume an unlimited amount of bytes).

It is unambiguous (e.g. 0 can be encoded in only one way instead of two).

Constructors

TinyVarInt

Fields

getTinyVarInt :: Word16

newtype FixedSizeInt a

A newtype wrapper for signifying that an integer should be serialized using a fixed amount of bytes.

Constructors

FixedSizeInt

Fields

getFixedSizeInt :: a

Primitives for limiting serialization

limitGet

Arguments

:: Int64
The upper limit on byte consumption

-> Get a
The decoder to isolate

-> Get a

Like isolate, but allows consuming less bytes than expected (just not more).

isolate64 :: Int64 -> Get a -> Get a

Like isolate, but works with Int64 only.