Pos.Block.Network

Description

Re-export of Pos.Block.Network.*

Documentation

announceBlock :: WorkMode ssc m => SendActions m -> MainBlockHeader ssc -> m ()

announceBlockOuts :: OutSpecs

handleHeadersCommunication :: WorkMode ssc m => ConversationActions (MsgHeaders ssc) MsgGetHeaders m -> m ()

blockListeners :: WorkMode ssc m => ([ListenerSpec m], OutSpecs)

blockStubListeners :: (SscHelpersClass ssc, WithLogger m) => Tagged ssc ([ListenerSpec m], OutSpecs)

retrievalWorker :: forall ssc m. (SscWorkersClass ssc, WorkMode ssc m) => (WorkerSpec m, OutSpecs)

handleUnsolicitedHeaders :: forall ssc m. WorkMode ssc m => NonEmpty (BlockHeader ssc) -> NodeId -> ConversationActions MsgGetHeaders (MsgHeaders ssc) m -> m ()

requestTip :: forall ssc m. WorkMode ssc m => NodeId -> ConversationActions MsgGetHeaders (MsgHeaders ssc) m -> m ()

requestHeaders :: forall ssc m. WorkMode ssc m => MsgGetHeaders -> Maybe (BlockHeader ssc) -> NodeId -> ConversationActions MsgGetHeaders (MsgHeaders ssc) m -> m ()

addToBlockRequestQueue :: forall ssc m. WorkMode ssc m => NewestFirst NE (BlockHeader ssc) -> Maybe (BlockHeader ssc) -> NodeId -> m ()

Given nonempty list of valid blockheaders and nodeid, this function will put them into download queue and they will be processed later. Second argument is optional recovery mode tip -- after pack of blocks is processed, next pack of headers will be requested until this header hash is received.

mkHeadersRequest :: WorkMode ssc m => Maybe HeaderHash -> m (Maybe MsgGetHeaders)

Make GetHeaders message using our main chain. This function chooses appropriate from hashes and puts them into GetHeaders message.

requestTipOuts :: OutSpecs

data MsgGetHeaders

GetHeaders message. Behaviour of the response depends on particular combination of mghFrom and mghTo.

mghTo resolves to some header (let's call it top for convenience) -- node's tip if it's Nothing, header with hash in Just if it's Just.

If mghFrom is empty, then semantics is "request to return header of block top".

Otherwise (if mghFrom isn't empty) it represents the set of checkpoints. Responding node will try to iterate headers from top to older until it reaches any checkpoint. If it finds checkpoint c, it returns all headers in range [c.next..top]. If it doesn't find any checkpoint or depth of searching exceeds recoveryHeadersMessage, it will try to find the newest checkpoint cc from mghFrom that's in main chain of responding node and then return at most recoveryHeadersMessage headers starting with cc as the oldest one, returning headers in range l2 = [cc.next..x] where x is either top (in case length l2 < recoveryHeadersMessage) or some arbitrary header (and length is precisely recoveryHeadersMessage).

Constructors

MsgGetHeaders

Fields

mghFrom :: ![HeaderHash]

mghTo :: !(Maybe HeaderHash)

data MsgGetBlocks

GetHeaders message (see protocol specification).

Constructors

MsgGetBlocks

Fields

mgbFrom :: !HeaderHash

mgbTo :: !HeaderHash

newtype MsgHeaders ssc

Headers message (see protocol specification).

Constructors

MsgHeaders (NewestFirst NE (BlockHeader ssc))

newtype MsgBlock s ssc

Block message (see protocol specification).

The s parameter is used for passing block size limit to deserialization instances (using Data.Reflection). Grep for reify and reflect to see usage examples.

Constructors

MsgBlock (Block ssc)