Pos.Constants

Description

Constants used by algorithm. See paper for more details. Some constants are parsed at compile-time (see Type). Others are derived from those.

Constants mentioned in paper

blkSecurityParam :: Integral a => a

Security parameter which is maximum number of blocks which can be rolled back.

slotSecurityParam :: Integral a => a

Security parameter expressed in number of slots. It uses chain quality property. It's basically 'blkSecurityParam / chain_quality'.

epochSlots :: Integral a => a

Number of slots inside one epoch.

networkDiameter :: Microsecond

Estimated time needed to broadcast message from one node to all other nodes. Also see ccNetworkDiameter.

SSC constants

sharedSeedLength :: Integral a => a

Length of shared seed.

mpcSendInterval :: Microsecond

Length of interval during which node should send her MPC message. Relevant only for one SSC implementation. Also see ccMpcSendInterval.

Dev/production mode, system start

isDevelopment :: Bool

True if current mode is Development.

staticSysStart :: Maybe Timestamp

System start time embeded into binary.

Genesis constants

genesisN :: Integral i => i

See ccGenesisN.

genesisSlotDuration :: Millisecond

Length of slot.

genesisMaxBlockSize :: Byte

Maximum size of a block (in bytes)

genesisMaxTxSize :: Byte

See ccGenesisMaxTxSize.

genesisMpcThd :: CoinPortion

See ccGenesisMpcThd.

genesisHeavyDelThd :: CoinPortion

See ccGenesisHeavyDelThd.

genesisUpdateVoteThd :: CoinPortion

See ccGenesisUpdateVoteThd.

genesisUpdateProposalThd :: CoinPortion

See ccGenesisUpdateProposalThd.

genesisUpdateImplicit :: Integral i => i

See ccGenesisUpdateImplicit.

genesisUpdateSoftforkThd :: CoinPortion

See ccGenesisUpdateSoftforkThd.

Other constants

maxLocalTxs :: Integral i => i

Maximum amount of transactions we have in storage (i.e. we can accept without putting them in block). There're next kind of storages in our implementation:

temporary storage of transactions

utxo map that corresponds to it

utxo of blocks in history

This constant is size of first set. Also see ccMaxLocalTxs.

neighborsSendThreshold :: Integral a => a

See ccNeighboursSendThreshold.

networkConnectionTimeout :: Microsecond

networkReceiveTimeout :: Microsecond

blockRetrievalQueueSize :: Integral a => a

propagationQueueSize :: Integral a => a

defaultPeers :: [DHTNode]

See ccDefaultPeers.

sysTimeBroadcastSlots :: Integral i => i

Time-lord node announces system start time by broadcast. She does it during first ccSysTimeBroadcastSlots slots.

vssMaxTTL :: Integral i => i

Max VSS certificate TTL (Ssc.GodTossing part)

vssMinTTL :: Integral i => i

Min VSS certificate TTL (Ssc.GodTossing part)

protocolMagic :: Int32

Protocol magic constant. Is put to block serialized version to distinguish testnet and realnet (for example, possible usages are wider).

enhancedMessageBroadcast :: Integral a => a

Setting this to true enables enhanced message broadcast

recoveryHeadersMessage :: Integral a => a

Maximum amount of headers node can put into headers message while in "after offline" or "recovery" mode. Should be more than blkSecurityParam.

kademliaDumpInterval :: Integral a => a

Interval for dumping state of Kademlia in slots

messageCacheTimeout :: Integral a => a

Timeout for caching system. Components that use caching on messages can use this timeout to invalidate caches.

Delegation

lightDlgConfirmationTimeout :: Integral a => a

Amount of time we hold confirmations for light PSKs.

Malicious activity detection constants

mdNoBlocksSlotThreshold :: Integral i => i

Number of slots used by malicious actions detection to check if we are not receiving generated blocks.

mdNoCommitmentsEpochThreshold :: Integral i => i

Number of epochs used by malicious actions detection to check if our commitments are not included in blockchain.

Update system constants

lastKnownBlockVersion :: BlockVersion

Last block version application is aware of.

curSoftwareVersion :: SoftwareVersion

Version of application (code running)

ourAppName :: ApplicationName

Name of our application.

appSystemTag :: SystemTag

NTP

ntpMaxError :: Microsecond

Inaccuracy in call threadDelay (actually it is error much less than 1 sec)

ntpResponseTimeout :: Microsecond

After making request to NTP servers, how long to wait for their response

ntpPollDelay :: Microsecond

How often send request to NTP server