Pos.Constants

Description

Constants used by algorithm. See paper for more details. Some constants are parsed at compile-time (see Type). Others are derived from those.

Constants mentioned in paper

blkSecurityParam :: Integral a => a

Security parameter which is maximum number of blocks which can be rolled back.

slotSecurityParam :: Integral a => a

Security parameter expressed in number of slots. It uses chain quality property. It's basically 'blkSecurityParam / chain_quality'.

slotDuration :: Microsecond

Length of slot. Also see ccSlotDurationSec.

epochSlots :: Integral a => a

Number of slots inside one epoch.

epochDuration :: Microsecond

Length of one epoch in Microseconds.

networkDiameter :: Microsecond

Estimated time needed to broadcast message from one node to all other nodes. Also see ccNetworkDiameter.

SSC constants

sharedSeedLength :: Integral a => a

Length of shared seed.

mpcSendInterval :: Microsecond

Length of interval during which node should send her MPC message. Relevant only for one SSC implementation. Also see ccMpcSendInterval.

mpcThreshold :: CoinPortion

Threshold value for mpc participation.

Dev/production mode, system start

isDevelopment :: Bool

True if current mode is Development.

staticSysStart :: Maybe Timestamp

System start time embeded into binary.

Other constants

genesisN :: Integral i => i

See ccGenesisN.

maxLocalTxs :: Integral i => i

Maximum amount of transactions we have in storage (i.e. we can accept without putting them in block). There're next kind of storages in our implementation:

temporary storage of transactions

utxo map that corresponds to it

utxo of blocks in history

This constant is size of first set. Also see ccMaxLocalTxs.

maxBlockProxySKs :: Integral i => i

Maximum number of PSKs allowed in block

neighborsSendThreshold :: Integral a => a

See ccNeighboursSendThreshold.

networkConnectionTimeout :: Microsecond

blockRetrievalQueueSize :: Integral a => a

defaultPeers :: [DHTNode]

See ccDefaultPeers.

sysTimeBroadcastSlots :: Integral i => i

Time-lord node announces system start time by broadcast. She does it during first ccSysTimeBroadcastSlots slots.

vssMaxTTL :: Integral i => i

Max VSS certificate TTL (Ssc.GodTossing part)

vssMinTTL :: Integral i => i

Min VSS certificate TTL (Ssc.GodTossing part)

protocolMagic :: Int32

Protocol magic constant. Is put to block serialized version to distinguish testnet and realnet (for example, possible usages are wider).

enhancedMessageBroadcast :: Integral a => a

Setting this to true enables enhanced message broadcast

delegationThreshold :: CoinPortion

Portion of total stake necessary to vote for or against update.

recoveryHeadersMessage :: Integral a => a

Maximum amount of headers node can put into headers message while in "after offline" or "recovery" mode. Should be more than blkSecurityParam.

kademliaDumpInterval :: Integral a => a

Interval for dumping state of Kademlia in slots

Malicious activity detection constants

mdNoBlocksSlotThreshold :: Integral i => i

Number of slots used by malicious actions detection to check if we are not receiving generated blocks.

mdNoCommitmentsEpochThreshold :: Integral i => i

Number of epochs used by malicious actions detection to check if our commitments are not included in blockchain.

Update system constants

lastKnownBlockVersion :: BlockVersion

Last block version application is aware of.

curSoftwareVersion :: SoftwareVersion

Version of application (code running)

appSystemTag :: SystemTag

updateServers :: [String]

Update servers

updateProposalThreshold :: CoinPortion

Portion of total stake such that block containing UpdateProposal must contain positive votes for this proposal from stakeholders owning at least this amount of stake.

updateVoteThreshold :: CoinPortion

Portion of total stake necessary to vote for or against update.

updateImplicitApproval :: Integral i => i

Number of slots after which update is implicitly approved unless it has more negative votes than positive.

NTP

ntpMaxError :: Microsecond

Inaccuracy in call threadDelay (actually it is error much less than 1 sec)

ntpResponseTimeout :: Microsecond

How often request to NTP server and response collection

ntpPollDelay :: Microsecond

How often send request to NTP server