Pos.Context

Description

Re-exports of Pos.Context.* functionality.

Documentation

class WithNodeContext ssc m | m -> ssc where

Class for something that has NodeContext inside.

Minimal complete definition

getNodeContext

Methods

getNodeContext :: m (NodeContext ssc)

type LrcSyncData = (Bool, EpochIndex)

Data used for LRC syncronization. First value is False iff LRC is running now. Second value is last epoch for which we have already computed LRC.

data NodeContext ssc

NodeContext contains runtime context of node.

Constructors

NodeContext

Fields

ncJLFile :: !(Maybe (MVar FilePath))

ncSscContext :: !(SscNodeContext ssc)

ncBlkSemaphore :: !(MVar HeaderHash)

Semaphore which manages access to block application. Stored hash is a hash of last applied block.

ncLrcSync :: !(TVar LrcSyncData)

Primitive for synchronization with LRC.

ncUserSecret :: !(TVar UserSecret)

Secret keys (and path to file) which are used to send transactions

ncBlockRetrievalQueue :: !(TBQueue (NodeId, NewestFirst NE (BlockHeader ssc)))

Concurrent queue that holds block headers that are to be downloaded.

ncRecoveryHeader :: !(TMVar (NodeId, BlockHeader ssc))

In case of recovery mode this variable holds the latest header hash we know about so we can do chained block requests. Invariant: this mvar is full iff we're more than recoveryHeadersMessage blocks deep relatively to some valid header and we're downloading blocks. Every time we get block that's more difficult than this one, we overwrite. Every time we process some blocks and fail or see that we've downloaded this header, we clean mvar.

ncProgressHeader :: !(TMVar (BlockHeader ssc))

Header of the last block that was downloaded in retrieving queue. Is needed to show smooth prorgess on the frontend.

ncUpdateSemaphore :: !(MVar ConfirmedProposalState)

A semaphore which is unlocked when update data is downloaded and ready to apply.

ncInvPropagationQueue :: !RelayInvQueue

ncLoggerConfig :: !LoggerConfig

Logger config, as taken/read from CLI.

ncNodeParams :: !NodeParams

Params node is launched with

ncShutdownFlag :: !(TVar Bool)

If this flag is True, then workers should stop.

ncShutdownNotifyQueue :: !(TBQueue ())

A queue which is used to count how many workers have successfully terminated.

ncSendLock :: !(Maybe (MVar ()))

Exclusive lock for sending messages to other nodes (if Nothing, no lock used).

ncStartTime :: !UTCTime

Time when node was started (NodeContext initialized).

ncPublicKey :: NodeContext ssc -> PublicKey

Generate PublicKey from SecretKey of NodeContext.

ncPubKeyAddress :: NodeContext ssc -> Address

Generate Address from SecretKey of NodeContext

ncGenesisLeaders :: NodeContext ssc -> SlotLeaders

ncGenesisUtxo :: NodeContext ssc -> Utxo

ncSystemStart :: NodeContext __ -> Timestamp

data NodeParams

Contains algorithm specific & storage parameters for Node.

Constructors

NodeParams

Fields

npDbPathM :: !FilePath

Path to node's database.

npRebuildDb :: !Bool

True if data-base should be rebuilt

npSystemStart :: !Timestamp

System start

npSecretKey :: !SecretKey

Primary secret key of node

npUserSecret :: !UserSecret

All node secret keys

npBaseParams :: !BaseParams

See BaseParams

npCustomUtxo :: !Utxo

predefined custom utxo

npTimeLord :: !Bool

True if node started as time-lord

npJLFile :: !(Maybe FilePath)

npAttackTypes :: ![AttackType]

List of attack types used by malicious emulation

npAttackTargets :: ![AttackTarget]

List of targets to attack by malicious emulation

npPropagation :: !Bool

Whether to propagate txs, ssc data, blocks to neighbors

npUpdatePath :: !FilePath

Path to update installer executable, downloaded by update system

npUpdateWithPkg :: !Bool

If True then use installer update mechanism

npUpdateServers :: ![Text]

List of update server URLs

npReportServers :: ![Text]

List of report server URLs

data BaseParams

Contains basic & networking parameters for running node.

Constructors

BaseParams

Fields

bpIpPort :: !NetworkAddress

Port to run on

bpDHTPeers :: ![DHTNode]

Peers passed from CLI

bpDHTKey :: !(Maybe DHTKey)

bpDHTExplicitInitial :: !Bool

bpLoggingParams :: !LoggingParams

Logger parameters

bpKademliaDump :: !FilePath

Path to kademlia dump file

type RelayInvQueue = TBQueue SomeInvMsg

Queue of InvMsges which should be propagated.

data SomeInvMsg

Constructors

(Message (InvOrData tag key contents), Bi (InvOrData tag key contents), Buildable tag, Buildable key, Message (ReqMsg key tag), Bi (ReqMsg key tag)) => SomeInvMsg !(InvOrData tag key contents)

Genesis

genesisUtxoM :: (Functor m, WithNodeContext ssc m) => m Utxo

genesisLeadersM :: (Functor m, WithNodeContext ssc m) => m SlotLeaders

Block semaphore.

putBlkSemaphore :: (MonadIO m, WithNodeContext ssc m) => HeaderHash -> m ()

readBlkSemaphore :: (MonadIO m, WithNodeContext ssc m) => m HeaderHash

takeBlkSemaphore :: (MonadIO m, WithNodeContext ssc m) => m HeaderHash

LRC synchronization

waitLrc :: (MonadIO m, WithNodeContext ssc m) => EpochIndex -> m ()

Block until LRC data is available for given epoch.

lrcActionOnEpoch :: (MonadIO m, WithNodeContext ssc m, MonadThrow m) => EpochIndex -> (EpochIndex -> m (Maybe a)) -> m a

lrcActionOnEpochReason :: (MonadIO m, WithNodeContext ssc m, MonadThrow m) => EpochIndex -> Text -> (EpochIndex -> m (Maybe a)) -> m a

Misc

getUptime :: (MonadIO m, WithNodeContext ssc m) => m Microsecond

Returns node uptime based on current time and ncStartTime.

isRecoveryMode :: (MonadIO m, WithNodeContext ssc m) => m Bool

Returns if ncRecoveryHeader is Just which is equivalent to "we're in recovery mode".

newtype ContextHolder ssc m a

Wrapper for monadic action which brings NodeContext.

Constructors

ContextHolder

Fields

getContextHolder :: ReaderT (NodeContext ssc) m a

runContextHolder :: NodeContext ssc -> ContextHolder ssc m a -> m a

Run ContextHolder action.