Pos.DB.Block

Documentation

data BlundAux

Structure describes location of block and undo with specific header hash. Block and corresponding Undo are stored in same file.

Constructors

BlundAux

Fields

blundFileNum :: !Word32

In which file the blund is stored.

blockOffset :: !Word32

Offset of blob of the block in a file.

blockLen :: !Word32

Length of blob of the block.

undoOffset :: !Word32

Offset of blob of the undo in a file.

undoLen :: !Word32

Length of blob of the undo.

getBlock :: (SscHelpersClass ssc, MonadDB m) => HeaderHash -> m (Maybe (Block ssc))

Get block with given hash from Block DB.

getBlockHeader :: (SscHelpersClass ssc, MonadDB m) => HeaderHash -> m (Maybe (BlockHeader ssc))

Returns header of block that was requested from Block DB.

getUndo :: MonadDB m => HeaderHash -> m (Maybe Undo)

Get undo data for block with given hash from Block DB.

getBlockWithUndo :: (SscHelpersClass ssc, MonadDB m) => HeaderHash -> m (Maybe (Block ssc, Undo))

Retrieves block and undo together.

putBlund :: (SscHelpersClass ssc, MonadDB m) => Block ssc -> Undo -> m ()

Put given block, its metadata and Undo data into Block DB.

prepareBlockDB :: forall ssc m. (SscHelpersClass ssc, MonadDB m) => GenesisBlock ssc -> m ()

Load data

loadBlundsWhile :: (SscHelpersClass ssc, MonadDB m) => (Block ssc -> Bool) -> HeaderHash -> m (NewestFirst [] (Blund ssc))

Load blunds starting from block with header hash equal to given hash and while predicate is true.

loadBlundsByDepth :: (SscHelpersClass ssc, MonadDB m) => Word -> HeaderHash -> m (NewestFirst [] (Blund ssc))

Load blunds which have depth less than given.

loadBlocksWhile :: (SscHelpersClass ssc, MonadDB m) => (Block ssc -> Bool) -> HeaderHash -> m (NewestFirst [] (Block ssc))

Load blocks starting from block with header hash equal to given hash and while predicate is true.

loadHeadersWhile :: (SscHelpersClass ssc, MonadDB m) => (BlockHeader ssc -> Bool) -> HeaderHash -> m (NewestFirst [] (BlockHeader ssc))

Load headers starting from block with header hash equal to given hash and while predicate is true.

loadHeadersByDepth :: (SscHelpersClass ssc, MonadDB m) => Word -> HeaderHash -> m (NewestFirst [] (BlockHeader ssc))

Load headers which have depth less than given.

loadHeadersByDepthWhile :: (SscHelpersClass ssc, MonadDB m) => (BlockHeader ssc -> Bool) -> Word -> HeaderHash -> m (NewestFirst [] (BlockHeader ssc))

Load headers which have depth less than given and match some criterion.