Pos.DB.GState

Description

GState DB stores whole global state, i. e. result of application of all blocks up to tip.

Getters

getTotalFtsStake :: MonadDB ssc m => m Coin

Get total amount of stake to be used for follow-the-satoshi. It's different from total amount of coins in the system.

getFtsStake :: MonadDB ssc m => StakeholderId -> m (Maybe Coin)

Get stake owne by given stakeholder (according to rules used for FTS).

getFtsStakeFromDB :: (MonadIO m, MonadThrow m) => StakeholderId -> DB ssc -> m (Maybe Coin)

Operations

data BalancesOp

Constructors

PutFtsSum !Coin

PutFtsStake !StakeholderId !Coin

Initialization

prepareGStateBalances :: forall ssc m. MonadDB ssc m => Utxo -> m ()

Iteration

data BalIter

runBalanceIterator :: forall m ssc a. MonadDB ssc m => DBnIterator ssc BalIter a -> m a

runBalanceMapIterator :: forall v m ssc a. MonadDB ssc m => DBnMapIterator ssc BalIter v a -> (IterType BalIter -> v) -> m a

Sanity checks

sanityCheckBalances :: (MonadMask m, MonadDB ssc m, WithLogger m) => m ()

resolveForwardLink :: (HasHeaderHash a, MonadDB ssc m) => a -> m (Maybe HeaderHash)

Tries to retrieve next block using current one (given a block/header).

isBlockInMainChain :: (HasHeaderHash a, MonadDB ssc m) => a -> m Bool

Check if given hash representing block is in main chain.

data BlockExtraOp

Constructors

AddForwardLink HeaderHash HeaderHash
Adds or overwrites forward link

RemoveForwardLink HeaderHash
Removes forward link

SetInMainChain Bool HeaderHash
Enables or disables "in main chain" status of the block

loadHeadersUpWhile :: (SscHelpersClass ssc, MonadDB ssc m, HasHeaderHash a) => a -> (BlockHeader ssc -> Int -> Bool) -> m (OldestFirst [] (BlockHeader ssc))

Returns headers loaded up.

loadBlocksUpWhile :: (SscHelpersClass ssc, MonadDB ssc m, HasHeaderHash a) => a -> (Block ssc -> Int -> Bool) -> m (OldestFirst [] (Block ssc))

Returns blocks loaded up.

prepareGStateBlockExtra :: MonadDB ssc m => HeaderHash -> m ()

Getters

getTip :: MonadDB ssc m => m HeaderHash

Get current tip from GState DB.

getBot :: MonadDB ssc m => m HeaderHash

Get the hash of the first genesis block from GState DB.

Initialization

prepareGStateCommon :: forall ssc m. MonadDB ssc m => HeaderHash -> m ()

Put missing initial common data into GState DB.

Helpers

getBi :: (MonadDB ssc m, Bi v) => ByteString -> m (Maybe v)

putBi :: (MonadDB ssc m, Bi v) => ByteString -> v -> m ()

delete :: MonadDB ssc m => ByteString -> m ()

writeBatchGState :: (RocksBatchOp a, MonadDB ssc m) => [a] -> m ()

Operations

data CommonOp

Constructors

PutTip HeaderHash

getPSKByIssuerAddressHash :: MonadDB ssc m => StakeholderId -> m (Maybe ProxySKHeavy)

Retrieves certificate by issuer address (hash of public key) if present.

getPSKByIssuer :: MonadDB ssc m => PublicKey -> m (Maybe ProxySKHeavy)

Retrieves certificate by issuer public key if present.

isIssuerByAddressHash :: MonadDB ssc m => StakeholderId -> m Bool

data DelegationOp

Constructors

AddPSK !ProxySKHeavy
Adds PSK. Overwrites if present.

DelPSK !PublicKey
Removes PSK by issuer PK.

runPskIterator :: forall m ssc a. MonadDB ssc m => DBnIterator ssc PskIter a -> m a

runPskMapIterator :: forall v m ssc a. MonadDB ssc m => DBnMapIterator ssc PskIter v a -> (IterType PskIter -> v) -> m a

prepareGStateDB :: forall ssc m. (WithNodeContext ssc m, MonadDB ssc m) => HeaderHash -> m ()

Put missing initial data into GState DB.

sanityCheckGStateDB :: forall ssc m. (MonadDB ssc m, MonadMask m, WithLogger m) => m ()

Check that GState DB is consistent.

usingGStateSnapshot :: (MonadDB ssc m, MonadMask m) => m a -> m a

Getters

getAdoptedBV :: MonadDB ssc m => m BlockVersion

Get last adopted block version.

getAdoptedBVData :: MonadDB ssc m => m BlockVersionData

Get state of last adopted BlockVersion.

getAdoptedBVFull :: MonadDB ssc m => m (BlockVersion, BlockVersionData)

Get last adopted BlockVersion and data associated with it.

getBVState :: MonadDB ssc m => BlockVersion -> m (Maybe BlockVersionState)

Get BlockVersionState associated with given BlockVersion.

getProposalState :: MonadDB ssc m => UpId -> m (Maybe ProposalState)

Get state of UpdateProposal for given UpId

getAppProposal :: MonadDB ssc m => ApplicationName -> m (Maybe UpId)

Get UpId of current proposal for given appName

getProposalStateByApp :: MonadDB ssc m => ApplicationName -> m (Maybe ProposalState)

Get state of Update Proposal for given AppName

getConfirmedSV :: MonadDB ssc m => ApplicationName -> m (Maybe NumSoftwareVersion)

Get last confirmed SoftwareVersion of given application.

getMaxBlockSize :: MonadDB ssc m => m Byte

Get maximum block size (in bytes).

getSlottingData :: MonadDB ssc m => m SlottingData

Get most recent SlottingData.

Operations

data UpdateOp

Constructors

PutProposal !ProposalState

DeleteProposal !UpId !ApplicationName

ConfirmVersion !SoftwareVersion

DelConfirmedVersion !ApplicationName

AddConfirmedProposal !ConfirmedProposalState

DelConfirmedProposal !SoftwareVersion

SetAdopted !BlockVersion BlockVersionData

SetBVState !BlockVersion !BlockVersionState

DelBV !BlockVersion

PutSlottingData !SlottingData

Initialization

prepareGStateUS :: forall ssc m. MonadDB ssc m => Timestamp -> m ()

Iteration

data PropIter

runProposalMapIterator :: forall v m ssc a. MonadDB ssc m => DBnMapIterator ssc PropIter v a -> (IterType PropIter -> v) -> m a

runProposalIterator :: forall m ssc a. MonadDB ssc m => DBnIterator ssc PropIter a -> m a

getOldProposals :: forall ssc m. MonadDB ssc m => SlotId -> m [UndecidedProposalState]

Get all proposals which were issued no later than given slot.

getDeepProposals :: forall ssc m. MonadDB ssc m => ChainDifficulty -> m [DecidedProposalState]

Get all decided proposals which were accepted deeper than given difficulty.

data ConfPropIter

getConfirmedProposals :: MonadDB ssc m => Maybe NumSoftwareVersion -> m [ConfirmedProposalState]

Get confirmed proposals which update our application and have version bigger than argument (or all proposals if Nothing is passed). For instance, current software version can be passed to this function to get all proposals with bigger version.

data BVIter

getProposedBVs :: MonadDB ssc m => m [BlockVersion]

Get all proposed BlockVersions.

getConfirmedBVStates :: MonadDB ssc m => m [(BlockVersion, BlockVersionState)]

Get all confirmed BlockVersions and their states.

getProposedBVStates :: MonadDB ssc m => m [BlockVersionState]

Getters

getGenUtxo :: (MonadThrow m, MonadDB ssc m) => m Utxo

Get genesis utxo

getTxOutFromDB :: (MonadIO m, MonadThrow m) => TxIn -> DB ssc -> m (Maybe TxOutAux)

getTxOut :: MonadDB ssc m => TxIn -> m (Maybe TxOutAux)

Operations

data UtxoOp

Constructors

DelTxIn !TxIn

AddTxOut !TxIn !TxOutAux

Initialization

prepareGStateUtxo :: forall ssc m. MonadDB ssc m => Utxo -> m ()

Iteration

data UtxoIter

runUtxoIterator :: forall m ssc a. MonadDB ssc m => DBnIterator ssc UtxoIter a -> m a

runUtxoMapIterator :: forall v m ssc a. MonadDB ssc m => DBnMapIterator ssc UtxoIter v a -> (IterType UtxoIter -> v) -> m a

getFilteredUtxo :: MonadDB ssc m => Address -> m Utxo

Get small sub-utxo containing only outputs of given address

Sanity checks

sanityCheckUtxo :: (MonadDB ssc m, WithLogger m) => Coin -> m ()