Pos.Lrc

Description

Leaders and richmen computation.

Documentation

data LrcConsumer m

Datatype for LRC computation client. If you want to compute richmen, you should add such client to LRC framework

Constructors

LrcConsumer

Fields

lcThreshold :: Coin -> Coin

Function which defines threshold depends on total stake

lcIfNeedCompute :: EpochIndex -> m Bool

Function which defines necessity of richmen computation

lcComputedCallback :: EpochIndex -> Coin -> RichmenStake -> m ()

Callback which will be called when richmen computed

lcConsiderDelegated :: Bool

Whether delegated stake should be considered

lrcConsumerFromComponent :: forall c m. RichmenComponent c => (EpochIndex -> m Bool) -> (EpochIndex -> Coin -> RichmenStake -> m ()) -> LrcConsumer m

Create LrcConsumer using constants from RichmenComponent and given actions.

lrcConsumerFromComponentSimple :: forall c ssc m. (RichmenComponent c, MonadDB ssc m) => LrcConsumer m

Create simple LrcConsumer using constants from RichmenComponent which uses only LRC DB.

allLrcConsumers :: (SscWorkersClass ssc, WorkMode ssc m) => [LrcConsumer m]

findRichmenStake

Arguments

:: MonadIterator (StakeholderId, Coin) m

=> Coin
Eligibility threshold

-> m RichmenStake

Find nodes which have at least 'eligibility threshold' coins.

findRichmenPure :: [(StakeholderId, Coin)] -> (Coin -> Coin) -> RichmenType -> FullRichmenData

Pure version of findRichmen which uses in-memory Utxo.

findAllRichmenMaybe

Arguments

:: (MonadDB ssc m, MonadIterator (StakeholderId, Coin) m)

=> Maybe Coin
Eligibility threshold (optional)

-> Maybe Coin
Delegation threshold (optional)

-> m (RichmenStake, RichmenStake)

Function considers all variants of computation and compute using one pass by stake DB and one pass by delegation DB.

findDelegatedRichmen :: (MonadDB ssc m, MonadIterator (StakeholderId, Coin) m) => Coin -> m RichmenStake

Find delegated richmen.

data RichmenType

Constructors

RTUsual

RTDelegation (HashMap StakeholderId [StakeholderId])

module Pos.Lrc.FollowTheSatoshi

type Richmen = NonEmpty StakeholderId

Addresses which have enough stake for participation in SSC.

type RichmenSet = HashSet StakeholderId

Hashset of richmen.

type RichmenStake = HashMap StakeholderId Coin

Richmen with Stake

toRichmen :: RichmenStake -> Richmen

type FullRichmenData = (Coin, RichmenStake)

Full richmen data consists of total stake at some point and stake distribution among richmen.