Pos.Reporting.Methods

Description

Methods of reporting different unhealthy behaviour to server.

Documentation

sendReportNode :: (MonadIO m, MonadCatch m, WithNodeContext її m) => ReportType -> m ()

Sends node's logs, taking LoggerConfig from NodeContext, retrieving all logger files from it. List of servers is also taken from node's configuration.

getNodeInfo :: WorkMode ssc m => m Text

Retrieves node info that we would like to know whne analyzing malicious behavior of node.

sendReport :: (MonadIO m, MonadCatch m) => [FilePath] -> ReportType -> Text -> Version -> String -> m ()

Given logs files list and report type, sends reports to URI asked. All files _must_ exist. Report server URI should be in form like "http(s):/host:port" without specified endpoint.

retrieveLogFiles :: LoggerTree -> [([Text], FilePath)]

Given logger config, retrieves all (logger name, filepath) for every logger that has file handle.

chooseLogFiles :: MonadIO m => FilePath -> m [FilePath]

Retrieves real filepathes of logs given filepathes from log description. Example: there's component.log in config, but this function will return [component.log.122, component.log.123].