Pos.Slotting

Description

Slotting functionality.

Documentation

class Monad m => MonadSlotsData m where

MonadSlotsData provides access to data necessary for slotting to work.

Methods

getSlottingData :: m SlottingData

getSlottingData :: (MonadTrans t, MonadSlotsData m', t m' ~ m) => m SlottingData

class MonadSlotsData m => MonadSlots m where

Type class providing information about current slot.

Methods

getCurrentSlot :: m SlotId

getCurrentSlot :: (MonadTrans t, MonadSlots m', t m' ~ m) => m SlotId

newtype DBSlotsData m a

Monad transformer which provides SlottingData using DB.

Constructors

DBSlotsData

Fields

runDBSlotsData :: m a

data NtpSlottingState

Data needed for the slotting algorithm to work.

type NtpSlottingVar = TVar NtpSlottingState

newtype NtpSlotting m a

Monad transformer which implements NTP-based solution for slotting.

Constructors

NtpSlotting

Fields

getNtpSlotting :: ReaderT NtpSlottingVar m a

runNtpSlotting :: Mockable CurrentTime m => NtpSlotting m a -> m a

runNtpSlottingFromVar :: NtpSlottingVar -> NtpSlotting m a -> m a

data SlottingData

Constructors

SlottingData

Fields

sdEpoch :: !EpochIndex

sdSlotDuration :: !Millisecond

sdEpochStart :: !EpochIndex

Helpers using MonadSlots

getCurrentSlotFlat :: MonadSlots m => m FlatSlotId

Get flat id of current slot based on MonadSlots.

getSlotStart :: MonadSlots m => SlotId -> m Timestamp

Get timestamp when given slot starts.

Worker which ticks when slot starts

onNewSlot :: OnNewSlot ssc m => Bool -> (SlotId -> m ()) -> m ()

Run given action as soon as new slot starts, passing SlotId to it. This function uses Mockable and assumes consistency between MonadSlots and Mockable implementations.

onNewSlotImpl :: OnNewSlot ssc m => Bool -> Bool -> (SlotId -> m ()) -> m ()