Pos.Slotting

Description

Slotting functionality.

Documentation

class Monad m => MonadSlotsData m where

MonadSlotsData provides access to data necessary for slotting to work.

Methods

getSlottingData :: m SlottingData

waitPenultEpochEquals :: EpochIndex -> m ()

putSlottingData :: SlottingData -> m ()

getSlottingData :: (MonadTrans t, MonadSlotsData m', t m' ~ m) => m SlottingData

waitPenultEpochEquals :: (MonadTrans t, MonadSlotsData m', t m' ~ m) => EpochIndex -> m ()

putSlottingData :: (MonadTrans t, MonadSlotsData m', t m' ~ m) => SlottingData -> m ()

class MonadSlotsData m => MonadSlots m where

Type class providing information about current slot.

Methods

getCurrentSlot :: m (Maybe SlotId)

getCurrentSlotBlocking :: m SlotId

Blocking version of getCurrentSlot. This function doesn't return until current slot is known.

getCurrentSlotInaccurate :: m SlotId

This function tries to predict current slot as accurately as it can. If getCurrentTime returns unreliable time, then function returns last known slot If our slotting data into DB is outdated, then function tries to extrapolate slot using last know slotting data

currentTimeSlotting :: m Timestamp

slottingWorkers :: [m ()]

getCurrentSlot :: (MonadTrans t, MonadSlots m', t m' ~ m) => m (Maybe SlotId)

getCurrentSlotBlocking :: (MonadTrans t, MonadSlots m', t m' ~ m) => m SlotId

Blocking version of getCurrentSlot. This function doesn't return until current slot is known.

currentTimeSlotting :: (MonadTrans t, MonadSlots m', t m' ~ m) => m Timestamp

slottingWorkers :: (MonadTrans t, MonadSlots m', t m' ~ m) => [m ()]

getCurrentSlotInaccurate :: (MonadTrans t, MonadSlots m', t m' ~ m) => m SlotId

This function tries to predict current slot as accurately as it can. If getCurrentTime returns unreliable time, then function returns last known slot If our slotting data into DB is outdated, then function tries to extrapolate slot using last know slotting data

data SlottingError

Type aggregating run-time errors related to Slotting.

Constructors

SEUnknownSlotStart !SlotId

newtype SlottingHolder m a

Monad transformer which provides SlottingData using DB.

Constructors

SlottingHolder

Fields

getSlottingHolder :: ReaderT SlottingVar m a

type SlottingVar = TVar SlottingData

runSlottingHolder :: SlottingVar -> SlottingHolder m a -> m a

Run USHolder using existing SlottingVar.

mkSlottingVar :: MonadDB ε m => m SlottingVar

Create new SlottingVar using data from DB.

data NtpSlottingState

Data needed for the slotting algorithm to work.

type NtpSlottingVar = TVar NtpSlottingState

newtype NtpSlotting m a

Monad transformer which implements NTP-based solution for slotting.

Constructors

NtpSlotting

Fields

getNtpSlotting :: ReaderT NtpSlottingVar m a

mkNtpSlottingVar :: (MonadIO m, Mockable CurrentTime m, WithLogger m, Mockable Delay m, Mockable Fork m, Mockable Throw m, Mockable Catch m) => m NtpSlottingVar

runNtpSlotting :: NtpSlottingVar -> NtpSlotting m a -> m a

data EpochSlottingData

Data which is necessary for slotting and corresponds to a particular epoch.

Constructors

EpochSlottingData

Fields

esdSlotDuration :: !Millisecond

Slot duration actual for given epoch.

esdStart :: !Timestamp

Time when epoch starts.

data SlottingData

Data necessary for slotting to work which is basically part of GState.

Constructors

SlottingData

Fields

sdPenult :: !EpochSlottingData

EpochSlottingData for penult (i. e. sdPenultEpoch) epoch.

sdLast :: !EpochSlottingData

EpochSlottingData for last epoch. Last means 'sdPenultEpoch + 1'.

sdPenultEpoch :: !EpochIndex

Penult epoch for which EpochSlottingData is known.

Helpers using 'MonadSlots[Data]'

getCurrentSlotFlat :: MonadSlots m => m (Maybe FlatSlotId)

Get flat id of current slot based on MonadSlots.

getSlotStart :: MonadSlotsData m => SlotId -> m (Maybe Timestamp)

Get timestamp when given slot starts.

getSlotStartEmpatically :: (MonadSlotsData m, MonadThrow m) => SlotId -> m Timestamp

Get timestamp when given slot starts empatically, which means that function throws exception when slot start is unknown.

getLastKnownSlotDuration :: MonadSlotsData m => m Millisecond

Get last known slot duration.

Worker which ticks when slot starts

onNewSlot :: OnNewSlot ssc m => Bool -> (SlotId -> m ()) -> m ()

Run given action as soon as new slot starts, passing SlotId to it. This function uses Mockable and assumes consistency between MonadSlots and Mockable implementations.

onNewSlotImpl :: forall ssc m. OnNewSlot ssc m => Bool -> Bool -> (SlotId -> m ()) -> m ()

Worker which logs beginning of new slot

logNewSlotWorker :: OnNewSlot ssc m => m ()

Waiting for system start

waitSystemStart :: (WithNodeContext ssc m, Mockable Delay m, WithLogger m, MonadSlots m) => m ()

Wait until system starts. This function is useful if node is launched before 0-th epoch starts.