Pos.Slotting

Description

Slotting functionality.

Documentation

class Monad m => MonadSlots m where

Type class providing information about time when system started functioning.

Methods

getSystemStartTime :: m Timestamp

getCurrentTime :: m Timestamp

getCurrentSlot :: m SlotId

getSystemStartTime :: (MonadTrans t, MonadSlots m', t m' ~ m) => m Timestamp

getCurrentTime :: (MonadTrans t, MonadSlots m', t m' ~ m) => m Timestamp

getCurrentSlot :: (MonadTrans t, MonadSlots m', t m' ~ m) => m SlotId

getCurrentSlotFlat :: MonadSlots m => m FlatSlotId

Get flat id of current slot based on MonadSlots.

getSlotStart :: MonadSlots m => SlotId -> m Timestamp

Get timestamp when given slot starts.

getCurrentSlotUsingNtp :: (MonadSlots m, Mockable CurrentTime m) => SlotId -> (Microsecond, Microsecond) -> m SlotId

onNewSlot :: (MonadIO m, MonadSlots m, MonadCatch m, WithLogger m, Mockable Fork m, Mockable Delay m) => Bool -> (SlotId -> m ()) -> m a

Run given action as soon as new slot starts, passing SlotId to it. This function uses Mockable and assumes consistency between MonadSlots and Mockable implementations.

onNewSlotWithLogging :: (MonadIO m, MonadSlots m, MonadCatch m, WithLogger m, Mockable Fork m, Mockable Delay m) => Bool -> (SlotId -> m ()) -> m a

Same as onNewSlot, but also logs debug information.