Pos.Ssc.GodTossing

Description

Re-exports of GodTossing modules.

GodTossing is a coin tossing with guaranteed output delivery. Nodes exchange commitments, openings, and shares, and in the end arrive at a shared seed.

See https://eprint.iacr.org/2015/889.pdf (“A Provably Secure Proof-of-Stake Blockchain Protocol”), section 4 for more details.

Documentation

data CommitmentOpening

Pair of Commitment and Opening.

Constructors

CommitmentOpening

Fields

coCommitment :: !Commitment

coOpening :: !Opening

Helpers

genCommitmentAndOpening :: (MonadFail m, MonadIO m) => Threshold -> NonEmpty (AsBinary VssPublicKey) -> m (Commitment, Opening)

Generate securely random SharedSeed.

isCommitmentId :: SlotId -> Bool

isCommitmentIdx :: LocalSlotIndex -> Bool

isOpeningId :: SlotId -> Bool

isOpeningIdx :: LocalSlotIndex -> Bool

isSharesId :: SlotId -> Bool

isSharesIdx :: LocalSlotIndex -> Bool

mkSignedCommitment :: Bi Commitment => SecretKey -> EpochIndex -> Commitment -> SignedCommitment

Make signed commitment from commitment and epoch index using secret key.

secretToSharedSeed :: Secret -> SharedSeed

Convert Secret to SharedSeed.

CommitmentsMap

insertSignedCommitment :: SignedCommitment -> CommitmentsMap -> CommitmentsMap

Safely insert SignedCommitment into CommitmentsMap.

deleteSignedCommitment :: StakeholderId -> CommitmentsMap -> CommitmentsMap

Safely delete SignedCommitment from CommitmentsMap.

diffCommMap :: CommitmentsMap -> CommitmentsMap -> CommitmentsMap

Compute difference of two CommitmentsMaps.

intersectCommMap :: CommitmentsMap -> CommitmentsMap -> CommitmentsMap

Compute intersection of two CommitmentsMaps.

intersectCommMapWith :: (map -> HashMap StakeholderId x) -> CommitmentsMap -> map -> CommitmentsMap

Generalized version of intersectCommMap which makes it possible to intersect with different maps.

Verification and Checks

checkCertTTL :: EpochIndex -> VssCertificate -> Bool

Check that the VSS certificate has valid TTL: i. e. it is in '[vssMinTTL, vssMaxTTL]'.

verifyCommitment :: Commitment -> Bool

Verify that Commitment is correct.

#verifyEncShare

verifyCommitmentSignature :: Bi Commitment => EpochIndex -> SignedCommitment -> Bool

Verify signature in SignedCommitment using epoch index.

#checkSig

verifySignedCommitment :: Bi Commitment => EpochIndex -> SignedCommitment -> VerificationRes

Verify SignedCommitment using public key and epoch index.

#verifyCommitmentSignature #verifyCommitment

verifyOpening :: Commitment -> Opening -> Bool

Verify that Secret provided with Opening corresponds to given commitment.

#verifySecretProof

Payload and proof

_gpCertificates :: GtPayload -> VssCertificatesMap

mkGtProof :: GtPayload -> GtProof

Construct GtProof from GtPayload.

Commitments

data Commitment

Commitment is a message generated during the first stage of GodTossing. It contains encrypted shares and proof of secret. Invariant which must be ensured: commShares is not empty.

Constructors

Commitment

Fields

commExtra :: !(AsBinary SecretSharingExtra)

commProof :: !(AsBinary SecretProof)

commShares :: !(HashMap (AsBinary VssPublicKey) (AsBinary EncShare))

type CommitmentSignature = Signature (EpochIndex, Commitment)

Signature which ensures that commitment was generated by node with given public key for given epoch.

type SignedCommitment = (PublicKey, Commitment, CommitmentSignature)

data CommitmentsMap

CommitmentsMap is a wrapper for 'HashMap StakeholderId SignedCommitment' which ensures that keys are consistent with values, i. e. PublicKey from SignedCommitment corresponds to key which is StakeholderId.

mkCommitmentsMap :: [SignedCommitment] -> CommitmentsMap

Safe constructor of CommitmentsMap.

mkCommitmentsMapUnsafe :: HashMap StakeholderId SignedCommitment -> CommitmentsMap

Unsafe straightforward constructor of CommitmentsMap.

Openings

newtype Opening

Opening reveals secret.

Constructors

Opening

Fields

getOpening :: AsBinary Secret

type OpeningsMap = HashMap StakeholderId Opening

Shares

type InnerSharesMap = HashMap StakeholderId (AsBinary Share)

type SharesMap = HashMap StakeholderId InnerSharesMap

Each node generates a SharedSeed, breaks it into Shares, and sends those encrypted shares to other nodes. In a SharesMap, for each node we collect shares which said node has received and decrypted.

Specifically, if node identified by Address X has received a share from node identified by key Y, this share will be at sharesMap ! X ! Y.

Vss certificates

data VssCertificate

VssCertificate allows VssPublicKey to participate in MPC. Each stakeholder should create a Vss keypair, sign VSS public key with signing key and send it into blockchain.

A public key of node is included in certificate in order to enable validation of it using only node's P2PKH address. Expiry epoch is last epoch when certificate is valid, expiry epoch is included in certificate and signature.

Other nodes accept this certificate if it is valid and if node has enough stake.

Invariant: 'checkSig vcSigningKey (vcVssKey, vcExpiryEpoch) vcSignature'.

mkVssCertificate :: SecretKey -> AsBinary VssPublicKey -> EpochIndex -> VssCertificate

Make VssCertificate valid up to given epoch using SecretKey to sign data.

recreateVssCertificate :: MonadFail m => AsBinary VssPublicKey -> EpochIndex -> Signature (AsBinary VssPublicKey, EpochIndex) -> PublicKey -> m VssCertificate

Recreate VssCertificate from its contents. This function main fail if data is invalid.

getCertId :: VssCertificate -> StakeholderId

type VssCertificatesMap = HashMap StakeholderId VssCertificate

VssCertificatesMap contains all valid certificates collected during some period of time.

mkVssCertificatesMap :: [VssCertificate] -> VssCertificatesMap

Safe constructor of VssCertificatesMap. TODO: wrap into newtype.

Payload

data GtPayload

Payload included into blocks.

Constructors

CommitmentsPayload !CommitmentsMap !VssCertificatesMap

OpeningsPayload !OpeningsMap !VssCertificatesMap

SharesPayload !SharesMap !VssCertificatesMap

CertificatesPayload !VssCertificatesMap

data GtProof

Proof of GtPayload.

Constructors

CommitmentsProof !(Hash CommitmentsMap) !(Hash VssCertificatesMap)

OpeningsProof !(Hash OpeningsMap) !(Hash VssCertificatesMap)

SharesProof !(Hash SharesMap) !(Hash VssCertificatesMap)

CertificatesProof !(Hash VssCertificatesMap)

Misc

type NodeSet = HashSet StakeholderId

data SeedError

Data type for error during seed calculation.

Constructors

ExtraneousOpenings (HashSet StakeholderId)
Some nodes in the OpeningsMap aren't in the set of participants

ExtraneousShares (HashSet StakeholderId)
Some nodes in the SharesMap aren't in the set of participants

NoParticipants
There were no participants so a random string couldn't be generated

BrokenCommitment StakeholderId
Commitment can't be deserialized or didn't match secret (either recovered or in openings)

NoSecretFound StakeholderId
Secret couldn't be recovered, or wasn't found in either OpeningsMap or SharesMap

BrokenSecret StakeholderId
Secret can't be deserialized

BrokenShare StakeholderId
Share can't be deserialized

hasCommitment :: StakeholderId -> GtGlobalState -> Bool

hasOpening :: StakeholderId -> GtGlobalState -> Bool

hasShares :: StakeholderId -> GtGlobalState -> Bool

hasVssCertificate :: StakeholderId -> GtGlobalState -> Bool

GtPayload

verifyGtPayload :: MonadError TossVerFailure m => Either EpochIndex (MainBlockHeader ssc) -> GtPayload -> m ()

VSS

vssThreshold :: Integral a => a -> Threshold

Figure out the threshold (i.e. how many secret shares would be required to recover each node's secret) using number of participants.

getStableCertsPure :: EpochIndex -> VssCertData -> VssCertificatesMap

genesisCertificates :: VssCertificatesMap

Certificates in genesis represented as VssCertificatesMap.

genesisVssKeyPairs :: [VssKeyPair]

List of VssKeyPair in genesis.

'Inv|Req|Data' processing.

sscIsDataUseful :: (WithLogger m, MonadDB SscGodTossing m, WithNodeContext kek m, MonadSlots m, MonadSscMem SscGodTossing m) => GtTag -> StakeholderId -> m Bool

Check whether SSC data with given tag and public key can be added to current local data.

sscProcessCommitment :: forall m. GtDataProcessingMode m => SignedCommitment -> m ()

Process SignedCommitment received from network, checking it against current state (global + local) and adding to local state if it's valid.

sscProcessOpening :: GtDataProcessingMode m => StakeholderId -> Opening -> m ()

Process Opening received from network, checking it against current state (global + local) and adding to local state if it's valid.

sscProcessShares :: GtDataProcessingMode m => StakeholderId -> InnerSharesMap -> m ()

Process InnerSharesMap received from network, checking it against current state (global + local) and adding to local state if it's valid.

sscProcessCertificate :: GtDataProcessingMode m => VssCertificate -> m ()

Process VssCertificate received from network, checking it against current state (global + local) and adding to local state if it's valid.

Garbage collection worker

localOnNewSlot :: MonadSscMem SscGodTossing m => SlotId -> m ()

Clean-up some data when new slot starts. This function is only needed for garbage collection, it doesn't affect validity of local data. Currently it does nothing, but maybe later we'll decide to do clean-up.

Instances

instance SscLocalDataClass SscGodTossing

data GtLocalData

Constructors

GtLocalData

Fields

_ldModifier :: !TossModifier

TossModifier which also serves as mempool of GT data, because for GodTossing modifier and mempool are same.

_ldEpoch :: !EpochIndex

Epoch for which this mempool can be used to form payload.

ldModifier :: Lens' GtLocalData TossModifier

ldEpoch :: Lens' GtLocalData EpochIndex

getOurCommitment :: MonadDB SscGodTossing m => EpochIndex -> m (Maybe SignedCommitment)

Get our commitment for given epoch if it's known.

getOurOpening :: MonadDB SscGodTossing m => EpochIndex -> m (Maybe Opening)

Get our opening corresponding for given epoch if it's known.

putOurSecret :: MonadDB SscGodTossing m => SignedCommitment -> Opening -> EpochIndex -> m ()

Put our secret for given epoch.

calculateSeed

Arguments

:: CommitmentsMap
All participating nodes

-> OpeningsMap

-> SharesMap
Decrypted shares

-> Either SeedError SharedSeed

Calculate SharedSeed. SharedSeed is a random bytestring that all nodes generate together and agree on.

TODO: do we need to check secrets' lengths? Probably not.

Instances

instance SscGStateClass SscGodTossing

getGlobalCerts :: (MonadSscMem SscGodTossing m, MonadIO m) => SlotId -> m VssCertificatesMap

gtGetGlobalState :: (MonadSscMem SscGodTossing m, MonadIO m) => m GtGlobalState

getStableCerts :: (MonadSscMem SscGodTossing m, MonadIO m) => EpochIndex -> m VssCertificatesMap

Get stable VSS certificates for given epoch.

data SscGodTossing

Data type which represents shared seed calculation tag in -XTypeApplication hacks with type families.

Instances

instance Ssc SscGodTossing

data GtTag

Tag corresponding to GodTossing data.

Constructors

CommitmentMsg

OpeningMsg

SharesMsg

VssCertificateMsg

data GtMsgContents

Data message. Can be used to send actual data.

Constructors

MCCommitment !SignedCommitment

MCOpening !StakeholderId !Opening

MCShares !StakeholderId !InnerSharesMap

MCVssCertificate !VssCertificate

msgContentsTag :: GtMsgContents -> GtTag

GtTag appropriate for given DataMsg.

Instance types

data GtGlobalState

Global state of GodTossing, contains relevant SSC data from blocks.

Constructors

GtGlobalState

Fields

_gsCommitments :: !CommitmentsMap

Commitments are added during the first phase of epoch.

_gsOpenings :: !OpeningsMap

Openings are added during the second phase of epoch.

_gsShares :: !SharesMap

Decrypted shares to be used in the third phase.

_gsVssCertificates :: !VssCertData

Vss certificates are added at any time if they are valid and received from stakeholders.

data GtContext

Constructors

GtContext

Fields

gtcVssKeyPair :: !VssKeyPair

Vss key pair used for MPC.

gtcParticipateSsc :: !(TVar Bool)

Flag which determines whether we want to participate in SSC.

data GtParams

Constructors

GtParams

Fields

gtpSscEnabled :: !Bool

Whether node should participate in SSC in case SSC requires participation.

gtpVssKeyPair :: !VssKeyPair

Key pair used for secret sharing

data GtSecretStorage

Constructors

GtSecretStorage

Fields

gssCommitment :: !SignedCommitment

Our commitment.

gssOpening :: !Opening

Corresponding opening

gssEpoch :: !EpochIndex

Epoch for which this secret were generated

Lenses

GtPayload

gsCommitments :: Lens' GtGlobalState CommitmentsMap

gsOpenings :: Lens' GtGlobalState OpeningsMap

gsShares :: Lens' GtGlobalState SharesMap

gsVssCertificates :: Lens' GtGlobalState VssCertData

createGtContext :: MonadIO m => GtParams -> m GtContext

data VssCertData

Wrapper around VssCertificate with TTL. Every VssCertificate has own TTL. Wrapper supports simple HashMap operations. Wrapper holds VssCertificatesMap and Set of certificates sorted by expiry epoch.

Constructors

VssCertData

Fields

lastKnownEoS :: !EpochOrSlot

Last known slot, every element of expirySlotSet > lastKnownEoS

certs :: !VssCertificatesMap

Not expired certificates

whenInsMap :: !(HashMap StakeholderId EpochOrSlot)

Slot when certs was inserted. It is needed for deletion from insSlotSet (by StakeholderId).

whenInsSet :: !(Set (EpochOrSlot, StakeholderId))

Set of pairs (insertion slot, address hash) Every element of insSlotSet <= lastKnownEoS

whenExpire :: !(Set (EpochOrSlot, StakeholderId))

Set of pairs (expiry slot, address hash). Expiry slot is first slot when certificate expires. Pairs are sorted by expiry slot (in increasing order, so the oldest certificate is first element).

expiredCerts :: !(Set (EpochOrSlot, (StakeholderId, EpochOrSlot, VssCertificate)))

Set of expired certs for current lastKnownEoS. We store only certificates which expried no earlier than in previous epoch. Set (full expired slot, (id, insertion slot, cert))

empty :: VssCertData

Create empty VssCertData.

insert :: VssCertificate -> VssCertData -> VssCertData

Remove old certificate corresponding to the specified StakeholderId and insert new certificate.

lookup :: StakeholderId -> VssCertData -> Maybe VssCertificate

Lookup certificate corresponding to the specified StakeholderId.

lookupExpiryEpoch :: StakeholderId -> VssCertData -> Maybe EpochIndex

Lookup expiry epoch of certificate corresponding to the specified StakeholderId.

setLastKnownSlot :: SlotId -> VssCertData -> VssCertData

setLastKnownEoS :: EpochOrSlot -> VssCertData -> VssCertData

Set last known slot (lks). 1. If new lks is bigger than lastKnownEoS then some expired certificates will be removed. 2. If new lks is less than lastKnownEoS then some inserted after nlks certificates will be removed (and whenExpire) also will be updated.

keys :: VssCertData -> [StakeholderId]

Ids of stakeholders issued certificates.

member :: StakeholderId -> VssCertData -> Bool

Return True if the specified address hash is present in the map, False otherwise.

expiryEpoch :: VssCertificate -> EpochIndex

Convert expiry epoch of certificate to FlatSlotId.

expiryEoS :: VssCertificate -> EpochOrSlot

Functions which delete certificates. Be careful

delete :: StakeholderId -> VssCertData -> VssCertData

Delete certificate corresponding to the specified StakeholderId. This function is dangerous, because after using it you can't rollback deleted certificates. Use carefully.

difference :: VssCertData -> HashMap StakeholderId a -> VssCertData

This function is dangerous, because after you using it you can't rollback deleted certificates. Use carefully.

filter :: (StakeholderId -> Bool) -> VssCertData -> VssCertData

Filtering the certificates. This function is dangerous, because after you using it you can't rollback deleted certificates. Use carefully.