Pos.Txp

Description

Txp system reexports.

Documentation

txInToPair :: TxIn -> (TxId, Word32)

Make a pair from TxIn.

txOutStake :: TxOutAux -> [(StakeholderId, Coin)]

Use this function if you need to know how a TxOut distributes stake (e.g. for the purpose of running follow-the-satoshi).

mkTxProof :: TxPayload -> TxProof

Construct TxProof which proves given TxPayload.

mkTxPayload :: [(Tx, TxWitness, TxDistribution)] -> TxPayload

Construct TxPayload from transactions and their witnesses and distributions.

data VTxGlobalContext

Global context data needed for script execution -- is same for the whole transaction. VT stands for "Verify Tx". Currently empty.

Constructors

VTxGlobalContext

data VTxLocalContext

Local context data for scripts -- differs per input.

Constructors

VTxLocalContext

Fields

vtlTxOut :: TxOutAux

Transaction output

verifyTx

Arguments

:: Monad m

=> Bool
Verify that script & address versions in tx are known

-> VTxGlobalContext

-> (TxIn -> m (Maybe VTxLocalContext))

-> TxAux

-> m (Either [Text] TxUndo)

CHECK: Verify Tx correctness using magic function which resolves input into Address and Coin. And also the following checks:

sum of inputs >= sum of outputs;

every input has a proper witness verifying that input;

script witnesses have matching script versions;

every input is a known unspent output.

Note that verifyTx doesn't attempt to verify scripts with versions higher than maximum script version we can handle. That's because we want blocks with such transactions to be accepted (to avoid hard forks). However, we won't include such transactions into blocks when we're creating a block.

verifyTxPure

Arguments

:: Bool
Verify that script & address versions in tx are known

-> VTxGlobalContext

-> (TxIn -> Maybe VTxLocalContext)

-> TxAux

-> Either [Text] TxUndo

topsortTxs :: forall a. (a -> WithHash Tx) -> [a] -> Maybe [a]

Does topological sort on things that contain transactions – e.g. can be used both for sorting [Tx] and [(Tx, TxWitness)].

(Backwards dfs from every node with reverse visiting order recording. Returns nothing on loop encountered. Return order is head-first.)

type TxId = Hash Tx

Represents transaction identifier as Hash of Tx.

Witness

data TxInWitness

A witness for a single input.

Constructors

PkWitness

Fields

twKey :: !PublicKey

twSig :: !TxSig

ScriptWitness

Fields

twValidator :: !Script

twRedeemer :: !Script

RedeemWitness

Fields

twRedeemKey :: !RedeemPublicKey

twRedeemSig :: !(RedeemSignature TxSigData)

UnknownWitnessType !Word8 !ByteString

type TxWitness = Vector TxInWitness

A witness is a proof that a transaction is allowed to spend the funds it spends (by providing signatures, redeeming scripts, etc). A separate proof is provided for each input.

newtype TxDistribution

Distribution of “fake” stake that follow-the-satoshi would use for a particular transaction. Length of stored list must be same as length of _txOutputs of corresponding transaction.

Constructors

TxDistribution

Fields

getTxDistribution :: NonEmpty [(StakeholderId, Coin)]

type TxSigData = (TxId, Word32, Hash (NonEmpty TxOut), Hash TxDistribution)

Data that is being signed when creating a TxSig.

type TxSig = Signature TxSigData

Signature of addrId.

Tx parts

data TxIn

Transaction input.

Constructors

TxIn

Fields

txInHash :: !TxId

Which transaction's output is used

txInIndex :: !Word32

Index of the output in transaction's outputs

data TxOut

Transaction output.

Constructors

TxOut

Fields

txOutAddress :: !Address

txOutValue :: !Coin

type TxOutAux = (TxOut, [(StakeholderId, Coin)])

type TxAttributes = Attributes ()

Represents transaction attributes: map from 1-byte integer to arbitrary-type value. To be used for extending transaction with new fields via softfork.

Tx

data Tx

Transaction.

NB: transaction witnesses are stored separately.

Constructors

UnsafeTx

Fields

_txInputs :: !(NonEmpty TxIn)

Inputs of transaction.

_txOutputs :: !(NonEmpty TxOut)

Outputs of transaction.

_txAttributes :: !TxAttributes

Attributes of transaction

type TxAux = (Tx, TxWitness, TxDistribution)

Transaction + auxiliary data

mkTx :: MonadFail m => NonEmpty TxIn -> NonEmpty TxOut -> TxAttributes -> m Tx

Create valid Tx or fail. Verify inputs and outputs are non empty; have enough coins.

txInputs :: Lens' Tx (NonEmpty TxIn)

txOutputs :: Lens' Tx (NonEmpty TxOut)

txAttributes :: Lens' Tx TxAttributes

txF :: Format r (Tx -> r)

Specialized formatter for Tx.

txaF :: Format r (TxAux -> r)

Specialized formatter for Tx with auxiliary data

Payload and proof

data TxProof

Constructors

TxProof

Fields

txpNumber :: !Word32

txpRoot :: !(MerkleRoot Tx)

txpWitnessesHash :: !(Hash [TxWitness])

data TxPayload

Constructors

TxPayload

Fields

_txpTxs :: !(MerkleTree Tx)

Transactions are the main payload.

_txpDistributions :: ![TxDistribution]

Distributions for P2SH addresses in transaction outputs. * length mbTxAddrDistributions == length mbTxs * i-th element is Just if at least one output of i-th transaction is P2SH * n-th element of i-th element is Just if n-th output of i-th transaction is P2SH Ask @neongreen if you don't understand wtf is going on. Basically, address distributions are needed so that (potential) receivers of P2SH funds would count as stakeholders.

_txpWitnesses :: ![TxWitness]

Transaction witnesses. Invariant: there are as many witnesses as there are transactions in the block. This is checked during deserialisation. We can't put them into the same Merkle tree with transactions, as the whole point of segwit is to separate transactions and witnesses.

TODO: should they be put into a separate Merkle tree or left as a list?

txpTxs :: Lens' TxPayload (MerkleTree Tx)

txpWitnesses :: Lens' TxPayload [TxWitness]

txpDistributions :: Lens' TxPayload [TxDistribution]

Undo

type TxUndo = [TxOutAux]

Particular undo needed for transactions

type TxpUndo = [TxUndo]

data TxpError

Constructors

TxpInternalError !Text
Something bad happened inside Txp

txVerifyBlocks :: forall ssc m. TxpWorkMode m => OldestFirst NE (Block ssc) -> m (Either Text (OldestFirst NE TxpUndo))

Verify chain of blocks and return transaction undos of blocks.

txApplyBlocks :: TxpWorkMode m => OldestFirst NE (Blund ssc) -> m SomeBatchOp

Apply chain of definitely valid blocks to state on transactions processing.

txRollbackBlocks :: (WithLogger m, MonadDB m) => NewestFirst NE (Blund ssc) -> m SomeBatchOp

Rollback chain of blocks.

txProcessTransaction :: TxpLocalWorkMode m => (TxId, TxAux) -> m ()

txNormalize :: (MonadDB m, MonadTxpMem m) => m ()

Recompute UtxoView by current MemPool | 2. Remove invalid transactions from MemPool | 3. Set new tip to txp local data

class Monad m => MonadTxpMem m where

Reduced equivalent of MonadReader TxpLocalData m.

Methods

askTxpMem :: m TxpLocalData

Retrieve TxpLocalData.

askTxpMem :: (MonadTrans t, MonadTxpMem m', t m' ~ m) => m TxpLocalData

Retrieve TxpLocalData.

getUtxoModifier :: (MonadTxpMem m, MonadIO m) => m UtxoModifier

getLocalTxsNUndo :: (MonadIO m, MonadTxpMem m) => m ([(TxId, TxAux)], HashMap TxId [TxOutAux])

getMemPool :: (MonadIO m, MonadTxpMem m) => m MemPool

getLocalTxs :: (MonadIO m, MonadTxpMem m) => m [(TxId, TxAux)]

modifyTxpLocalData :: (MonadIO m, MonadTxpMem m) => (TxpLocalDataPure -> (a, TxpLocalDataPure)) -> m a

setTxpLocalData :: (MonadIO m, MonadTxpMem m) => TxpLocalDataPure -> m ()

newtype TxpHolder m a

Constructors

TxpHolder

Fields

getTxpHolder :: ReaderT TxpLocalData m a

runTxpHolder :: MonadIO m => UtxoModifier -> HeaderHash -> TxpHolder m a -> m a

runLocalTxpHolder :: MonadIO m => TxpHolder m a -> UtxoModifier -> m a

Local run needed for validation txs. For validation need only UtxoModifier.

runTxpHolderReader :: TxpLocalData -> TxpHolder m a -> m a

data TxpLocalData

LocalData of transactions processing. There are two invariants which must hold for local data (where um is UtxoModifier, memPool is MemPool and tip is HeaderHash): 1. Suppose blks is sequence of blocks from the very beginning up to tip. If one applies blks to genesis Utxo, resulting Utxo (let's call it utxo1) will be such that all transactions from memPool are valid with respect to it. 2. If one applies all transactions from memPool to utxo1, resulting Utxo will be equivalent to um with respect to MonadUtxo.

Real data inside TxpHolder.

Constructors

TxpLocalData

Fields

txpUtxoModifier :: !(TVar UtxoModifier)

txpMemPool :: !(TVar MemPool)

txpUndos :: !(TVar UndoMap)

txpTip :: !(TVar HeaderHash)

type TxpLocalDataPure = (UtxoModifier, MemPool, UndoMap, HeaderHash)

Pure version of TxpLocalData.

txListeners :: WorkMode ssc m => m ([ListenerSpec m], OutSpecs)

txStubListeners :: WithLogger m => ([ListenerSpec m], OutSpecs)

data TxMsgTag

Constructors

TxMsgTag

data TxMsgContents

Data message. Can be used to send one transaction per message.

Constructors

TxMsgContents

Fields

dmTx :: !Tx

dmWitness :: !TxWitness

dmDistribution :: !TxDistribution

class Monad m => MonadUtxoRead m where

Methods

utxoGet :: TxIn -> m (Maybe TxOutAux)

utxoGet :: (MonadTrans t, MonadUtxoRead m', t m' ~ m) => TxIn -> m (Maybe TxOutAux)

class MonadUtxoRead m => MonadUtxo m where

Methods

utxoPut :: TxIn -> TxOutAux -> m ()

utxoPut :: (MonadTrans t, MonadUtxo m', t m' ~ m) => TxIn -> TxOutAux -> m ()

utxoDel :: TxIn -> m ()

utxoDel :: (MonadTrans t, MonadUtxo m', t m' ~ m) => TxIn -> m ()

class Monad m => MonadBalancesRead m where

Methods

getStake :: StakeholderId -> m (Maybe Coin)

getTotalStake :: m Coin

getStake :: (MonadTrans t, MonadBalancesRead m', t m' ~ m) => StakeholderId -> m (Maybe Coin)

getTotalStake :: (MonadTrans t, MonadBalancesRead m', t m' ~ m) => m Coin

class MonadBalancesRead m => MonadBalances m where

Methods

setStake :: StakeholderId -> Coin -> m ()

setTotalStake :: Coin -> m ()

setStake :: (MonadTrans t, MonadBalances m', t m' ~ m) => StakeholderId -> Coin -> m ()

setTotalStake :: (MonadTrans t, MonadBalances m', t m' ~ m) => Coin -> m ()

class Monad m => MonadTxPool m where

Methods

hasTx :: TxId -> m Bool

putTxWithUndo :: TxId -> TxAux -> TxUndo -> m ()

poolSize :: m Int

hasTx :: (MonadTrans t, MonadTxPool m', t m' ~ m) => TxId -> m Bool

putTxWithUndo :: (MonadTrans t, MonadTxPool m', t m' ~ m) => TxId -> TxAux -> TxUndo -> m ()

poolSize :: (MonadTrans t, MonadTxPool m', t m' ~ m) => m Int

newtype DBTxp m a

Constructors

DBTxp

Fields

runDBTxp :: m a

data TxpVerFailure

Result of transaction processing

Constructors

TxpKnown
Transaction is already in the storage (cache)

TxpInvalid !Text

TxpOverwhelmed
Local transaction storage is full -- can't accept more txs

TxpCantTopsort

TxpInvalidUndoLength !Int !Int

verifyTxp :: GlobalTxpMode m => [TxAux] -> m TxpUndo

Verify transactions correctness with respect to Utxo applying them one-by-one. Note: transactions must be topsorted to pass check. Warning: this function may apply some transactions and fail eventually. Use it only on temporary data.

applyTxp :: GlobalTxpMode m => [(TxAux, TxUndo)] -> m ()

Apply transactions from one block.

rollbackTxp :: GlobalTxpMode m => [(TxAux, TxUndo)] -> m ()

Rollback transactions from one block.

normalizeTxp :: LocalTxpMode m => [(TxId, TxAux)] -> m ()

Get rid of invalid transactions. All valid transactions will be added to mem pool and applied to utxo.

processTx :: LocalTxpMode m => (TxId, TxAux) -> m ()

newtype TxpT m a

Monad transformer which stores TxpModifier and implements writable MonadTxp.

WARNING

This transformer uses StateT and is intended for single-threaded usage only. Used for block application now.

Constructors

TxpT

Fields

getTxpT :: StateT TxpModifier m a

runTxpT :: TxpModifier -> TxpT m a -> m (a, TxpModifier)

runTxpTGlobal :: Functor m => BalancesView -> TxpT m a -> m (a, TxpModifier)

runTxpTLocal :: Functor m => UtxoModifier -> MemPool -> UndoMap -> TxpT m a -> m (a, TxpModifier)

execTxpTLocal :: Functor m => UtxoModifier -> MemPool -> UndoMap -> TxpT m a -> m TxpModifier

type Utxo = Map TxIn TxOutAux

Unspent transaction outputs.

Transaction inputs are identified by (transaction ID, index in list of output) pairs.

formatUtxo :: Utxo -> Builder

Format Utxo map as json.

utxoF :: Format r (Utxo -> r)

Specialized formatter for Utxo.

data MemPool

Constructors

MemPool

Fields

_mpLocalTxs :: !TxMap

_mpLocalTxsSize :: !Int

length is O(n) for HashMap so we store it explicitly.

mpLocalTxs :: Lens' MemPool TxMap

mpLocalTxsSize :: Lens' MemPool Int

type TxMap = HashMap TxId TxAux

data BalancesView

Constructors

BalancesView

Fields

_bvStakes :: !(HashMap StakeholderId Coin)

_bvTotal :: !Coin

bvStakes :: Lens' BalancesView (HashMap StakeholderId Coin)

bvTotal :: Lens' BalancesView Coin

type UndoMap = HashMap TxId TxUndo

type UtxoModifier = MapModifier TxIn TxOutAux

data TxpModifier

Constructors

TxpModifier

Fields

_txmUtxoModifier :: !UtxoModifier

_txmBalances :: !BalancesView

_txmMemPool :: !MemPool

_txmUndos :: !UndoMap

txmUtxoModifier :: Lens' TxpModifier UtxoModifier

txmBalances :: Lens' TxpModifier BalancesView

txmMemPool :: Lens' TxpModifier MemPool

txmUndos :: Lens' TxpModifier UndoMap

newtype UtxoReaderT m a

Constructors

UtxoReaderT

Fields

getUtxoReaderT :: ReaderT Utxo m a

runUtxoReaderT :: UtxoReaderT m a -> Utxo -> m a

type UtxoReader = UtxoReaderT Identity

runUtxoReader :: UtxoReader a -> Utxo -> a

newtype UtxoStateT m a

Constructors

UtxoStateT

Fields

getUtxoStateT :: StateT Utxo m a

runUtxoStateT :: UtxoStateT m a -> Utxo -> m (a, Utxo)

evalUtxoStateT :: Monad m => UtxoStateT m a -> Utxo -> m a

execUtxoStateT :: Monad m => UtxoStateT m a -> Utxo -> m Utxo

type UtxoState = UtxoStateT Identity

runUtxoState :: UtxoState a -> Utxo -> (a, Utxo)

evalUtxoState :: UtxoState a -> Utxo -> a

execUtxoState :: UtxoState a -> Utxo -> Utxo

applyTxToUtxoPure :: WithHash Tx -> TxDistribution -> Utxo -> Utxo

Pure version of applyTxToUtxo.

verifyTxUtxoPure :: Bool -> Utxo -> TxAux -> VerificationRes

Pure version of verifyTxUtxo.

verifyTxUtxo :: (MonadUtxoRead m, MonadError TxpVerFailure m) => Bool -> TxAux -> m TxUndo

Verify single Tx using MonadUtxoRead as TxIn resolver.

applyTxToUtxo :: MonadUtxo m => WithHash Tx -> TxDistribution -> m ()

Remove unspent outputs used in given transaction, add new unspent outputs.

rollbackTxUtxo :: (MonadError TxpVerFailure m, MonadUtxo m) => (TxAux, TxUndo) -> m ()

Pure

deleteTxIn :: TxIn -> Utxo -> Utxo

Delete given TxIn from Utxo if any.

findTxIn :: TxIn -> Utxo -> Maybe TxOutAux

Find transaction input in Utxo assuming it is valid.

belongsTo :: TxOutAux -> Address -> Bool

A predicate for TxOut which selects outputs for given address

filterUtxoByAddr :: Address -> Utxo -> Utxo

Select only TxOuts for given addresses

utxoToStakes :: Utxo -> HashMap StakeholderId Coin