Pos.Txp

Description

Txp system reexports.

Documentation

module Pos.Txp.Core

data TxpError

Constructors

TxpInternalError !Text
Something bad happened inside Txp

txpGlobalSettings :: TxpGlobalSettings

Settings used for global transactions data processing used by a simple full node.

Helpers

data ApplyBlocksSettings extra m

Constructors

ApplyBlocksSettings

Fields

absApplySingle :: TxpBlund -> m ()

absExtraOperations :: extra -> SomeBatchOp

applyBlocksWith :: (TxpGlobalApplyMode m, Default extra) => ApplyBlocksSettings extra (ToilT extra (DBTxp m)) -> OldestFirst NE TxpBlund -> m SomeBatchOp

blundToAuxNUndo :: TxpBlund -> [(TxAux, TxUndo)]

genericToilModifierToBatch :: (e -> SomeBatchOp) -> GenericToilModifier e -> SomeBatchOp

Convert GenericToilModifier to batch of database operations.

runToilAction :: (MonadDB m, Default e) => ToilT e (DBTxp m) a -> m (a, GenericToilModifier e)

Run action which requires toil interfaces.

txProcessTransaction :: TxpLocalWorkMode m => (TxId, TxAux) -> m ()

txNormalize :: (MonadDB m, MonadTxpMem () m) => m ()

Recompute UtxoView by current MemPool | 2. Remove invalid transactions from MemPool | 3. Set new tip to txp local data

class Monad m => MonadTxpMem extra m | m -> extra where

Reduced equivalent of MonadReader (GenericTxpLocalData mw) m.

Methods

askTxpMem :: m (GenericTxpLocalData extra)

Retrieve GenericTxpLocalData.

askTxpMem :: (MonadTrans t, MonadTxpMem extra m', t m' ~ m) => m (GenericTxpLocalData extra)

Retrieve GenericTxpLocalData.

getUtxoModifier :: (MonadTxpMem e m, MonadIO m) => m UtxoModifier

getLocalTxsNUndo :: (MonadIO m, MonadTxpMem e m) => m ([(TxId, TxAux)], HashMap TxId (NonEmpty TxOutAux))

getMemPool :: (MonadIO m, MonadTxpMem e m) => m MemPool

getLocalTxs :: (MonadIO m, MonadTxpMem e m) => m [(TxId, TxAux)]

getLocalTxsMap :: (MonadIO m, MonadTxpMem e m) => m (HashMap TxId TxAux)

getTxpExtra :: (MonadIO m, MonadTxpMem e m) => m e

modifyTxpLocalData :: (MonadIO m, MonadTxpMem ext m) => (GenericTxpLocalDataPure ext -> (a, GenericTxpLocalDataPure ext)) -> m a

setTxpLocalData :: (MonadIO m, MonadTxpMem ext m) => GenericTxpLocalDataPure ext -> m ()

newtype TxpHolder ext m a

Constructors

TxpHolder

Fields

getTxpHolder :: ReaderT (GenericTxpLocalData ext) m a

mkTxpLocalData :: (Default e, MonadIO m) => UtxoModifier -> HeaderHash -> m (GenericTxpLocalData e)

runTxpHolder :: GenericTxpLocalData ext -> TxpHolder ext m a -> m a

data GenericTxpLocalData extra

LocalData of transactions processing. There are two invariants which must hold for local data (where um is UtxoModifier, memPool is MemPool and tip is HeaderHash): 1. Suppose blks is sequence of blocks from the very beginning up to tip. If one applies blks to genesis Utxo, resulting Utxo (let's call it utxo1) will be such that all transactions from memPool are valid with respect to it. 2. If one applies all transactions from memPool to utxo1, resulting Utxo will be equivalent to um with respect to MonadUtxo.

Memory state of Txp. Generic version.

Constructors

TxpLocalData

Fields

txpUtxoModifier :: !(TVar UtxoModifier)

txpMemPool :: !(TVar MemPool)

txpUndos :: !(TVar UndoMap)

txpTip :: !(TVar HeaderHash)

txpExtra :: !(TVar extra)

type GenericTxpLocalDataPure extra = (UtxoModifier, MemPool, UndoMap, HeaderHash, extra)

Pure version of GenericTxpLocalData.

type TxpLocalData = GenericTxpLocalData ()

Memory state of Txp. This version is used by actual Txp implementation.

type TxpLocalDataPure = GenericTxpLocalDataPure ()

Pure version of TxpLocalData.

txListeners :: WorkMode ssc m => m ([ListenerSpec m], OutSpecs)

txStubListeners :: WithLogger m => ([ListenerSpec m], OutSpecs)

data TxMsgTag

Constructors

TxMsgTag

data TxMsgContents

Data message. Can be used to send one transaction per message.

Constructors

TxMsgContents

Fields

dmTx :: !Tx

dmWitness :: !TxWitness

dmDistribution :: !TxDistribution

module Pos.Txp.Settings

class Monad m => MonadUtxoRead m where

Methods

utxoGet :: TxIn -> m (Maybe TxOutAux)

utxoGet :: (MonadTrans t, MonadUtxoRead m', t m' ~ m) => TxIn -> m (Maybe TxOutAux)

class MonadUtxoRead m => MonadUtxo m where

Methods

utxoPut :: TxIn -> TxOutAux -> m ()

utxoPut :: (MonadTrans t, MonadUtxo m', t m' ~ m) => TxIn -> TxOutAux -> m ()

utxoDel :: TxIn -> m ()

utxoDel :: (MonadTrans t, MonadUtxo m', t m' ~ m) => TxIn -> m ()

class Monad m => MonadBalancesRead m where

Methods

getStake :: StakeholderId -> m (Maybe Coin)

getTotalStake :: m Coin

getStake :: (MonadTrans t, MonadBalancesRead m', t m' ~ m) => StakeholderId -> m (Maybe Coin)

getTotalStake :: (MonadTrans t, MonadBalancesRead m', t m' ~ m) => m Coin

class MonadBalancesRead m => MonadBalances m where

Methods

setStake :: StakeholderId -> Coin -> m ()

setTotalStake :: Coin -> m ()

setStake :: (MonadTrans t, MonadBalances m', t m' ~ m) => StakeholderId -> Coin -> m ()

setTotalStake :: (MonadTrans t, MonadBalances m', t m' ~ m) => Coin -> m ()

class Monad m => MonadToilEnv m where

Type class which lets get some environmental data needed for transactions processing.

Methods

getToilEnv :: m ToilEnv

getToilEnv :: (MonadTrans t, MonadToilEnv m', t m' ~ m) => m ToilEnv

class Monad m => MonadTxPool m where

Methods

hasTx :: TxId -> m Bool

poolSize :: m Int

putTxWithUndo :: TxId -> TxAux -> TxUndo -> m ()

hasTx :: (MonadTrans t, MonadTxPool m', t m' ~ m) => TxId -> m Bool

poolSize :: (MonadTrans t, MonadTxPool m', t m' ~ m) => m Int

putTxWithUndo :: (MonadTrans t, MonadTxPool m', t m' ~ m) => TxId -> TxAux -> TxUndo -> m ()

newtype DBTxp m a

Constructors

DBTxp

Fields

runDBTxp :: m a

data ToilVerFailure

Result of transaction processing

Constructors

ToilKnown
Transaction is already in the storage (cache)

ToilTipsMismatch

Fields

ttmOldTip :: !HeaderHash

ttmNewTip :: !HeaderHash

ToilOverwhelmed
Local transaction storage is full -- can't accept more txs

ToilNotUnspent !TxIn
Tx input is not a known unspent input.

ToilOutGTIn

Fields

tInputSum :: !Integer

tOutputSum :: !Integer

ToilInconsistentTxAux !Text

ToilInvalidOutputs !Text

ToilInvalidInputs ![Text]

ToilTooLargeTx

Fields

ttltSize :: !Byte

ttltLimit :: !Byte

type GlobalToilMode m = (MonadUtxo m, MonadBalances m, MonadToilEnv m, WithLogger m)

verifyToil :: (GlobalToilMode m, MonadError ToilVerFailure m) => [TxAux] -> m TxpUndo

Verify transactions correctness with respect to Utxo applying them one-by-one. Note: transactions must be topsorted to pass check. Warning: this function may apply some transactions and fail eventually. Use it only on temporary data.

applyToil :: GlobalToilMode m => [(TxAux, TxUndo)] -> m ()

Apply transactions from one block. They must be valid (for example, it implies topological sort).

rollbackToil :: GlobalToilMode m => [(TxAux, TxUndo)] -> m ()

Rollback transactions from one block.

type LocalToilMode m = (MonadUtxo m, MonadToilEnv m, MonadTxPool m)

normalizeToil :: LocalToilMode m => [(TxId, TxAux)] -> m ()

Get rid of invalid transactions. All valid transactions will be added to mem pool and applied to utxo.

processTx :: (LocalToilMode m, MonadError ToilVerFailure m) => (TxId, TxAux) -> m TxUndo

Verify one transaction and also add it to mem pool and apply to utxo if transaction is valid.

verifyAndApplyTx :: (MonadUtxo m, MonadToilEnv m, MonadError ToilVerFailure m) => Bool -> (TxId, TxAux) -> m TxUndo

newtype ToilT ext m a

Monad transformer which stores ToilModifier and implements writable Toil type classes.

WARNING

This transformer uses StateT and is intended for single-threaded usage only. Used for block application now.

Constructors

ToilT

Fields

getToilT :: StateT (GenericToilModifier ext) m a

runToilT :: GenericToilModifier ext -> ToilT ext m a -> m (a, GenericToilModifier ext)

Run ToilT using specified modifier.

runToilTGlobal :: (Default ext, Functor m) => ToilT ext m a -> m (a, GenericToilModifier ext)

Run ToilT using empty modifier. Should be used for global transaction processing.

runToilTLocal :: Functor m => UtxoModifier -> MemPool -> UndoMap -> ToilT () m a -> m (a, ToilModifier)

Run ToilT using empty balances modifier. Should be used for local transaction processing.

execToilTLocal :: Functor m => UtxoModifier -> MemPool -> UndoMap -> ToilT () m a -> m ToilModifier

Execute ToilT using empty balances modifier. Should be used for local transaction processing.

runToilTLocalExtra :: Functor m => UtxoModifier -> MemPool -> UndoMap -> extra -> ToilT extra m a -> m (a, GenericToilModifier extra)

Like runToilTLocal, but takes extra data as argument.

type Utxo = Map TxIn TxOutAux

Unspent transaction outputs.

Transaction inputs are identified by (transaction ID, index in list of output) pairs.

formatUtxo :: Utxo -> Builder

Format Utxo map as json.

utxoF :: Format r (Utxo -> r)

Specialized formatter for Utxo.

data MemPool

Constructors

MemPool

Fields

_mpLocalTxs :: !TxMap

_mpLocalTxsSize :: !Int

length is O(n) for HashMap so we store it explicitly.

mpLocalTxs :: Lens' MemPool TxMap

mpLocalTxsSize :: Lens' MemPool Int

type TxMap = HashMap TxId TxAux

data BalancesView

Constructors

BalancesView

Fields

_bvStakes :: !(HashMap StakeholderId Coin)

_bvTotal :: !(Maybe Coin)

bvStakes :: Lens' BalancesView (HashMap StakeholderId Coin)

bvTotal :: Lens' BalancesView (Maybe Coin)

type UndoMap = HashMap TxId TxUndo

type UtxoModifier = MapModifier TxIn TxOutAux

data GenericToilModifier extension

Constructors

ToilModifier

Fields

_tmUtxo :: !UtxoModifier

_tmBalances :: !BalancesView

_tmMemPool :: !MemPool

_tmUndos :: !UndoMap

_tmExtra :: !extension

type ToilModifier = GenericToilModifier ()

tmUtxo :: forall extension. Lens' (GenericToilModifier extension) UtxoModifier

tmBalances :: forall extension. Lens' (GenericToilModifier extension) BalancesView

tmMemPool :: forall extension. Lens' (GenericToilModifier extension) MemPool

tmUndos :: forall extension. Lens' (GenericToilModifier extension) UndoMap

tmExtra :: forall extension extension. Lens (GenericToilModifier extension) (GenericToilModifier extension) extension extension

Env

data ToilEnv

Environment used by Toil.

Constructors

ToilEnv

Fields

teMaxTxSize :: !Byte

data VTxContext

Global context data needed for tx verification. VT stands for "Verify Tx". To be populated in further versions.

Constructors

VTxContext

Fields

vtcVerifyVersions :: !Bool

Verify that script & address versions in tx are known

verifyTxUtxo :: (MonadUtxoRead m, MonadError ToilVerFailure m) => VTxContext -> TxAux -> m TxUndo

CHECK: Verify Tx correctness using MonadUtxoRead. Specifically there are the following checks:

every input is a known unspent output;

sum of inputs >= sum of outputs;

every input has a proper witness verifying that input;

script witnesses have matching script versions;

if vtcVerifyVersions is True, addresses' versions are verified to be known.

Note that verifyTxUtxo doesn't attempt to verify scripts with versions higher than maximum script version we can handle. That's because we want blocks with such transactions to be accepted (to avoid hard forks). However, we won't include such transactions into blocks when we're creating a block.

applyTxToUtxo :: MonadUtxo m => WithHash Tx -> TxDistribution -> m ()

Remove unspent outputs used in given transaction, add new unspent outputs.

rollbackTxUtxo :: MonadUtxo m => (TxAux, TxUndo) -> m ()

Rollback application of given transaction to Utxo using Undo data. This function assumes that transaction has been really applied and doesn't check anything.

newtype UtxoReaderT m a

Constructors

UtxoReaderT

Fields

getUtxoReaderT :: ReaderT Utxo m a

runUtxoReaderT :: UtxoReaderT m a -> Utxo -> m a

type UtxoReader = UtxoReaderT Identity

runUtxoReader :: UtxoReader a -> Utxo -> a

newtype UtxoStateT m a

Constructors

UtxoStateT

Fields

getUtxoStateT :: StateT Utxo m a

runUtxoStateT :: UtxoStateT m a -> Utxo -> m (a, Utxo)

evalUtxoStateT :: Monad m => UtxoStateT m a -> Utxo -> m a

execUtxoStateT :: Monad m => UtxoStateT m a -> Utxo -> m Utxo

type UtxoState = UtxoStateT Identity

runUtxoState :: UtxoState a -> Utxo -> (a, Utxo)

evalUtxoState :: UtxoState a -> Utxo -> a

execUtxoState :: UtxoState a -> Utxo -> Utxo

applyTxToUtxoPure :: WithHash Tx -> TxDistribution -> Utxo -> Utxo

Pure version of applyTxToUtxo.

verifyTxUtxoPure :: MonadError ToilVerFailure m => VTxContext -> Utxo -> TxAux -> m TxUndo

Pure version of verifyTxUtxo.

filterUtxoByAddr :: Address -> Utxo -> Utxo

Select only TxOuts for given addresses

utxoToStakes :: Utxo -> HashMap StakeholderId Coin

Convert Utxo to map from StakeholderId to stake.