Pos.Txp

Description

Txp system reexports.

Documentation

data TxpLDWrap ssc

Real data inside TxpLDHolder

Constructors

TxpLDWrap

Fields

utxoView :: !(TVar (UtxoView ssc))

memPool :: !(TVar MemPool)

undos :: !(TVar (HashMap TxId [TxOutAux]))

ldTip :: !(TVar HeaderHash)

class Monad m => MonadTxpLD ssc m | m -> ssc where

Methods

getTxpLDWrap :: m (TxpLDWrap ssc)

setUtxoView :: UtxoView ssc -> m ()

setMemPool :: MemPool -> m ()

modifyTxpLD :: (TxpLD ssc -> (a, TxpLD ssc)) -> m a

modifyTxpLD_ :: (TxpLD ssc -> TxpLD ssc) -> m ()

getTxpLD :: m (TxpLD ssc)

setTxpLD :: TxpLD ssc -> m ()

getTxpLDWrap :: (MonadTrans t, MonadTxpLD ssc m', t m' ~ m) => m (TxpLDWrap ssc)

setUtxoView :: (MonadTrans t, MonadTxpLD ssc m', t m' ~ m) => UtxoView ssc -> m ()

setMemPool :: (MonadTrans t, MonadTxpLD ssc m', t m' ~ m) => MemPool -> m ()

modifyTxpLD :: (MonadTrans t, MonadTxpLD ssc m', t m' ~ m) => (TxpLD ssc -> (a, TxpLD ssc)) -> m a

getTxpLD :: (MonadTrans t, MonadTxpLD ssc m', t m' ~ m) => m (TxpLD ssc)

type TxpLD ssc = (UtxoView ssc, MemPool, HashMap TxId [TxOutAux], HeaderHash)

LocalData of transactions processing. There are two invariants which must hold for local data (where uv is UtxoView, memPool is MemPool and tip is HeaderHash): 1. Suppose blks is sequence of blocks from the very beggining up to tip. If one applies blks to genesis Utxo, resulting Utxo (let's call it utxo1) will be such that all transactions from memPool are valid with respect to it. 2. If one applies all transactions from memPool to utxo1, resulting Utxo will be equivalent to uv with respect to MonadUtxo.

getLocalTxs :: MonadTxpLD ssc m => m [(TxId, TxAux)]

getLocalUndo :: MonadTxpLD ssc m => m (HashMap TxId [TxOutAux])

getLocalTxsNUndo :: MonadTxpLD ssc m => m ([(TxId, TxAux)], HashMap TxId [TxOutAux])

getUtxoView :: MonadTxpLD ssc m => m (UtxoView ssc)

getMemPool :: MonadTxpLD ssc m => m MemPool

data TxpError

Constructors

TxpCantApplyBlocks Text
Can't apply blocks to state of transactions processing.

newtype TxpLDHolder ssc m a

Constructors

TxpLDHolder

Fields

getTxpLDHolder :: ReaderT (TxpLDWrap ssc) m a

runTxpLDHolder :: MonadIO m => UtxoView ssc -> HeaderHash -> TxpLDHolder ssc m a -> m a

runLocalTxpLDHolder :: MonadIO m => TxpLDHolder ssc m a -> UtxoView ssc -> m a

Local run needed for validation txs. For validation need only UtxoView.

data TxpLDWrap ssc

Real data inside TxpLDHolder

Constructors

TxpLDWrap

Fields

utxoView :: !(TVar (UtxoView ssc))

memPool :: !(TVar MemPool)

undos :: !(TVar (HashMap TxId [TxOutAux]))

ldTip :: !(TVar HeaderHash)

runTxpLDHolderReader :: TxpLDWrap ssc -> TxpLDHolder ssc m a -> m a

txVerifyBlocks :: forall ssc m. MonadDB ssc m => OldestFirst NE (Block ssc) -> m (Either Text (OldestFirst NE TxUndo))

Verify whether sequence of blocks can be applied to current Tx state. This function doesn't make pure checks for transactions, they are assumed to be done earlier.

txApplyBlocks :: TxpWorkMode ssc m => OldestFirst NE (Blund ssc) -> m (OldestFirst NE SomePrettyBatchOp)

Apply chain of definitely valid blocks to state on transactions processing.

processTx :: MinTxpWorkMode ssc m => (TxId, TxAux) -> m ProcessTxRes

txRollbackBlocks :: (WithLogger m, MonadDB ssc m) => NewestFirst NE (Blund ssc) -> m (NonEmpty SomePrettyBatchOp)

normalizeTxpLD :: (MonadDB ssc m, MonadTxpLD ssc m) => m ()

Recompute UtxoView by current MemPool | 2. Remove invalid transactions from MemPool

data TxMsgTag

Constructors

TxMsgTag

data TxMsgContents

Data message. Can be used to send one transaction per message.

Constructors

TxMsgContents

Fields

dmTx :: !Tx

dmWitness :: !TxWitness

dmDistribution :: !TxDistribution

data UtxoView ssc

Constructors

UtxoView

Fields

addUtxo :: !(HashMap TxIn TxOutAux)

delUtxo :: !(HashSet TxIn)

utxoDB :: !(DB ssc)

data MemPool

Constructors

MemPool

Fields

localTxs :: !TxMap

localTxsSize :: !Int

length is O(n) for HashMap so we store it explicitly.

type TxMap = HashMap TxId TxAux

data ProcessTxRes

Result of transaction processing

Constructors

PTRadded
Transaction has ben successfully added to the storage

PTRknown
Transaction is already in the storage (cache)

PTRinvalid !Text
Can't add transaction

PTRoverwhelmed
Local transaction storage is full -- can't accept more txs

mkPTRinvalid :: [Text] -> ProcessTxRes

Make ProcessTxRes from list of error messages using PTRinvalid constructor. Intended to be used with VerificationRes. Note: this version forces computation of all error messages. It can be made more efficient but less informative by using head, for example.

getTxOut :: (MonadIO m, MonadThrow m) => TxIn -> UtxoView ssc -> m (Maybe TxOutAux)

putTxOut :: TxIn -> TxOutAux -> UtxoView ssc -> UtxoView ssc

delTxIn :: TxIn -> UtxoView ssc -> UtxoView ssc

createFromDB :: DB ssc -> UtxoView ssc

class Monad m => MonadBalances ssc m | m -> ssc where

Minimal complete definition

setStake, setTotalStake, getStake, getTotalStake

Methods

setStake :: StakeholderId -> Coin -> m ()

setTotalStake :: Coin -> m ()

getStake :: StakeholderId -> m (Maybe Coin)

getTotalStake :: m Coin

type BalancesHolder ssc = StateT (BalancesView ssc)

data BalancesView ssc

Constructors

BalancesView

Fields

_stakes :: !(HashMap StakeholderId Coin)

_total :: !Coin

_balDB :: !(DB ssc)