Pos.Types.Address

Documentation

data Address

Address is where you can send coins.

Constructors

PubKeyAddress

Fields

addrKeyHash :: !(AddressHash PublicKey)

addrPkAttributes :: !(Attributes AddrPkAttrs)

ScriptAddress

Fields

addrScriptHash :: !(AddressHash Script)

UnknownAddressType !Word8 !ByteString

data AddrPkAttrs

Constructors

AddrPkAttrs

Fields

addrPkDerivationPath :: Maybe [Word32]

addressF :: Bi Address => Format r (Address -> r)

Specialized formatter for Address.

addressDetailedF :: Format r (Address -> r)

A formatter showing guts of an Address.

checkPubKeyAddress :: PublicKey -> Address -> Bool

Check if given Address is created from given PublicKey

checkScriptAddress :: Bi Script => Script -> Address -> Bool

Check if given Address is created from given validation script

checkUnknownAddressType :: Word8 -> Address -> Bool

Check if given Address has given type

makePubKeyAddress :: PublicKey -> Address

A function for making an address from PublicKey

makePubKeyHdwAddress

Arguments

:: PublicKey

-> [Word32]
Derivation path

-> Address

A function for making an HDW address

makeScriptAddress :: Bi Script => Script -> Address

A function for making an address from a validation script

decodeTextAddress :: Bi Address => Text -> Either Text Address

type StakeholderId = AddressHash PublicKey

Stakeholder identifier (stakeholders are identified by their public keys)

Internals

type AddressHash = AbstractHash Blake2s_224

addressHash :: Bi a => a -> AddressHash a

unsafeAddressHash :: Bi a => a -> AddressHash b

Orphan instances

Buildable AddrPkAttrs

Methods

build :: AddrPkAttrs -> Builder

Bi Address => Buildable Address

Methods

build :: Address -> Builder

Bi Address => Hashable Address

Methods

hashWithSalt :: Int -> Address -> Int

hash :: Address -> Int