Pos.Types.Arbitrary

Description

Arbitrary instances for core types for using in tests and benchmarks

Documentation

newtype BadSigsTx

Ill-formed Tx with bad signatures.

Constructors

BadSigsTx

Fields

getBadSigsTx :: [((Tx, TxDistribution), TxIn, TxOutAux, TxInWitness)]

newtype GoodTx

Well-formed transaction Tx.

Constructors

GoodTx

Fields

getGoodTx :: [((Tx, TxDistribution), TxIn, TxOutAux, TxInWitness)]

newtype SmallBadSigsTx

Constructors

SmallBadSigsTx BadSigsTx

newtype SmallHashMap

Constructors

SmallHashMap (HashMap PublicKey (HashMap PublicKey (AsBinary Share)))

newtype SmallGoodTx

Constructors

SmallGoodTx GoodTx

Orphan instances

Random LocalSlotIndex

Methods

randomR :: RandomGen g => (LocalSlotIndex, LocalSlotIndex) -> g -> (LocalSlotIndex, g)

random :: RandomGen g => g -> (LocalSlotIndex, g)

randomRs :: RandomGen g => (LocalSlotIndex, LocalSlotIndex) -> g -> [LocalSlotIndex]

randoms :: RandomGen g => g -> [LocalSlotIndex]

randomRIO :: (LocalSlotIndex, LocalSlotIndex) -> IO LocalSlotIndex

randomIO :: IO LocalSlotIndex

Random EpochIndex

Methods

randomR :: RandomGen g => (EpochIndex, EpochIndex) -> g -> (EpochIndex, g)

random :: RandomGen g => g -> (EpochIndex, g)

randomRs :: RandomGen g => (EpochIndex, EpochIndex) -> g -> [EpochIndex]

randoms :: RandomGen g => g -> [EpochIndex]

randomRIO :: (EpochIndex, EpochIndex) -> IO EpochIndex

randomIO :: IO EpochIndex

Arbitrary Millisecond

Methods

arbitrary :: Gen Millisecond

shrink :: Millisecond -> [Millisecond]

Arbitrary Microsecond

Methods

arbitrary :: Gen Microsecond

shrink :: Microsecond -> [Microsecond]

Arbitrary Script

Methods

arbitrary :: Gen Script

shrink :: Script -> [Script]

Arbitrary SoftwareVersion

Methods

arbitrary :: Gen SoftwareVersion

shrink :: SoftwareVersion -> [SoftwareVersion]

Arbitrary ApplicationName

Methods

arbitrary :: Gen ApplicationName

shrink :: ApplicationName -> [ApplicationName]

Arbitrary BlockVersion

Methods

arbitrary :: Gen BlockVersion

shrink :: BlockVersion -> [BlockVersion]

Arbitrary ChainDifficulty

Methods

arbitrary :: Gen ChainDifficulty

shrink :: ChainDifficulty -> [ChainDifficulty]

Arbitrary Address

Methods

arbitrary :: Gen Address

shrink :: Address -> [Address]

Arbitrary Timestamp

Methods

arbitrary :: Gen Timestamp

shrink :: Timestamp -> [Timestamp]

Arbitrary EpochOrSlot

Methods

arbitrary :: Gen EpochOrSlot

shrink :: EpochOrSlot -> [EpochOrSlot]

Arbitrary SlotId

Methods

arbitrary :: Gen SlotId

shrink :: SlotId -> [SlotId]

Arbitrary LocalSlotIndex

Methods

arbitrary :: Gen LocalSlotIndex

shrink :: LocalSlotIndex -> [LocalSlotIndex]

Arbitrary EpochIndex

Methods

arbitrary :: Gen EpochIndex

shrink :: EpochIndex -> [EpochIndex]

Arbitrary CoinPortion

Methods

arbitrary :: Gen CoinPortion

shrink :: CoinPortion -> [CoinPortion]

Arbitrary Coin

Methods

arbitrary :: Gen Coin

shrink :: Coin -> [Coin]

Arbitrary Tx
Arbitrary transactions generated from this instance will only be valid with regards to verifyTxAlone

Methods

arbitrary :: Gen Tx

shrink :: Tx -> [Tx]

Arbitrary TxOut

Methods

arbitrary :: Gen TxOut

shrink :: TxOut -> [TxOut]

Arbitrary TxIn

Methods

arbitrary :: Gen TxIn

shrink :: TxIn -> [TxIn]

Arbitrary TxDistribution

Methods

arbitrary :: Gen TxDistribution

shrink :: TxDistribution -> [TxDistribution]

Arbitrary TxInWitness

Methods

arbitrary :: Gen TxInWitness

shrink :: TxInWitness -> [TxInWitness]

Arbitrary SharedSeed

Methods

arbitrary :: Gen SharedSeed

shrink :: SharedSeed -> [SharedSeed]

Arbitrary a_6989586621679163413 => Arbitrary (FixedSizeInt a_6989586621679163413)

Methods

arbitrary :: Gen (FixedSizeInt a_6989586621679163413)

shrink :: FixedSizeInt a_6989586621679163413 -> [FixedSizeInt a_6989586621679163413]

Arbitrary a_6989586621679163865 => Arbitrary (SignedVarInt a_6989586621679163865)

Methods

arbitrary :: Gen (SignedVarInt a_6989586621679163865)

shrink :: SignedVarInt a_6989586621679163865 -> [SignedVarInt a_6989586621679163865]

Arbitrary a_6989586621679163866 => Arbitrary (UnsignedVarInt a_6989586621679163866)

Methods

arbitrary :: Gen (UnsignedVarInt a_6989586621679163866)

shrink :: UnsignedVarInt a_6989586621679163866 -> [UnsignedVarInt a_6989586621679163866]