Pos.Types.Block.Instances

Documentation

data family BodyProof p :: *

Proof of data stored in the body. Ensures immutability.

data family ConsensusData p :: *

Consensus data which can be used to check consensus properties.

data family Body p :: *

Body contains payload and other heavy data.

MainBlock and GenesisBlock lenses

blockLeaderKey :: Lens' (MainBlock ssc) PublicKey

Lens from MainBlock to PublicKey.

blockLeaders :: Lens' (GenesisBlock ssc) SlotLeaders

Lens from GenesisBlock to SlotLeaders.

blockMpc :: Lens' (MainBlock ssc) (SscPayload ssc)

Lens from MainBlock to SscPayload.

blockSignature :: Lens' (MainBlock ssc) (BlockSignature ssc)

Lens from MainBlock to Signature.

blockSlot :: Lens' (MainBlock ssc) SlotId

Lens from MainBlock to SlotId.

blockTxs :: Lens' (MainBlock ssc) (MerkleTree Tx)

Lens from MainBlock to MerkleTree.

blockTxas :: Getter (MainBlock ssc) [TxAux]

Getter from MainBlock to a list of transactions together with auxiliary data.

blockProxySKs :: Lens' (MainBlock ssc) [ProxySKHeavy]

Lens from MainBlock to ProxySKHeavy list.

gcdDifficulty :: Lens' (ConsensusData (GenesisBlockchain ssc)) ChainDifficulty

Lens for ChainDifficulty of GenesisBlockchain in ConsensusData.

gcdEpoch :: Lens' (ConsensusData (GenesisBlockchain ssc)) EpochIndex

Lens for EpochIndex of GenesisBlockchain in ConsensusData.

getBlockHeader :: Block ssc -> BlockHeader ssc

Take BlockHeader from either GenesisBlock or MainBlock.

headerLeaderKey :: Lens' (MainBlockHeader ssc) PublicKey

Lens from MainBlockHeader to PublicKey.

headerSignature :: Lens' (MainBlockHeader ssc) (BlockSignature ssc)

Lens from MainBlockHeader to Signature.

headerSlot :: Lens' (MainBlockHeader ssc) SlotId

Lens from MainBlockHeader to SlotId.

mbMpc :: Lens' (Body (MainBlockchain ssc)) (SscPayload ssc)

Lens for SscPayload in main block body.

mbTxs :: Lens' (Body (MainBlockchain ssc)) (MerkleTree Tx)

Lens for transaction tree in main block body.

mbWitnesses :: Lens' (Body (MainBlockchain ssc)) [TxWitness]

Lens for witness list in main block body.

mbProxySKs :: Lens' (Body (MainBlockchain ssc)) [ProxySKHeavy]

Lens for ProxySKs in main block body.

mbUpdatePayload :: Lens' (Body (MainBlockchain ssc)) UpdatePayload

Lens for UpdatePayload in main block body.

mcdSlot :: Lens' (ConsensusData (MainBlockchain ssc)) SlotId

Lens for SlotId of MainBlockchain in ConsensusData.

mcdLeaderKey :: Lens' (ConsensusData (MainBlockchain ssc)) PublicKey

Lens for PublicKey of MainBlockchain in ConsensusData.

mcdDifficulty :: Lens' (ConsensusData (MainBlockchain ssc)) ChainDifficulty

Lens for ChainDifficulty of MainBlockchain in ConsensusData.

mcdSignature :: Lens' (ConsensusData (MainBlockchain ssc)) (BlockSignature ssc)

Lens for Signature of MainBlockchain in ConsensusData.

HeaderHash related functions

blockHeader :: Getter (Block ssc) (BlockHeader ssc)

blockHeaderHash :: BiHeader ssc => BlockHeader ssc -> HeaderHash

This function is required because type inference fails in attempts to hash only Right or Left.

Orphan instances

HasEpochIndex SlotId

Methods

epochIndexL :: Lens' SlotId EpochIndex

HasHeaderHash HeaderHash

Methods

headerHash :: HeaderHash -> HeaderHash

headerHashG :: Getter HeaderHash HeaderHash

Eq (Body (GenesisBlockchain ssc))

Methods

(==) :: Body (GenesisBlockchain ssc) -> Body (GenesisBlockchain ssc) -> Bool

(/=) :: Body (GenesisBlockchain ssc) -> Body (GenesisBlockchain ssc) -> Bool

(Eq (SscPayload ssc), Ssc ssc) => Eq (Body (MainBlockchain ssc))

Methods

(==) :: Body (MainBlockchain ssc) -> Body (MainBlockchain ssc) -> Bool

(/=) :: Body (MainBlockchain ssc) -> Body (MainBlockchain ssc) -> Bool

Eq (ConsensusData (GenesisBlockchain ssc))

Methods

(==) :: ConsensusData (GenesisBlockchain ssc) -> ConsensusData (GenesisBlockchain ssc) -> Bool

(/=) :: ConsensusData (GenesisBlockchain ssc) -> ConsensusData (GenesisBlockchain ssc) -> Bool

Eq (ConsensusData (MainBlockchain ssc))

Methods

(==) :: ConsensusData (MainBlockchain ssc) -> ConsensusData (MainBlockchain ssc) -> Bool

(/=) :: ConsensusData (MainBlockchain ssc) -> ConsensusData (MainBlockchain ssc) -> Bool

Eq (BodyProof (GenesisBlockchain ssc))

Methods

(==) :: BodyProof (GenesisBlockchain ssc) -> BodyProof (GenesisBlockchain ssc) -> Bool

(/=) :: BodyProof (GenesisBlockchain ssc) -> BodyProof (GenesisBlockchain ssc) -> Bool

Ssc ssc => Eq (BodyProof (MainBlockchain ssc))

Methods

(==) :: BodyProof (MainBlockchain ssc) -> BodyProof (MainBlockchain ssc) -> Bool

(/=) :: BodyProof (MainBlockchain ssc) -> BodyProof (MainBlockchain ssc) -> Bool

Show (Body (GenesisBlockchain ssc))

Methods

showsPrec :: Int -> Body (GenesisBlockchain ssc) -> ShowS

show :: Body (GenesisBlockchain ssc) -> String

showList :: [Body (GenesisBlockchain ssc)] -> ShowS

Ssc ssc => Show (Body (MainBlockchain ssc))

Methods

showsPrec :: Int -> Body (MainBlockchain ssc) -> ShowS

show :: Body (MainBlockchain ssc) -> String

showList :: [Body (MainBlockchain ssc)] -> ShowS

Show (ConsensusData (GenesisBlockchain ssc))

Methods

showsPrec :: Int -> ConsensusData (GenesisBlockchain ssc) -> ShowS

show :: ConsensusData (GenesisBlockchain ssc) -> String

showList :: [ConsensusData (GenesisBlockchain ssc)] -> ShowS

Show (ConsensusData (MainBlockchain ssc))

Methods

showsPrec :: Int -> ConsensusData (MainBlockchain ssc) -> ShowS

show :: ConsensusData (MainBlockchain ssc) -> String

showList :: [ConsensusData (MainBlockchain ssc)] -> ShowS

Show (BodyProof (GenesisBlockchain ssc))

Methods

showsPrec :: Int -> BodyProof (GenesisBlockchain ssc) -> ShowS

show :: BodyProof (GenesisBlockchain ssc) -> String

showList :: [BodyProof (GenesisBlockchain ssc)] -> ShowS

Ssc ssc => Show (BodyProof (MainBlockchain ssc))

Methods

showsPrec :: Int -> BodyProof (MainBlockchain ssc) -> ShowS

show :: BodyProof (MainBlockchain ssc) -> String

showList :: [BodyProof (MainBlockchain ssc)] -> ShowS

Generic (Body (GenesisBlockchain ssc))

Associated Types

type Rep (Body (GenesisBlockchain ssc)) :: * -> *

Methods

from :: Body (GenesisBlockchain ssc) -> Rep (Body (GenesisBlockchain ssc)) x

to :: Rep (Body (GenesisBlockchain ssc)) x -> Body (GenesisBlockchain ssc)

Generic (Body (MainBlockchain ssc))

Associated Types

type Rep (Body (MainBlockchain ssc)) :: * -> *

Methods

from :: Body (MainBlockchain ssc) -> Rep (Body (MainBlockchain ssc)) x

to :: Rep (Body (MainBlockchain ssc)) x -> Body (MainBlockchain ssc)

Generic (ConsensusData (GenesisBlockchain ssc))

Associated Types

type Rep (ConsensusData (GenesisBlockchain ssc)) :: * -> *

Methods

from :: ConsensusData (GenesisBlockchain ssc) -> Rep (ConsensusData (GenesisBlockchain ssc)) x

to :: Rep (ConsensusData (GenesisBlockchain ssc)) x -> ConsensusData (GenesisBlockchain ssc)

Generic (ConsensusData (MainBlockchain ssc))

Associated Types

type Rep (ConsensusData (MainBlockchain ssc)) :: * -> *

Methods

from :: ConsensusData (MainBlockchain ssc) -> Rep (ConsensusData (MainBlockchain ssc)) x

to :: Rep (ConsensusData (MainBlockchain ssc)) x -> ConsensusData (MainBlockchain ssc)

Generic (BodyProof (GenesisBlockchain ssc))

Associated Types

type Rep (BodyProof (GenesisBlockchain ssc)) :: * -> *

Methods

from :: BodyProof (GenesisBlockchain ssc) -> Rep (BodyProof (GenesisBlockchain ssc)) x

to :: Rep (BodyProof (GenesisBlockchain ssc)) x -> BodyProof (GenesisBlockchain ssc)

Generic (BodyProof (MainBlockchain ssc))

Associated Types

type Rep (BodyProof (MainBlockchain ssc)) :: * -> *

Methods

from :: BodyProof (MainBlockchain ssc) -> Rep (BodyProof (MainBlockchain ssc)) x

to :: Rep (BodyProof (MainBlockchain ssc)) x -> BodyProof (MainBlockchain ssc)

(~) * (BHeaderHash b) HeaderHash => HasPrevBlock (GenericBlockHeader b)

Methods

prevBlockL :: Lens' (GenericBlockHeader b) HeaderHash

(~) * (BHeaderHash b) HeaderHash => HasPrevBlock (GenericBlock b)

Methods

prevBlockL :: Lens' (GenericBlock b) HeaderHash

Blockchain (GenesisBlockchain ssc)

Associated Types

data BodyProof (GenesisBlockchain ssc) :: *

data ConsensusData (GenesisBlockchain ssc) :: *

type ExtraHeaderData (GenesisBlockchain ssc) :: *

type BBlockHeader (GenesisBlockchain ssc) :: *

type BHeaderHash (GenesisBlockchain ssc) :: *

data Body (GenesisBlockchain ssc) :: *

type ExtraBodyData (GenesisBlockchain ssc) :: *

type BBlock (GenesisBlockchain ssc) :: *

Methods

mkBodyProof :: Body (GenesisBlockchain ssc) -> BodyProof (GenesisBlockchain ssc)

checkBodyProof :: Body (GenesisBlockchain ssc) -> BodyProof (GenesisBlockchain ssc) -> Bool

verifyBBlock :: GenericBlock (GenesisBlockchain ssc) -> Either Text ()

(SscHelpersClass ssc, Bi TxWitness, Bi UpdatePayload, Bi EpochIndex) => Blockchain (MainBlockchain ssc)

Associated Types

data BodyProof (MainBlockchain ssc) :: *

data ConsensusData (MainBlockchain ssc) :: *

type ExtraHeaderData (MainBlockchain ssc) :: *

type BBlockHeader (MainBlockchain ssc) :: *

type BHeaderHash (MainBlockchain ssc) :: *

data Body (MainBlockchain ssc) :: *

type ExtraBodyData (MainBlockchain ssc) :: *

type BBlock (MainBlockchain ssc) :: *

Methods

mkBodyProof :: Body (MainBlockchain ssc) -> BodyProof (MainBlockchain ssc)

checkBodyProof :: Body (MainBlockchain ssc) -> BodyProof (MainBlockchain ssc) -> Bool

verifyBBlock :: GenericBlock (MainBlockchain ssc) -> Either Text ()

HasEpochIndex (GenesisBlock ssc)

Methods

epochIndexL :: Lens' (GenesisBlock ssc) EpochIndex

HasEpochIndex (GenesisBlockHeader ssc)

Methods

epochIndexL :: Lens' (GenesisBlockHeader ssc) EpochIndex

HasEpochIndex (MainBlock ssc)

Methods

epochIndexL :: Lens' (MainBlock ssc) EpochIndex

HasEpochIndex (MainBlockHeader ssc)

Methods

epochIndexL :: Lens' (MainBlockHeader ssc) EpochIndex

HasEpochOrSlot (GenesisBlock ssc)

Methods

_getEpochOrSlot :: GenesisBlock ssc -> Either EpochIndex SlotId

getEpochOrSlot :: GenesisBlock ssc -> EpochOrSlot

epochOrSlotG :: Getter (GenesisBlock ssc) EpochOrSlot

HasEpochOrSlot (GenesisBlockHeader ssc)

Methods

_getEpochOrSlot :: GenesisBlockHeader ssc -> Either EpochIndex SlotId

getEpochOrSlot :: GenesisBlockHeader ssc -> EpochOrSlot

epochOrSlotG :: Getter (GenesisBlockHeader ssc) EpochOrSlot

HasEpochOrSlot (MainBlock ssc)

Methods

_getEpochOrSlot :: MainBlock ssc -> Either EpochIndex SlotId

getEpochOrSlot :: MainBlock ssc -> EpochOrSlot

epochOrSlotG :: Getter (MainBlock ssc) EpochOrSlot

HasEpochOrSlot (MainBlockHeader ssc)

Methods

_getEpochOrSlot :: MainBlockHeader ssc -> Either EpochIndex SlotId

getEpochOrSlot :: MainBlockHeader ssc -> EpochOrSlot

epochOrSlotG :: Getter (MainBlockHeader ssc) EpochOrSlot

HasDifficulty (ConsensusData (GenesisBlockchain ssc))

Methods

difficultyL :: Lens' (ConsensusData (GenesisBlockchain ssc)) ChainDifficulty

HasDifficulty (ConsensusData (MainBlockchain ssc))

Methods

difficultyL :: Lens' (ConsensusData (MainBlockchain ssc)) ChainDifficulty

HasDifficulty (Block ssc)

Methods

difficultyL :: Lens' (Block ssc) ChainDifficulty

HasDifficulty (BlockHeader ssc)

Methods

difficultyL :: Lens' (BlockHeader ssc) ChainDifficulty

HasDifficulty (GenesisBlock ssc)

Methods

difficultyL :: Lens' (GenesisBlock ssc) ChainDifficulty

HasDifficulty (GenesisBlockHeader ssc)

Methods

difficultyL :: Lens' (GenesisBlockHeader ssc) ChainDifficulty

HasDifficulty (MainBlock ssc)

Methods

difficultyL :: Lens' (MainBlock ssc) ChainDifficulty

HasDifficulty (MainBlockHeader ssc)

Methods

difficultyL :: Lens' (MainBlockHeader ssc) ChainDifficulty

BiHeader ssc => HasHeaderHash (Block ssc)

Methods

headerHash :: Block ssc -> HeaderHash

headerHashG :: Getter (Block ssc) HeaderHash

BiHeader ssc => HasHeaderHash (BlockHeader ssc)

Methods

headerHash :: BlockHeader ssc -> HeaderHash

headerHashG :: Getter (BlockHeader ssc) HeaderHash

BiHeader ssc => HasHeaderHash (GenesisBlock ssc)

Methods

headerHash :: GenesisBlock ssc -> HeaderHash

headerHashG :: Getter (GenesisBlock ssc) HeaderHash

BiHeader ssc => HasHeaderHash (GenesisBlockHeader ssc)

Methods

headerHash :: GenesisBlockHeader ssc -> HeaderHash

headerHashG :: Getter (GenesisBlockHeader ssc) HeaderHash

BiHeader ssc => HasHeaderHash (MainBlock ssc)

Methods

headerHash :: MainBlock ssc -> HeaderHash

headerHashG :: Getter (MainBlock ssc) HeaderHash

BiHeader ssc => HasHeaderHash (MainBlockHeader ssc)

Methods

headerHash :: MainBlockHeader ssc -> HeaderHash

headerHashG :: Getter (MainBlockHeader ssc) HeaderHash

Ssc ssc => NFData (Body (GenesisBlockchain ssc))

Methods

rnf :: Body (GenesisBlockchain ssc) -> ()

Ssc ssc => NFData (Body (MainBlockchain ssc))

Methods

rnf :: Body (MainBlockchain ssc) -> ()

Ssc ssc => NFData (ConsensusData (GenesisBlockchain ssc))

Methods

rnf :: ConsensusData (GenesisBlockchain ssc) -> ()

Ssc ssc => NFData (ConsensusData (MainBlockchain ssc))

Methods

rnf :: ConsensusData (MainBlockchain ssc) -> ()

Ssc ssc => NFData (BodyProof (GenesisBlockchain ssc))

Methods

rnf :: BodyProof (GenesisBlockchain ssc) -> ()

Ssc ssc => NFData (BodyProof (MainBlockchain ssc))

Methods

rnf :: BodyProof (MainBlockchain ssc) -> ()

Ssc ssc => NFData (GenesisBlock ssc)

Methods

rnf :: GenesisBlock ssc -> ()

Ssc ssc => NFData (MainBlock ssc)

Methods

rnf :: MainBlock ssc -> ()

BiSsc ssc => Buildable (BlockHeader ssc)

Methods

build :: BlockHeader ssc -> Builder

BiSsc ssc => Buildable (GenesisBlock ssc)

Methods

build :: GenesisBlock ssc -> Builder

BiSsc ssc => Buildable (GenesisBlockHeader ssc)

Methods

build :: GenesisBlockHeader ssc -> Builder

(Bi UpdateProposal, BiSsc ssc) => Buildable (MainBlock ssc)

Methods

build :: MainBlock ssc -> Builder

BiSsc ssc => Buildable (MainBlockHeader ssc)

Methods

build :: MainBlockHeader ssc -> Builder

(HasPrevBlock s, HasPrevBlock s') => HasPrevBlock (Either s s')

Methods

prevBlockL :: Lens' (Either s s') HeaderHash

HasPrevBlock s => HasPrevBlock (s, z)
Class for something that has previous block (lens to Hash for this block).

Methods

prevBlockL :: Lens' (s, z) HeaderHash

(HasEpochIndex a, HasEpochIndex b) => HasEpochIndex (Either a b)

Methods

epochIndexL :: Lens' (Either a b) EpochIndex

(HasEpochOrSlot a, HasEpochOrSlot b) => HasEpochOrSlot (Either a b)

Methods

_getEpochOrSlot :: Either a b -> Either EpochIndex SlotId

getEpochOrSlot :: Either a b -> EpochOrSlot

epochOrSlotG :: Getter (Either a b) EpochOrSlot