Pos.Types.Types

Description

Definitions of the most fundamental types.

Documentation

type TxAttributes = Attributes ()

Represents transaction attributes: map from 1-byte integer to arbitrary-type value. To be used for extending transaction with new fields via softfork.

data TxInWitness

A witness for a single input.

Constructors

PkWitness

Fields

twKey :: !PublicKey

twSig :: !TxSig

ScriptWitness

Fields

twValidator :: !Script

twRedeemer :: !Script

UnknownWitnessType !Word8 !ByteString

type TxWitness = Vector TxInWitness

A witness is a proof that a transaction is allowed to spend the funds it spends (by providing signatures, redeeming scripts, etc). A separate proof is provided for each input.

newtype TxDistribution

Distribution of “fake” stake that follow-the-satoshi would use for a particular transaction.

Constructors

TxDistribution

Fields

getTxDistribution :: [[(StakeholderId, Coin)]]

type TxSigData = (TxId, Word32, Hash [TxOut], Hash TxDistribution)

Data that is being signed when creating a TxSig.

type TxSig = Signature TxSigData

Signature of addrId.

type TxId = Hash Tx

Represents transaction identifier as Hash of Tx.

data TxIn

Transaction input.

Constructors

TxIn

Fields

txInHash :: !TxId

Which transaction's output is used

txInIndex :: !Word32

Index of the output in transaction's outputs

toPair :: TxIn -> (TxId, Word32)

Make pair from TxIn

data TxOut

Transaction output.

Constructors

TxOut

Fields

txOutAddress :: !Address

txOutValue :: !Coin

txOutStake :: TxOutAux -> [(StakeholderId, Coin)]

Use this function if you need to know how a TxOut distributes stake (e.g. for the purpose of running follow-the-satoshi).

data Tx

Transaction.

NB: transaction witnesses are stored separately.

Constructors

Tx

Fields

txInputs :: ![TxIn]

Inputs of transaction.

txOutputs :: ![TxOut]

Outputs of transaction.

txAttributes :: !TxAttributes

Attributes of transaction

_txInputs :: Lens' Tx [TxIn]

_txOutputs :: Lens' Tx [TxOut]

_txAttributes :: Lens' Tx TxAttributes

txF :: Format r (Tx -> r)

Specialized formatter for Tx.

txaF :: Bi Script => Format r (TxAux -> r)

Specialized formatter for Tx with auxiliary data

type TxAux = (Tx, TxWitness, TxDistribution)

Transaction + auxiliary data

type TxOutAux = (TxOut, [(StakeholderId, Coin)])

type Utxo = Map (TxId, Word32) TxOutAux

Unspent transaction outputs.

Transaction inputs are identified by (transaction ID, index in list of output) pairs.

formatUtxo :: Utxo -> Builder

Format Utxo map as json.

utxoF :: Format r (Utxo -> r)

Specialized formatter for Utxo.

type TxUndo = [[TxOutAux]]

Particular undo needed for transactions

newtype SharedSeed

This is a shared seed used for follow-the-satoshi. This seed is randomly generated by each party and eventually they agree on the same value.

Constructors

SharedSeed

Fields

getSharedSeed :: ByteString

type SlotLeaders = NonEmpty StakeholderId

NonEmpty list of slot leaders.

type ProxySigLight a = ProxySignature (EpochIndex, EpochIndex) a

Proxy signature used in csl -- holds a pair of epoch indices. Block is valid if it's epoch index is inside this range.

type ProxySKLight = ProxySecretKey (EpochIndex, EpochIndex)

Same alias for the proxy secret key (see ProxySigLight).

type ProxySigHeavy a = ProxySignature EpochIndex a

Simple proxy signature without ttl/epoch index constraints. EpochIndex inside is needed for replay attack prevention.

type ProxySKHeavy = ProxySecretKey EpochIndex

Correspondent SK for no-ttl proxy signature scheme.

type ProxySKEither = Either ProxySKLight ProxySKHeavy

Some proxy secret key.