Pos.Update

Description

Whole Update System in one module :)

Documentation

module Pos.Update.Core

downloadUpdate :: WorkMode ssc m => ConfirmedProposalState -> m ()

Download and save archive update by given ConfirmedProposalState

downloadHash :: [Text] -> Hash LByteString -> IO (Either Text LByteString)

Download a file by its hash.

Tries all servers in turn, fails if none of them work.

usApplyBlocks :: (MonadThrow m, USGlobalApplyMode ssc m) => OldestFirst NE (Block ssc) -> Maybe PollModifier -> m [SomeBatchOp]

Apply chain of definitely valid blocks to US part of GState DB and to US local data. This function assumes that no other thread applies block in parallel. It also assumes that parent of oldest block is current tip. If verification is done prior to application, one can pass PollModifier obtained from verification to this function.

usCanCreateBlock :: (WithLogger m, WithNodeContext ssc m, MonadDB ssc m) => m Bool

Checks whether our software can create block according to current global state.

usRollbackBlocks :: forall ssc m. USGlobalApplyMode ssc m => NewestFirst NE (Block ssc, USUndo) -> m [SomeBatchOp]

Revert application of given blocks to US part of GState DB and US local data. The caller must ensure that the tip stored in DB is headerHash of head.

usVerifyBlocks :: USGlobalVerifyMode ssc m => OldestFirst NE (Block ssc) -> m (PollModifier, OldestFirst NE USUndo)

Verify whether sequence of blocks can be applied to US part of current GState DB. This function doesn't make pure checks, they are assumed to be done earlier, most likely during objects construction.

Proposals

isProposalNeeded :: (MonadIO m, MonadUSMem m) => UpId -> m Bool

This function returns true if update proposal with given identifier should be requested.

getLocalProposalNVotes :: (MonadIO m, MonadUSMem m) => UpId -> m (Maybe (UpdateProposal, [UpdateVote]))

Get update proposal with given id if it is known.

processProposal :: USLocalLogicMode ssc m => UpdateProposal -> m (Either PollVerFailure ())

Process proposal received from network, checking it against current state (global + local) and adding to local state if it's valid with respect to it. If proposal is added to store, 'Right ()' is returned. Otherwise 'Left err' is returned and err lets caller decide whether sender could be sure that error would happen.

Votes

isVoteNeeded :: USLocalLogicMode ssc m => UpId -> PublicKey -> Bool -> m Bool

This function returns true if update vote proposal with given identifier issued by stakeholder with given PublicKey and with given decision should be requested.

getLocalVote :: (MonadIO m, MonadUSMem m) => UpId -> PublicKey -> Bool -> m (Maybe UpdateVote)

Get update vote for proposal with given id from given issuer and with given decision if it is known.

processVote :: USLocalLogicMode ssc m => UpdateVote -> m (Either PollVerFailure ())

Process vote received from network, checking it against current state (global + local) and adding to local state if it's valid with respect to it. If vote is added to store, 'Right ()' is returned. Otherwise 'Left err' is returned and err lets caller decide whether sender could be sure that error would happen.

Normalization

usNormalize :: USLocalLogicMode ssc m => m ()

Remove local data from memory state to make it consistent with current GState. This function assumes that GState is locked. It tries to leave as much data as possible. It assumes that blkSemaphore is taken.

processNewSlot :: USLocalLogicMode ssc m => SlotId -> m ()

Update memory state to make it correct for given slot.

usPreparePayload :: USLocalLogicMode ssc m => SlotId -> m (Maybe UpdatePayload)

Prepare UpdatePayload for inclusion into new block with given SlotId. This function assumes that blkSemaphore is taken and nobody can apply/rollback blocks in parallel. Sometimes payload can't be created. It can happen if we are trying to create block for slot which has already passed, for example.

usLrcConsumer :: MonadDB ssc m => LrcConsumer m

Consumer will be called on every Richmen computation.

class Monad m => MonadUSMem m where

Reduced equivalent of MonadReader MemVar m.

Methods

askUSMemVar :: m MemVar

Retrieve MemVar.

askUSMemVar :: (MonadTrans t, MonadUSMem m', t m' ~ m) => m MemVar

Retrieve MemVar.

askUSMemState :: MonadUSMem m => m (TVar MemState)

withUSLock :: (MonadUSMem m, MonadIO m, MonadMask m) => m a -> m a

modifyMemPool :: UpdatePayload -> PollModifier -> MemPool -> MemPool

Modify MemPool using UpdatePayload and PollModifier.

UpdatePayload is used to add new data to MemPool. Data must be verified by caller. It's added directly to MemPool.

PollModifier is used to remove or modify some data in MemPool. All deleted proposals and votes for them are removed. TODO [CSL-625] Deleted and modified votes from non-deleted proposals are not handled properly.

newtype USHolder m a

Trivial monad transformer based on ReaderT (MemVar).

Constructors

USHolder

Fields

getUSHolder :: ReaderT MemVar m a

runUSHolder :: MonadIO m => USHolder m a -> m a

Run USHolder using default (empty) MemState.

runUSHolderFromVar :: MemVar -> USHolder m a -> m a

Run USHolder using existing MemVar.

data MemPool

MemPool is data maintained by node to be included into block and relayed to other nodes.

Constructors

MemPool

Fields

mpProposals :: !UpdateProposals

mpLocalVotes :: !LocalVotes

type UpdateProposals = HashMap UpId UpdateProposal

type LocalVotes = HashMap UpId (HashMap PublicKey UpdateVote)

data MemState

MemState contains all in-memory data necesary for Update System.

Constructors

MemState

Fields

msSlot :: !SlotId

Slot for which data is valid. In reality EpochIndex should be enough, but we sometimes overgeneralize things.

msTip :: !HeaderHash

Tip for which data is valid.

msPool :: !MemPool

Pool of data to be included into block.

msModifier :: !PollModifier

Modifier of GState corresponding to msPool.

data MemVar

MemVar uses concurrency primitives and stores MemState.

Constructors

MemVar

Fields

mvState :: !(TVar MemState)

MemState itself.

mvLock :: !Lock

Lock for modifting MemState.

mkMemState :: MemState

newMemVar :: MonadIO m => m MemVar

usListeners :: WorkMode ssc m => m ([ListenerSpec m], OutSpecs)

Listeners for requests related to update system

usStubListeners :: WithLogger m => ([ListenerSpec m], OutSpecs)

data ProposalMsgTag

Tag for proposal messages

Constructors

ProposalMsgTag

data VoteMsgTag

Tag for vote messages

Constructors

VoteMsgTag

class (Monad m, WithLogger m) => MonadPollRead m where

Type class which provides function necessary for read-only verification of US data.

Methods

getBVState :: BlockVersion -> m (Maybe BlockVersionState)

Retrieve state of given block version.

getProposedBVs :: m [BlockVersion]

Retrieve all proposed block versions.

getConfirmedBVStates :: m [(BlockVersion, BlockVersionState)]

Get all confirmed BlockVersions and their states.

getAdoptedBVFull :: m (BlockVersion, BlockVersionData)

Retrieve last adopted block version and its state.

getLastConfirmedSV :: ApplicationName -> m (Maybe NumSoftwareVersion)

Get numeric component of last confirmed version of application

hasActiveProposal :: ApplicationName -> m Bool

Check if given application has an active (non-confirmed) proposal

getProposal :: UpId -> m (Maybe ProposalState)

Get active proposal

getConfirmedProposals :: m [ConfirmedProposalState]

Get all known confirmed proposals.

getEpochTotalStake :: EpochIndex -> m (Maybe Coin)

Get total stake from distribution corresponding to give epoch

getRichmanStake :: EpochIndex -> StakeholderId -> m (Maybe Coin)

Get stake of ricmhan corresponding to given epoch (if she is really rich)

getOldProposals :: SlotId -> m [UndecidedProposalState]

Get all proposals which are in undecided state and were included into block with slot less than or equal to given.

getDeepProposals :: ChainDifficulty -> m [DecidedProposalState]

Get all proposals which are in decided state and become decided deeper than given ChainDifficulty.

getBlockIssuerStake :: EpochIndex -> StakeholderId -> m (Maybe Coin)

Get stake of issuer of one of the blocks created so far using stake distribution which is stable in given epoch. Only issuer of stable block can be passed to this function, otherwise Nothing will be returned.

getSlottingData :: m SlottingData

Get most recent SlottingData.

getAdoptedBV :: m BlockVersion

getAdoptedBVData :: m BlockVersionData

getBVState :: (MonadTrans t, MonadPollRead m', t m' ~ m) => BlockVersion -> m (Maybe BlockVersionState)

Retrieve state of given block version.

getProposedBVs :: (MonadTrans t, MonadPollRead m', t m' ~ m) => m [BlockVersion]

Retrieve all proposed block versions.

getConfirmedBVStates :: (MonadTrans t, MonadPollRead m', t m' ~ m) => m [(BlockVersion, BlockVersionState)]

Get all confirmed BlockVersions and their states.

getAdoptedBVFull :: (MonadTrans t, MonadPollRead m', t m' ~ m) => m (BlockVersion, BlockVersionData)

Retrieve last adopted block version and its state.

getLastConfirmedSV :: (MonadTrans t, MonadPollRead m', t m' ~ m) => ApplicationName -> m (Maybe Word32)

Get numeric component of last confirmed version of application

hasActiveProposal :: (MonadTrans t, MonadPollRead m', t m' ~ m) => ApplicationName -> m Bool

Check if given application has an active (non-confirmed) proposal

getProposal :: (MonadTrans t, MonadPollRead m', t m' ~ m) => UpId -> m (Maybe ProposalState)

Get active proposal

getConfirmedProposals :: (MonadTrans t, MonadPollRead m', t m' ~ m) => m [ConfirmedProposalState]

Get all known confirmed proposals.

getEpochTotalStake :: (MonadTrans t, MonadPollRead m', t m' ~ m) => EpochIndex -> m (Maybe Coin)

Get total stake from distribution corresponding to give epoch

getRichmanStake :: (MonadTrans t, MonadPollRead m', t m' ~ m) => EpochIndex -> StakeholderId -> m (Maybe Coin)

Get stake of ricmhan corresponding to given epoch (if she is really rich)

getOldProposals :: (MonadTrans t, MonadPollRead m', t m' ~ m) => SlotId -> m [UndecidedProposalState]

Get all proposals which are in undecided state and were included into block with slot less than or equal to given.

getDeepProposals :: (MonadTrans t, MonadPollRead m', t m' ~ m) => ChainDifficulty -> m [DecidedProposalState]

Get all proposals which are in decided state and become decided deeper than given ChainDifficulty.

getBlockIssuerStake :: (MonadTrans t, MonadPollRead m', t m' ~ m) => EpochIndex -> StakeholderId -> m (Maybe Coin)

Get stake of issuer of one of the blocks created so far using stake distribution which is stable in given epoch. Only issuer of stable block can be passed to this function, otherwise Nothing will be returned.

getSlottingData :: (MonadTrans t, MonadPollRead m', t m' ~ m) => m SlottingData

Get most recent SlottingData.

class MonadPollRead m => MonadPoll m where

Type class which provides function necessary for verification of US data with ability to modify state.

Methods

putBVState :: BlockVersion -> BlockVersionState -> m ()

Put state of BlockVersion overriding if it exists.

delBVState :: BlockVersion -> m ()

Delete BlockVersion and associated state.

setAdoptedBV :: BlockVersion -> m ()

Set last adopted block version. State is taken from competing states.

setLastConfirmedSV :: SoftwareVersion -> m ()

Set last confirmed version of application.

delConfirmedSV :: ApplicationName -> m ()

Del last confirmed version of application.

addConfirmedProposal :: ConfirmedProposalState -> m ()

Add new confirmed update proposal.

delConfirmedProposal :: SoftwareVersion -> m ()

Del confirmed update proposal (for rollback only).

addActiveProposal :: ProposalState -> m ()

Add new active proposal with its state.

deactivateProposal :: UpId -> m ()

Delete active proposal given its name and identifier.

setSlottingData :: SlottingData -> m ()

Set most recent SlottingData.

putBVState :: (MonadTrans t, MonadPoll m', t m' ~ m) => BlockVersion -> BlockVersionState -> m ()

Put state of BlockVersion overriding if it exists.

delBVState :: (MonadTrans t, MonadPoll m', t m' ~ m) => BlockVersion -> m ()

Delete BlockVersion and associated state.

setAdoptedBV :: (MonadTrans t, MonadPoll m', t m' ~ m) => BlockVersion -> m ()

Set last adopted block version. State is taken from competing states.

setLastConfirmedSV :: (MonadTrans t, MonadPoll m', t m' ~ m) => SoftwareVersion -> m ()

Set last confirmed version of application.

delConfirmedSV :: (MonadTrans t, MonadPoll m', t m' ~ m) => ApplicationName -> m ()

Del last confirmed version of application.

addConfirmedProposal :: (MonadTrans t, MonadPoll m', t m' ~ m) => ConfirmedProposalState -> m ()

Add new confirmed update proposal.

delConfirmedProposal :: (MonadTrans t, MonadPoll m', t m' ~ m) => SoftwareVersion -> m ()

Del confirmed update proposal (for rollback only).

addActiveProposal :: (MonadTrans t, MonadPoll m', t m' ~ m) => ProposalState -> m ()

Add new active proposal with its state.

deactivateProposal :: (MonadTrans t, MonadPoll m', t m' ~ m) => UpId -> m ()

Delete active proposal given its name and identifier.

setSlottingData :: (MonadTrans t, MonadPoll m', t m' ~ m) => SlottingData -> m ()

Set most recent SlottingData.

newtype DBPoll m a

Constructors

DBPoll

Fields

runDBPoll :: m a

data PollVerFailure

PollVerFailure represents all possible errors which can appear in Poll data verification.

Constructors

PollWrongScriptVersion

Fields

pwsvExpected :: !ScriptVersion

pwsvFound :: !ScriptVersion

pwsvUpId :: !UpId

PollWrongSlotDuration
Slot duration for this block version is already known and the one we saw doesn't match it

Fields

pwsdExpected :: !Millisecond

pwsdFound :: !Millisecond

pwsdUpId :: !UpId

PollWrongMaxBlockSize
Max block size for this block version is already known and the one we saw doesn't match it

Fields

pwmbsExpected :: !Byte

pwmbsFound :: !Byte

pwmbsUpId :: !UpId

PollLargeMaxBlockSize
A proposal tried to increase the block size limit more than it was allowed to

Fields

plmbsMaxPossible :: !Byte

plmbsFound :: !Byte

plmbsUpId :: !UpId

PollNotFoundScriptVersion !BlockVersion

PollSmallProposalStake

Fields

pspsThreshold :: !Coin

pspsActual :: !Coin

pspsUpId :: !UpId

PollNotRichman

Fields

pnrStakeholder :: !StakeholderId

pnrThreshold :: !Coin

pnrStake :: !(Maybe Coin)

PollUnknownProposal

Fields

pupStakeholder :: !StakeholderId

pupProposal :: !UpId

PollUnknownStakes !EpochIndex

Poll2ndActiveProposal !SoftwareVersion

PollWrongSoftwareVersion

Fields

pwsvStored :: !(Maybe NumSoftwareVersion)

pwsvApp :: !ApplicationName

pwsvGiven :: !NumSoftwareVersion

pwsvUpId :: !UpId

PollProposalIsDecided

Fields

ppidUpId :: !UpId

ppidStakeholder :: !StakeholderId

PollExtraRevote

Fields

perUpId :: !UpId

perStakeholder :: !StakeholderId

perDecision :: !Bool

PollWrongHeaderBlockVersion

Fields

pwhpvGiven :: !BlockVersion

pwhpvAdopted :: !BlockVersion

PollBadBlockVersion

Fields

pbpvUpId :: !UpId

pbpvGiven :: !BlockVersion

pbpvAdopted :: !BlockVersion

PollTooBigBlock

Fields

ptbbHash :: !HeaderHash

ptbbSize :: !Byte

ptbbLimit :: !Byte

PollInternalError !Text

verifyAndApplyUSPayload :: forall ssc m. (ApplyMode m, Ssc ssc) => Bool -> Either SlotId (MainBlockHeader ssc) -> UpdatePayload -> m ()

Verify UpdatePayload with respect to data provided by MonadPoll. If data is valid it is also applied. Otherwise PollVerificationFailure is thrown using MonadError type class. When first flag is true and proposal is present, genesisUpdateProposalThd is checked for it, otherwise it's not checked. When second argument is 'Left epoch', it means that temporary payload for given slot is applied. When it is 'Right header', it means that payload from block with given header is applied.

rollbackUS :: forall m. MonadPoll m => USUndo -> m ()

Rollback application of UpdatePayload in MonadPoll using payload itself and undo data.

normalizePoll :: forall ssc m. (MonadPoll m, Ssc ssc) => SlotId -> UpdateProposals -> LocalVotes -> m (UpdateProposals, LocalVotes)

Normalize given proposals and votes with respect to current Poll state, i. e. remove everything that is invalid. Valid data is applied. This function doesn't consider genesisUpdateProposalThd.

filterProposalsByThd :: forall m. MonadPollRead m => EpochIndex -> UpdateProposals -> m (UpdateProposals, HashSet UpId)

Base

canCreateBlockBV :: MonadPollRead m => BlockVersion -> m Bool

Check whether block with given BlockVersion can be created according to current Poll.

Specifically, one of the following conditions must be true. • Given block version is equal to last adopted block version. • '(major, minor)' of given block version is greater than '(major, minor)' of adopted block version and this block version is confirmed.

verifyBlockSize :: (MonadError PollVerFailure m, MonadPollRead m) => HeaderHash -> Byte -> m ()

Verify that size of block doesn't exceed currently adopted limit.

Softfork resolution

processGenesisBlock :: (MonadError PollVerFailure m, MonadPoll m) => EpochIndex -> m ()

Process creation of genesis block for given epoch.

recordBlockIssuance :: (MonadError PollVerFailure m, MonadPoll m) => StakeholderId -> BlockVersion -> SlotId -> HeaderHash -> m ()

Record the fact that main block with given version and leader has been issued by for the given slot.

modifyPollModifier :: PollModifier -> PollModifier -> PollModifier

Unite two PollModifiers. Second argument dominates, i. e. if there are two confliciting modifications, the second one wins.

newtype RollT m a

Constructors

RollT

Fields

getRollT :: StateT USUndo m a

runRollT :: RollT m a -> m (a, USUndo)

execRollT :: Monad m => RollT m a -> m USUndo

newtype PollT m a

Monad transformer which stores PollModifier and implements writable MonadPoll.

WARNING

This transformer uses StateT and is intended for single-threaded usage only.

Constructors

PollT

Fields

getPollT :: StateT PollModifier m a

runPollT :: PollModifier -> PollT m a -> m (a, PollModifier)

evalPollT :: Functor m => PollModifier -> PollT m a -> m a

execPollT :: Functor m => PollModifier -> PollT m a -> m PollModifier

Proposal state

data UndecidedProposalState

State of UpdateProposal which can't be classified as approved or rejected.

Constructors

UndecidedProposalState

Fields

upsVotes :: !StakeholderVotes

Votes given for this proposal.

upsProposal :: !UpdateProposal

Proposal itself.

upsSlot :: !SlotId

SlotId from block in which update was proposed.

upsPositiveStake :: !Coin

Total stake of all positive votes.

upsNegativeStake :: !Coin

Total stake of all negative votes.

upsExtra :: !(Maybe UpsExtra)

Extra data

data DecidedProposalState

State of UpdateProposal which can be classified as approved or rejected.

Constructors

DecidedProposalState

Fields

dpsDecision :: !Bool

Whether proposal is approved.

dpsUndecided :: !UndecidedProposalState

Corresponding UndecidedProposalState

dpsDifficulty :: !(Maybe ChainDifficulty)

Difficulty at which this proposal became approved/rejected. Can be Nothing in temporary state.

dpsExtra :: !(Maybe DpsExtra)

Extra data

data ProposalState

State of UpdateProposal.

Constructors

PSUndecided !UndecidedProposalState

PSDecided !DecidedProposalState

data UpsExtra

Extra data required by wallet, stored in UndecidedProposalState

Constructors

UpsExtra

Fields

ueProposedBlk :: !HeaderHash

Block in which this update was proposed

data DpsExtra

Extra data required by wallet, stored in DecidedProposalState.

Constructors

DpsExtra

Fields

deDecidedBlk :: !HeaderHash

HeaderHash of block in which this update was approved/rejected

deImplicit :: !Bool

Which way we approve/reject this update proposal: implicit or explicit

data ConfirmedProposalState

Information about confirmed proposals stored in DB.

Constructors

ConfirmedProposalState

Fields

cpsUpdateProposal :: !UpdateProposal

cpsImplicit :: !Bool

cpsProposed :: !HeaderHash

cpsDecided :: !HeaderHash

cpsConfirmed :: !HeaderHash

cpsAdopted :: !(Maybe HeaderHash)

cpsVotes :: !StakeholderVotes

cpsPositiveStake :: !Coin

cpsNegativeStake :: !Coin

cpsBlockVersion :: ConfirmedProposalState -> BlockVersion

Get BlockVersion from ConfirmedProposalState.

cpsSoftwareVersion :: ConfirmedProposalState -> SoftwareVersion

Get SoftwareVersion from ConfirmedProposalState.

psProposal :: ProposalState -> UpdateProposal

psVotes :: ProposalState -> StakeholderVotes

mkUProposalState :: SlotId -> UpdateProposal -> UndecidedProposalState

Make UndecidedProposalState from immutable data, i. e. SlotId and UpdateProposal.

BlockVersion state

data BlockVersionState

State of BlockVersion from update proposal.

Constructors

BlockVersionState

Fields

bvsData :: !BlockVersionData

BlockVersioData associated with this block version.

bvsIsConfirmed :: !Bool

Whether proposal with this block version is confirmed.

bvsIssuersStable :: !(HashSet StakeholderId)

Identifiers of stakeholders which issued stable blocks with this BlockVersion. Stability is checked by the same rules as used in LRC. That is, SlotId is considered. If block is created after crucial slot of i-th epoch, it is not stable when 'i+1'-th epoch starts.

bvsIssuersUnstable :: !(HashSet StakeholderId)

Identifiers of stakeholders which issued unstable blocks with this BlockVersion. See description of bvsIssuersStable for details.

bvsLastBlockStable :: !(Maybe HeaderHash)

Identifier of last block which modified set of bvsIssuersStable.

bvsLastBlockUnstable :: !(Maybe HeaderHash)

Identifier of last block which modified set of bvsIssuersUnstable.

bvsScriptVersion :: BlockVersionState -> ScriptVersion

bvsSlotDuration :: BlockVersionState -> Millisecond

bvsMaxBlockSize :: BlockVersionState -> Byte

Poll modifier

data PollModifier

PollModifier is used in verification. It represents operation which one should apply to global state to obtain result of application of MemPool or blocks which are verified.

Constructors

PollModifier

Fields

pmNewBVs :: !(HashMap BlockVersion BlockVersionState)

pmDelBVs :: !(HashSet BlockVersion)

pmAdoptedBVFull :: !(Maybe (BlockVersion, BlockVersionData))

pmNewConfirmed :: !(HashMap ApplicationName NumSoftwareVersion)

pmDelConfirmed :: !(HashSet ApplicationName)

pmNewConfirmedProps :: !(HashMap SoftwareVersion ConfirmedProposalState)

pmDelConfirmedProps :: !(HashSet SoftwareVersion)

pmNewActiveProps :: !(HashMap UpId ProposalState)

pmDelActiveProps :: !(HashSet UpId)

pmNewActivePropsIdx :: !(HashMap ApplicationName UpId)

pmDelActivePropsIdx :: !(HashMap ApplicationName UpId)

pmSlottingData :: !(Maybe SlottingData)

pmNewBVsL :: Lens' PollModifier (HashMap BlockVersion BlockVersionState)

pmDelBVsL :: Lens' PollModifier (HashSet BlockVersion)

pmAdoptedBVFullL :: Lens' PollModifier (Maybe (BlockVersion, BlockVersionData))

pmNewConfirmedL :: Lens' PollModifier (HashMap ApplicationName NumSoftwareVersion)

pmDelConfirmedL :: Lens' PollModifier (HashSet ApplicationName)

pmNewConfirmedPropsL :: Lens' PollModifier (HashMap SoftwareVersion ConfirmedProposalState)

pmDelConfirmedPropsL :: Lens' PollModifier (HashSet SoftwareVersion)

pmNewActivePropsL :: Lens' PollModifier (HashMap UpId ProposalState)

pmDelActivePropsL :: Lens' PollModifier (HashSet UpId)

pmNewActivePropsIdxL :: Lens' PollModifier (HashMap ApplicationName UpId)

pmDelActivePropsIdxL :: Lens' PollModifier (HashMap ApplicationName UpId)

pmSlottingDataL :: Lens' PollModifier (Maybe SlottingData)

Rollback

data PrevValue a

Previous value of something that could be missing.

Constructors

PrevValue a

NoExist

maybeToPrev :: Maybe a -> PrevValue a

data USUndo

Data necessary to unapply US data.

Constructors

USUndo

Fields

unChangedBV :: !(HashMap BlockVersion (PrevValue BlockVersionState))

unLastAdoptedBV :: !(Maybe BlockVersion)

unChangedProps :: !(HashMap UpId (PrevValue ProposalState))

unChangedSV :: !(HashMap ApplicationName (PrevValue NumSoftwareVersion))

unChangedConfProps :: !(HashMap SoftwareVersion (PrevValue ConfirmedProposalState))

unChangedSVL :: Lens' USUndo (HashMap ApplicationName (PrevValue NumSoftwareVersion))

unChangedPropsL :: Lens' USUndo (HashMap UpId (PrevValue ProposalState))

unChangedBVL :: Lens' USUndo (HashMap BlockVersion (PrevValue BlockVersionState))

unLastAdoptedBVL :: Lens' USUndo (Maybe BlockVersion)

unChangedConfPropsL :: Lens' USUndo (HashMap SoftwareVersion (PrevValue ConfirmedProposalState))

usWorkers :: WorkMode ssc m => ([WorkerSpec m], OutSpecs)

Update System related workers.