Pos.Util

Description

Miscellaneous unclassified utility functions.

Stuff for testing and benchmarking

class Nonrepeating a where

Type for generating list of unique (nonrepeating) elemets.

Minimal complete definition

nonrepeating

Methods

nonrepeating :: Int -> Gen [a]

class ArbitraryUnsafe a where

ArbitraryUnsafe class ~~~~~~~~~~~~~~~~~~~~~~~~

This class is the same as Arbitrary, except instances of this class for stuff like public/secret keys, VSS shares, commitments etc. are designed not to mimic real data as presisely as possible (using OpenSSL random), but rather to be simple and efficient.

This is especially useful for benchmarking.

Note: we don't need Generic to derive instances of ArbitraryUnsafe. We can either use one-line instance declaration, or -XStandaloneDeriving or -XDeriveAnyClass to write something like deriving (Arbitrary, ArbitraryUnsafe).

Methods

arbitraryUnsafe :: Gen a

arbitraryUnsafe :: Arbitrary a => Gen a

makeSmall :: Gen a -> Gen a

sublistN :: Int -> [a] -> Gen [a]

Choose a random (shuffled) subset of length n. Throws an error if there's not enough elements.

unsafeMakeList :: Text -> IO [a] -> [a]

Unsafely create list of n random values to be picked (see note in Arbitrary for explanation) Used because genSharedSecret already returns a list of EncShares, making the replicateM unneeded.

unsafeMakePool :: Text -> Int -> IO a -> [a]

Unsafely create pool of n random values to be picked (see note in Arbitrary for explanation)

arbitrarySizedS :: Int -> Gen ByteString

Make arbitrary ByteString of given length.

arbitrarySizedSL :: Int -> Gen ByteString

Make arbitrary Lazy of given length.

runGen :: Gen a -> a

Get something out of a quickcheck Gen without having to do IO

module Pos.Util.TimeLimit

Various

mappendPair :: (Monoid a, Monoid b) => (a, b) -> (a, b) -> (a, b)

mconcatPair :: (Monoid a, Monoid b) => [(a, b)] -> (a, b)

(<//>) :: String -> String -> String

Concatenates two url part using regular slash /. E.g. ".dir" // "file" = ".dir/file".

readerToState :: MonadState s m => Reader s a -> m a

Convert (Reader s) to any (MonadState s)

eitherPanic :: Show a => Text -> Either a b -> b

A helper for simple error handling in executables

inAssertMode :: Applicative m => m a -> m ()

This function performs checks at compile-time for different actions. May slowdown implementation. To disable such checks (especially in benchmarks) one should compile with: stack build --flag cardano-sl:-asserts

diffDoubleMap :: forall k1 k2 v. (Eq k1, Eq k2, Hashable k1, Hashable k2) => HashMap k1 (HashMap k2 v) -> HashMap k1 (HashMap k2 v) -> HashMap k1 (HashMap k2 v)

Remove elements which are in b from a

getKeys :: HashMap k v -> HashSet k

Create HashSet from HashMap's keys

maybeThrow :: (MonadThrow m, Exception e) => e -> Maybe a -> m a

maybeThrow' :: (Mockable Throw m, Exception e) => e -> Maybe a -> m a

parseIntegralSafe :: Integral a => Parser a

Lists

allDistinct :: Ord a => [a] -> Bool

NonEmpty

type NE = NonEmpty

neZipWith3 :: (x -> y -> z -> q) -> NonEmpty x -> NonEmpty y -> NonEmpty z -> NonEmpty q

Chronological sequences

newtype NewestFirst f a

Constructors

NewestFirst

Fields

getNewestFirst :: f a

newtype OldestFirst f a

Constructors

OldestFirst

Fields

getOldestFirst :: f a

toNewestFirst :: Chrono f => OldestFirst f a -> NewestFirst f a

toOldestFirst :: Chrono f => NewestFirst f a -> OldestFirst f a

Lenses

makeLensesData :: Name -> Name -> DecsQ

Make lenses for a data family instance.

magnify' :: MonadReader s m => LensLike' (Magnified (Reader s) a) s t -> ReaderT t Identity a -> m a

A magnify which in MonadReader.

_neHead :: Lens' (NonEmpty a) a

Lens for the head of NonEmpty.

We can't use _head because it doesn't work for NonEmpty: https://github.com/ekmett/lens/issues/636#issuecomment-213981096. Even if we could though, it wouldn't be a lens, only a traversal.

_neTail :: Lens' (NonEmpty a) [a]

Lens for the tail of NonEmpty.

_neLast :: Lens' (NonEmpty a) a

Lens for the last element of NonEmpty.

zoom' :: MonadState s m => LensLike' (Zoomed (State s) a) s t -> StateT t Identity a -> m a

A zoom which works in MonadState.

See https://github.com/ekmett/lens/issues/580. You might be surprised but actual zoom doesn't work in any MonadState, it only works in a handful of state monads and their combinations defined by Zoom.

Prettification

data Color :: *

Constructors

Black

Red

Green

Yellow

Blue

Magenta

Cyan

White

colorize :: Color -> Text -> Text

Prettify Text message with Vivid color.

withColoredMessages :: MonadIO m => Color -> Text -> m a -> m a

Write colored message, do some action, write colored message. Intended for debug only.

LRU

clearLRU :: Ord k => LRU k v -> LRU k v

Remove all items from LRU, retaining maxSize property.

spanSafe :: (a -> a -> Bool) -> NonEmpty a -> (NonEmpty a, [a])

Makes a span on the list, considering tail only. Predicate has list head as first argument. Used to take non-null prefix that depends on the first element.

eitherToVerRes :: Either Text a -> VerificationRes

Concurrency

clearMVar :: MonadIO m => MVar a -> m ()

forcePutMVar :: MonadIO m => MVar a -> a -> m ()

readMVarConditional :: MonadIO m => (x -> Bool) -> MVar x -> m x

Block until value in MVar satisfies given predicate. When value satisfies, it is returned.

readUntilEqualMVar :: (Eq a, MonadIO m) => (x -> a) -> MVar x -> a -> m x

Read until value is equal to stored value comparing by some function.

readTVarConditional :: MonadIO m => (x -> Bool) -> TVar x -> m x

Block until value in TVar satisfies given predicate. When value satisfies, it is returned.

readUntilEqualTVar :: (Eq a, MonadIO m) => (x -> a) -> TVar x -> a -> m x

Read until value is equal to stored value comparing by some function.

withReadLifted :: (MonadIO m, MonadMask m) => RWLock -> m a -> m a

withWriteLifted :: (MonadIO m, MonadMask m) => RWLock -> m a -> m a

Instances

Lift Byte

FromJSON Byte

ToJSON Byte

SafeCopy (NonEmpty a)

SafeCopy Microsecond

SafeCopy Millisecond

MonadFail (Either s), assuming IsString s

MonadFail ParsecT

MonadFail Dialog

MonadFail Transfer

MonadFail TimedIO

MonadFail ResponseT

MonadFail LoggerNameBox

Orphan instances

Lift Byte

Methods

lift :: Byte -> Q Exp

SafeCopy Millisecond

Methods

version :: Version Millisecond

kind :: Kind Millisecond

getCopy :: Contained (Get Millisecond)

putCopy :: Millisecond -> Contained Put

internalConsistency :: Consistency Millisecond

objectProfile :: Profile Millisecond

errorTypeName :: Proxy Millisecond -> String

SafeCopy Microsecond

Methods

version :: Version Microsecond

kind :: Kind Microsecond

getCopy :: Contained (Get Microsecond)

putCopy :: Microsecond -> Contained Put

internalConsistency :: Consistency Microsecond

objectProfile :: Profile Microsecond

errorTypeName :: Proxy Microsecond -> String

SafeCopy VerificationRes

Methods

version :: Version VerificationRes

kind :: Kind VerificationRes

getCopy :: Contained (Get VerificationRes)

putCopy :: VerificationRes -> Contained Put

internalConsistency :: Consistency VerificationRes

objectProfile :: Profile VerificationRes

errorTypeName :: Proxy VerificationRes -> String

ToJSON Byte

Methods

toJSON :: Byte -> Value

toEncoding :: Byte -> Encoding

toJSONList :: [Byte] -> Value

toEncodingList :: [Byte] -> Encoding

FromJSON Byte

Methods

parseJSON :: Value -> Parser Byte

parseJSONList :: Value -> Parser [Byte]

IsString s => MonadFail (Either s)

Methods

fail :: String -> Either s a

MonadFail m => MonadFail (LoggerNameBox m)

Methods

fail :: String -> LoggerNameBox m a

MonadFail m => MonadFail (ResourceT m)

Methods

fail :: String -> ResourceT m a

SafeCopy a => SafeCopy (NonEmpty a)

Methods

version :: Version (NonEmpty a)

kind :: Kind (NonEmpty a)

getCopy :: Contained (Get (NonEmpty a))

putCopy :: NonEmpty a -> Contained Put

internalConsistency :: Consistency (NonEmpty a)

objectProfile :: Profile (NonEmpty a)

errorTypeName :: Proxy (NonEmpty a) -> String

(Ord k, SafeCopy k, SafeCopy v) => SafeCopy (LRU k v)

Methods

version :: Version (LRU k v)

kind :: Kind (LRU k v)

getCopy :: Contained (Get (LRU k v))

putCopy :: LRU k v -> Contained Put

internalConsistency :: Consistency (LRU k v)

objectProfile :: Profile (LRU k v)

errorTypeName :: Proxy (LRU k v) -> String

MonadFail (ParsecT s u m)

Methods

fail :: String -> ParsecT s u m a