Pos.Util.JsonLog

Description

Monadic represantion of something that has json journaled log of operations.

Documentation

data JLEvent

Json log event.

Constructors

JLCreatedBlock JLBlock

JLAdoptedBlock BlockId

JLTpsStat Int

data JLBlock

Json log of one block with corresponding BlockId.

Constructors

JLBlock

Fields

jlHash :: BlockId

jlPrevBlock :: BlockId

jlTxs :: [TxId]

jlSlot :: JLSlotId

data JLTimedEvent

JLEvent with Timestamp -- corresponding time of this event.

Constructors

JLTimedEvent

Fields

jlTimestamp :: Integer

jlEvent :: JLEvent

jlCreatedBlock :: BiSsc ssc => Block ssc -> JLEvent

Return event of created block.

jlAdoptedBlock :: Ssc ssc => Block ssc -> JLEvent

Returns event of created Block.

class Monad m => MonadJL m where

Monad for things that can log Json log events.

Minimal complete definition

jlLog

Methods

jlLog :: JLEvent -> m ()

appendJL :: MonadIO m => FilePath -> JLEvent -> m ()

Append event into log by given FilePath.

fromJLSlotId :: JLSlotId -> SlotId

Get SlotId from JLSlotId.