Pos.Wallet

Documentation

data WalletParams

Constructors

WalletParams

Fields

wpDbPath :: !(Maybe FilePath)

wpRebuildDb :: !Bool

wpKeyFilePath :: !FilePath

wpSystemStart :: Timestamp

wpGenesisKeys :: !Bool

wpBaseParams :: !BaseParams

wpGenesisUtxo :: !Utxo

Genesis utxo

runRawRealWallet :: RealModeResources -> WalletParams -> ListenersWithOut WalletRealMode -> (ActionSpec WalletRealMode a, OutSpecs) -> Production a

runWalletRealMode :: RealModeResources -> WalletParams -> (ActionSpec WalletRealMode a, OutSpecs) -> Production a

WalletMode runner

runWalletReal :: RealModeResources -> WalletParams -> ([WorkerSpec WalletRealMode], OutSpecs) -> Production ()

runWallet :: WalletMode ssc m => ([WorkerSpec m], OutSpecs) -> (WorkerSpec m, OutSpecs)

submitVote :: (MinWorkMode m, MonadDBLimits m) => SendActions m -> [DHTNode] -> UpdateVote -> m ()

Send UpdateVote to given addresses

submitUpdateProposal :: (MinWorkMode m, MonadDBLimits m) => SendActions m -> SecretKey -> [DHTNode] -> UpdateProposal -> m ()

Send UpdateProposal with one positive vote to given addresses

sendVoteOuts :: OutSpecs

sendProposalOuts :: OutSpecs

class Monad m => MonadBalances m where

A class which have the methods to get state of address' balance

Methods

getOwnUtxo :: Address -> m Utxo

getBalance :: Address -> m Coin

getOwnUtxo :: (MonadTrans t, MonadBalances m', t m' ~ m) => Address -> m Utxo

class Monad m => MonadTxHistory m where

A class which have methods to get transaction history

Methods

getTxHistory :: SscHelpersClass ssc => Tagged ssc (Address -> Maybe (HeaderHash, Utxo) -> m TxHistoryAnswer)

saveTx :: (TxId, TxAux) -> m ()

getTxHistory :: (SscHelpersClass ssc, MonadTrans t, MonadTxHistory m', t m' ~ m) => Tagged ssc (Address -> Maybe (HeaderHash, Utxo) -> m TxHistoryAnswer)

saveTx :: (MonadTrans t, MonadTxHistory m', t m' ~ m) => (TxId, TxAux) -> m ()

class Monad m => MonadBlockchainInfo m where

Methods

networkChainDifficulty :: m (Maybe ChainDifficulty)

localChainDifficulty :: m ChainDifficulty

blockchainSlotDuration :: m Millisecond

connectedPeers :: m Word

networkChainDifficulty :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m (Maybe ChainDifficulty)

localChainDifficulty :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m ChainDifficulty

blockchainSlotDuration :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m Millisecond

connectedPeers :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m Word

class Monad m => MonadUpdates m where

Abstraction over getting update proposals

Methods

waitForUpdate :: m ConfirmedProposalState

applyLastUpdate :: m ()

waitForUpdate :: (MonadTrans t, MonadUpdates m', t m' ~ m) => m ConfirmedProposalState

applyLastUpdate :: (MonadTrans t, MonadUpdates m', t m' ~ m) => m ()

type WalletMode ssc m = (TxMode ssc m, MonadKeys m, MonadBlockchainInfo m, MonadUpdates m, WithWalletContext m, MonadDHT m, WithPeerState m)

type WalletRealMode = PeerStateHolder (KademliaDHT (KeyStorage (WalletDB (ContextHolder (LoggerNameBox Production)))))