Pos.Wallet

Documentation

data WalletParams

Constructors

WalletParams

Fields

wpDbPath :: !(Maybe FilePath)

wpRebuildDb :: !Bool

wpKeyFilePath :: !FilePath

wpSystemStart :: !Timestamp

wpGenesisKeys :: !Bool

wpBaseParams :: !BaseParams

wpGenesisUtxo :: !Utxo

Genesis utxo

runRawRealWallet :: RealModeResources -> WalletParams -> ListenersWithOut WalletRealMode -> (ActionSpec WalletRealMode a, OutSpecs) -> Production a

runWalletRealMode :: RealModeResources -> WalletParams -> (ActionSpec WalletRealMode a, OutSpecs) -> Production a

WalletMode runner

runWalletReal :: RealModeResources -> WalletParams -> ([WorkerSpec WalletRealMode], OutSpecs) -> Production ()

runWallet :: WalletMode ssc m => ([WorkerSpec m], OutSpecs) -> (WorkerSpec m, OutSpecs)

makePubKeyTx :: SafeSigner -> TxInputs -> TxOutputs -> TxAux

Makes a transaction which use P2PKH addresses as a source

makeMOfNTx :: Script -> [Maybe SafeSigner] -> TxInputs -> TxOutputs -> TxAux

submitTx :: TxMode ssc m => SendActions m -> SafeSigner -> [DHTNode] -> NonEmpty TxOutAux -> m (Either TxError TxAux)

Construct Tx using secret key and given list of desired outputs

submitRedemptionTx :: TxMode ssc m => SendActions m -> RedeemSecretKey -> [DHTNode] -> Address -> m (Either TxError TxAux)

Construct redemption Tx using redemption secret key and a output address

submitTxRaw :: (MinWorkMode m, MonadDBLimits m) => SendActions m -> [DHTNode] -> TxAux -> m ()

Send the ready-to-use transaction

createTx :: Utxo -> SafeSigner -> TxOutputs -> Either TxError TxAux

Make a multi-transaction using given secret key and info for outputs

createMOfNTx :: Utxo -> [(PublicKey, Maybe SafeSigner)] -> TxOutputs -> Either TxError TxAux

Make a transaction, using M-of-N script as a source

sendTxOuts :: OutSpecs

submitVote :: (MinWorkMode m, MonadDBLimits m) => SendActions m -> [DHTNode] -> UpdateVote -> m ()

Send UpdateVote to given addresses

submitUpdateProposal :: (MinWorkMode m, MonadDBLimits m) => SendActions m -> SecretKey -> [DHTNode] -> UpdateProposal -> m ()

Send UpdateProposal with one positive vote to given addresses

sendVoteOuts :: OutSpecs

sendProposalOuts :: OutSpecs

class Monad m => MonadBalances m where

A class which have the methods to get state of address' balance

Methods

getOwnUtxo :: Address -> m Utxo

getBalance :: Address -> m Coin

getOwnUtxo :: (MonadTrans t, MonadBalances m', t m' ~ m) => Address -> m Utxo

class Monad m => MonadTxHistory m where

A class which have methods to get transaction history

Methods

getTxHistory :: SscHelpersClass ssc => Tagged ssc (Address -> Maybe (HeaderHash, Utxo) -> m TxHistoryAnswer)

saveTx :: (TxId, TxAux) -> m ()

getTxHistory :: (SscHelpersClass ssc, MonadTrans t, MonadTxHistory m', t m' ~ m) => Tagged ssc (Address -> Maybe (HeaderHash, Utxo) -> m TxHistoryAnswer)

saveTx :: (MonadTrans t, MonadTxHistory m', t m' ~ m) => (TxId, TxAux) -> m ()

data TxHistoryAnswer

Constructors

TxHistoryAnswer

Fields

taLastCachedHash :: HeaderHash

taCachedNum :: Int

taCachedUtxo :: Utxo

taHistory :: [TxHistoryEntry]

class Monad m => MonadBlockchainInfo m where

Methods

networkChainDifficulty :: m (Maybe ChainDifficulty)

localChainDifficulty :: m ChainDifficulty

blockchainSlotDuration :: m Millisecond

connectedPeers :: m Word

networkChainDifficulty :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m (Maybe ChainDifficulty)

localChainDifficulty :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m ChainDifficulty

blockchainSlotDuration :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m Millisecond

connectedPeers :: (MonadTrans t, MonadBlockchainInfo m', t m' ~ m) => m Word

class Monad m => MonadUpdates m where

Abstraction over getting update proposals

Methods

waitForUpdate :: m ConfirmedProposalState

applyLastUpdate :: m ()

waitForUpdate :: (MonadTrans t, MonadUpdates m', t m' ~ m) => m ConfirmedProposalState

applyLastUpdate :: (MonadTrans t, MonadUpdates m', t m' ~ m) => m ()

type TxMode ssc m = (MinWorkMode m, MonadBalances m, MonadTxHistory m, MonadMockable m, MonadMask m, MonadDBLimits m)

type WalletMode ssc m = (TxMode ssc m, MonadKeys m, MonadBlockchainInfo m, MonadUpdates m, WithWalletContext m, MonadDHT m, WithPeerState m)

type WalletRealMode = PeerStateHolder (KademliaDHT (KeyStorage (WalletDB (ContextHolder (LoggerNameBox Production)))))