Pos.Wallet.Web

Description

Web part of wallet.

Documentation

type WalletApi = ("api" :> ("get_wallet" :> (Capture "address" CAddress :> Get '[JSON] (Either WalletError CWallet)))) :<|> (("api" :> ("get_wallets" :> Get '[JSON] (Either WalletError [CWallet]))) :<|> (("api" :> ("send" :> (Capture "from" CAddress :> (Capture "to" CAddress :> (Capture "amount" Coin :> Post '[JSON] (Either WalletError CTx)))))) :<|> (("api" :> ("send" :> (Capture "from" CAddress :> (Capture "to" CAddress :> (Capture "amount" Coin :> (Capture "currency" CCurrency :> (Capture "title" Text :> (Capture "description" Text :> Post '[JSON] (Either WalletError CTx))))))))) :<|> (("api" :> ("txhistory" :> (Capture "address" CAddress :> (Capture "skip" Word :> (Capture "limit" Word :> Get '[JSON] (Either WalletError ([CTx], Word))))))) :<|> (("api" :> ("search_txhistory" :> (Capture "address" CAddress :> (Capture "search" Text :> (Capture "skip" Word :> (Capture "limit" Word :> Get '[JSON] (Either WalletError ([CTx], Word)))))))) :<|> (("api" :> ("update_transaction" :> (Capture "address" CAddress :> (Capture "transaction" CTxId :> (ReqBody '[JSON] CTxMeta :> Post '[JSON] (Either WalletError ())))))) :<|> (("api" :> ("new_wallet" :> (ReqBody '[JSON] CWalletInit :> Post '[JSON] (Either WalletError CWallet)))) :<|> (("api" :> ("restore_wallet" :> (ReqBody '[JSON] CWalletInit :> Post '[JSON] (Either WalletError CWallet)))) :<|> (("api" :> ("update_wallet" :> (Capture "address" CAddress :> (ReqBody '[JSON] CWalletMeta :> Post '[JSON] (Either WalletError CWallet))))) :<|> (("api" :> ("delete_wallet" :> (Capture "address" CAddress :> Post '[JSON] (Either WalletError ())))) :<|> (("api" :> ("valid_address" :> (Capture "currency" CCurrency :> (Capture "address" Text :> Get '[JSON] (Either WalletError Bool))))) :<|> (("api" :> ("get_profile" :> Get '[JSON] (Either WalletError CProfile))) :<|> (("api" :> ("update_profile" :> (ReqBody '[JSON] CProfile :> Post '[JSON] (Either WalletError CProfile)))) :<|> (("api" :> ("redeem_ada" :> (ReqBody '[JSON] CWalletRedeem :> Post '[JSON] (Either WalletError CWallet)))) :<|> (("api" :> ("next_update" :> Get '[JSON] (Either WalletError CUpdateInfo))) :<|> (("api" :> ("apply_update" :> Post '[JSON] (Either WalletError ()))) :<|> (("api" :> ("slot_duration" :> Get '[JSON] (Either WalletError Word))) :<|> (("api" :> ("system_version" :> Get '[JSON] (Either WalletError SoftwareVersion))) :<|> ("api" :> ("postpone_update" :> (Capture "time" POSIXTime :> Post '[JSON] (Either WalletError ()))))))))))))))))))))))

Servant API which provides access to wallet.

walletApi :: Proxy WalletApi

Helper Proxy.

data CAddress

Client address

data CCurrency

currencies handled by client Note: Cardano does not deal with other currency than ADA yet

Constructors

ADA

BTC

ETH

data CHash

Client hash

data CTType

type of transactions It can be an input output exchange transaction CTInOut CTExMeta -- Ex == exchange

Constructors

CTIn CTxMeta

CTOut CTxMeta

data CProfile

Client profile (CP) all data of client are "meta data" - that is not provided by Cardano (Flow type: accountType)

Constructors

CProfile

Fields

cpName :: Text

cpEmail :: Text

cpPhoneNumber :: Text

cpPwHash :: CPwHash

cpPwCreated :: POSIXTime

cpLocale :: Text

cpPicture :: Text

type CPwHash = Text

Password hash of client profile

data CTx

Client transaction (CTx) Provides all Data about a transaction needed by client. It includes meta data which are not part of Cardano, too (Flow type: transactionType)

Constructors

CTx

Fields

ctId :: CTxId

ctAmount :: Coin

ctConfirmations :: Word

ctType :: CTType

data CTxId

Client transaction id

data CTxMeta

meta data of transactions

Constructors

CTxMeta

Fields

ctmCurrency :: CCurrency

ctmTitle :: Text

ctmDescription :: Text

ctmDate :: POSIXTime

data CTExMeta

meta data of exchanges

Constructors

CTExMeta

Fields

cexCurrency :: CCurrency

cexTitle :: Text

cexDescription :: Text

cexDate :: POSIXTime

cexRate :: Text

cexLabel :: Text

cexAddress :: CAddress

data CWallet

Client Wallet (CW) (Flow type: walletType)

Constructors

CWallet

Fields

cwAddress :: !CAddress

cwAmount :: !Coin

cwMeta :: !CWalletMeta

data CWalletType

A wallet can be used as personal or shared wallet

Constructors

CWTPersonal

CWTShared

data CWalletMeta

Meta data of CWallet Includes data which are not provided by Cardano

Constructors

CWalletMeta

Fields

cwType :: !CWalletType

cwCurrency :: !CCurrency

cwName :: !Text

data CWalletInit

Query data for wallet creation (wallet meta + backup phrase)

Constructors

CWalletInit

Fields

cwBackupPhrase :: !BackupPhrase

cwInitMeta :: !CWalletMeta

data CUpdateInfo

Update system data

Constructors

CUpdateInfo

Fields

cuiSoftwareVersion :: !SoftwareVersion

cuiBlockVesion :: !BlockVersion

cuiScriptVersion :: !ScriptVersion

cuiImplicit :: !Bool

cuiVotesFor :: !Int

cuiVotesAgainst :: !Int

cuiPositiveStake :: !Coin

cuiNegativeStake :: !Coin

data CWalletRedeem

Query data for redeem

Constructors

CWalletRedeem

Fields

crWalletId :: !CAddress

crSeed :: !Text

data NotifyEvent

Constructors

ConnectionOpened

NetworkDifficultyChanged ChainDifficulty
NewWalletTransaction CAddress | NewTransaction

LocalDifficultyChanged ChainDifficulty

ConnectedPeersChanged Word

UpdateAvailable

ConnectionClosed

addressToCAddress :: Address -> CAddress

transform Address into CAddress TODO: this is not complitely safe. If someone changes implementation of Buildable Address. It should be probably more safe to introduce `class PSSimplified` that would have the same implementation has it is with Buildable Address but then person will know it will probably change something for purescript.

cAddressToAddress :: CAddress -> Either Text Address

mkCTx

Arguments

:: Address
An address for which transaction info is forming

-> ChainDifficulty
Current chain difficulty (to get confirmations)

-> TxHistoryEntry
Tx history entry

-> CTxMeta
Transaction metadata

-> CTx

mkCTxId :: Text -> CTxId

txIdToCTxId :: TxId -> CTxId

transform TxId into CTxId

ctTypeMeta :: CTType -> CTxMeta

txContainsTitle :: Text -> CTx -> Bool

toCUpdateInfo :: ConfirmedProposalState -> CUpdateInfo

Creates CTUpdateInfo from ConfirmedProposalState

walletDocsText :: Text

data WalletError

Constructors

Internal !Text
Some internal error.

walletServeWebFull :: SscConstraint ssc => SendActions (RawRealMode ssc) -> Bool -> FilePath -> Bool -> Word16 -> RawRealMode ssc ()

walletServerOuts :: OutSpecs

walletServeWebLite :: SendActions WalletRealMode -> FilePath -> Bool -> Word16 -> WalletRealMode ()

walletServerOuts :: OutSpecs

walletApplication :: WalletWebMode ssc m => m (Server WalletApi) -> m Application

walletServer :: (Monad m, WalletWebMode ssc (WalletWebHandler m)) => SendActions m -> WalletWebHandler m (WalletWebHandler m :~> Handler) -> WalletWebHandler m (Server WalletApi)

walletServeImpl

Arguments

:: (MonadIO m, MonadMask m, WalletWebMode ssc (WalletWebHandler m))

=> WalletWebHandler m Application
Application getter

-> FilePath
Path to wallet acid-state

-> Bool
Rebuild flag for acid-state

-> Word16
Port to listen

-> m ()

walletServerOuts :: OutSpecs

data WalletWebSockets m a

Holder for web wallet data

type WebWalletSockets m = (MonadWalletWebSockets m, MonadIO m)

class Monad m => MonadWalletWebSockets m where

MonadWalletWebSockets stands for monad which is able to get web wallet sockets

Minimal complete definition

getWalletWebSockets

Methods

getWalletWebSockets :: m ConnectionsVar

type ConnectionsVar = TVar (Maybe Connection)

initWSConnection :: MonadIO m => m ConnectionsVar

closeWSConnection :: MonadIO m => ConnectionsVar -> m ()

upgradeApplicationWS :: ConnectionsVar -> Application -> Application

notify :: WebWalletSockets m => NotifyEvent -> m ()

runWalletWS :: ConnectionsVar -> WalletWebSockets m a -> m a

getWalletWebSocketsState :: Monad m => WalletWebSockets m ConnectionsVar

newtype WalletWebDB m a

Holder for web wallet data

Constructors

WalletWebDB

Fields

getWalletWebDB :: ReaderT WalletState m a

runWalletWebDB :: WalletState -> WalletWebDB m a -> m a

Execute WalletWebDB action with given WalletState

getWalletState :: Monad m => WalletWebDB m WalletState

type WalletState = ExtendedState WalletStorage

class Monad m => MonadWalletWebDB m where

MonadWalletWebDB stands for monad which is able to get web wallet state

Minimal complete definition

getWalletWebState

Methods

getWalletWebState :: m WalletState

type WebWalletModeDB m = (MonadWalletWebDB m, MonadIO m, MonadMockable m)

Constraint for working with web wallet DB

openState :: MonadIO m => Bool -> FilePath -> m WalletState

openMemState :: MonadIO m => m WalletState

closeState :: MonadIO m => WalletState -> m ()

Getters

getProfile :: WebWalletModeDB m => m (Maybe CProfile)

getWalletMetas :: WebWalletModeDB m => m [CWalletMeta]

getWalletMeta :: WebWalletModeDB m => CAddress -> m (Maybe CWalletMeta)

getTxMeta :: WebWalletModeDB m => CAddress -> CTxId -> m (Maybe CTxMeta)

getWalletHistory :: WebWalletModeDB m => CAddress -> m (Maybe [CTxMeta])

getUpdates :: WebWalletModeDB m => m [CUpdateInfo]

getNextUpdate :: WebWalletModeDB m => m (Maybe CUpdateInfo)

getPostponeUpdateUntil :: WebWalletModeDB m => m (Maybe POSIXTime)

Setters

createWallet :: WebWalletModeDB m => CAddress -> CWalletMeta -> m ()

setProfile :: WebWalletModeDB m => CProfile -> m ()

setWalletMeta :: WebWalletModeDB m => CAddress -> CWalletMeta -> m ()

setWalletTransactionMeta :: WebWalletModeDB m => CAddress -> CTxId -> CTxMeta -> m ()

setWalletHistory :: WebWalletModeDB m => CAddress -> [(CTxId, CTxMeta)] -> m ()

addOnlyNewTxMeta :: WebWalletModeDB m => CAddress -> CTxId -> CTxMeta -> m ()

removeWallet :: WebWalletModeDB m => CAddress -> m ()

addUpdate :: WebWalletModeDB m => CUpdateInfo -> m ()

removeNextUpdate :: WebWalletModeDB m => m ()

setPostponeUpdateUntil :: WebWalletModeDB m => POSIXTime -> m ()

removePostponeUpdateUntil :: WebWalletModeDB m => m ()