Pos.WorkMode

Description

WorkMode constraint. It is widely used in almost every our code. Simple alias for bunch of useful constraints. This module also contains new monads to extend functional capabilities inside do-block.

Documentation

type WorkMode ssc m = (WithLogger m, MonadIO m, MonadMockable m, MonadDHT m, MonadMask m, MonadSlots m, MonadDB ssc m, MonadTxpLD ssc m, MonadDelegation m, MonadUtxo m, MonadSscGS ssc m, SscStorageClass ssc, SscLocalDataClass ssc, SscHelpersClass ssc, MonadSscLD ssc m, WithNodeContext ssc m, MonadStats m, MonadJL m, WithKademliaDHTInstance m, MonadFail m, WithPeerState ssc m, MonadUSMem m)

Bunch of constraints to perform work for real world distributed system.

type MinWorkMode m = (WithLogger m, MonadMockable m, MonadDHT m, MonadIO m, MonadFail m)

More relaxed version of WorkMode.

Actual modes

type ProductionMode ssc = NoStatsT (RawRealMode ssc)

ProductionMode is an instance of WorkMode which is used (unsurprisingly) in production.

type RawRealMode ssc = PeerStateHolder ssc (KademliaDHT (USHolder (DelegationT (TxpLDHolder ssc (SscHolder ssc (ContextHolder ssc (DBHolder ssc (LoggerNameBox Production))))))))

RawRealMode is a basis for WorkModes used to really run system.

type ServiceMode = KademliaDHT (LoggerNameBox Production)

ServiceMode is the mode in which support nodes work.

type StatsMode ssc = StatsT (RawRealMode ssc)

StatsMode is used for remote benchmarking.

Orphan instances

(Monad m, WithNodeContext ssc m) => WithNodeContext ssc (KademliaDHT m)

Methods

getNodeContext :: KademliaDHT m (NodeContext ssc)

MonadDB ssc m => MonadDB ssc (KademliaDHT m)

Methods

getNodeDBs :: KademliaDHT m (NodeDBs ssc)

usingReadOptions :: ReadOptions -> ASetter' (NodeDBs ssc) (DB ssc) -> KademliaDHT m a -> KademliaDHT m a

usingWriteOptions :: WriteOptions -> ASetter' (NodeDBs ssc) (DB ssc) -> KademliaDHT m a -> KademliaDHT m a

MonadSscGS ssc m => MonadSscGS ssc (KademliaDHT m)

Methods

getGlobalState :: KademliaDHT m (SscGlobalState ssc)

setGlobalState :: SscGlobalState ssc -> KademliaDHT m ()

modifyGlobalState :: (SscGlobalState ssc -> (a, SscGlobalState ssc)) -> KademliaDHT m a

MonadSscRichmen m => MonadSscRichmen (KademliaDHT m)

MonadUtxo m => MonadUtxo (KademliaDHT m)

Methods

utxoPut :: TxIn -> TxOutAux -> KademliaDHT m ()

utxoDel :: TxIn -> KademliaDHT m ()

MonadUtxoRead m => MonadUtxoRead (KademliaDHT m)

Methods

utxoGet :: TxIn -> KademliaDHT m (Maybe TxOutAux)

MonadDelegation m => MonadDelegation (KademliaDHT m)

Methods

askDelegationState :: KademliaDHT m (TVar DelegationWrap)

MonadUSMem m => MonadUSMem (KademliaDHT m)

Methods

askUSMemVar :: KademliaDHT m MemVar

MonadJL m => MonadJL (KademliaDHT m)

Methods

jlLog :: JLEvent -> KademliaDHT m ()