Pos.WorkMode

Description

WorkMode constraint. It is widely used in almost every our code. Simple alias for bunch of useful constraints. This module also contains new monads to extend functional capabilities inside do-block.

Documentation

type WorkMode ssc m = (MinWorkMode m, MonadMask m, MonadSlots m, MonadDB m, MonadDBLimits m, MonadTxpMem m, MonadDelegation m, MonadSscMem ssc m, SscGStateClass ssc, SscLocalDataClass ssc, SscHelpersClass ssc, WithNodeContext ssc m, MonadStats m, MonadJL m, WithKademliaDHTInstance m, WithPeerState m, MonadUSMem m)

Bunch of constraints to perform work for real world distributed system.

type MinWorkMode m = (WithLogger m, MonadMockable m, MonadDHT m, MonadIO m, WithPeerState m)

More relaxed version of WorkMode.

Actual modes

type ProductionMode ssc = NoStatsT (RawRealMode ssc)

ProductionMode is an instance of WorkMode which is used (unsurprisingly) in production.

type RawRealMode ssc = PeerStateHolder (KademliaDHT (USHolder (DelegationT (TxpHolder (SscHolder ssc (NtpSlotting (SlottingHolder (ContextHolder ssc (DBHolder (LoggerNameBox Production))))))))))

RawRealMode is a basis for WorkModes used to really run system.

type ServiceMode = PeerStateHolder (KademliaDHT (LoggerNameBox Production))

ServiceMode is the mode in which support nodes work.

type StatsMode ssc = StatsT (RawRealMode ssc)

StatsMode is used for remote benchmarking.

Orphan instances

MonadSscMem ssc m => MonadSscMem ssc (PeerStateHolder m)

Methods

askSscMem :: PeerStateHolder m (SscState ssc)

(Monad m, WithNodeContext ssc m) => WithNodeContext ssc (PeerStateHolder m)

Methods

getNodeContext :: PeerStateHolder m (NodeContext ssc)

(Monad m, WithNodeContext ssc m) => WithNodeContext ssc (KademliaDHT m)

Methods

getNodeContext :: KademliaDHT m (NodeContext ssc)

MonadDB m => MonadDB (KademliaDHT m)

Methods

getNodeDBs :: KademliaDHT m NodeDBs

usingReadOptions :: ReadOptions -> ASetter' NodeDBs DB -> KademliaDHT m a -> KademliaDHT m a

usingWriteOptions :: WriteOptions -> ASetter' NodeDBs DB -> KademliaDHT m a -> KademliaDHT m a

MonadDB m => MonadDB (PeerStateHolder m)

Methods

getNodeDBs :: PeerStateHolder m NodeDBs

usingReadOptions :: ReadOptions -> ASetter' NodeDBs DB -> PeerStateHolder m a -> PeerStateHolder m a

usingWriteOptions :: WriteOptions -> ASetter' NodeDBs DB -> PeerStateHolder m a -> PeerStateHolder m a

(Monad m, WithKademliaDHTInstance m) => WithKademliaDHTInstance (PeerStateHolder m)

Methods

getKademliaDHTInstance :: PeerStateHolder m KademliaDHTInstance

MonadDHT m => MonadDHT (PeerStateHolder m)

Methods

joinNetwork :: [DHTNode] -> PeerStateHolder m ()

discoverPeers :: PeerStateHolder m [DHTNode]

getKnownPeers :: PeerStateHolder m [DHTNode]

currentNodeKey :: PeerStateHolder m DHTKey

dhtLoggerName :: Proxy (* -> *) (PeerStateHolder m) -> LoggerName

MonadDelegation m => MonadDelegation (KademliaDHT m)

Methods

askDelegationState :: KademliaDHT m (TVar DelegationWrap)

MonadDelegation m => MonadDelegation (PeerStateHolder m)

Methods

askDelegationState :: PeerStateHolder m (TVar DelegationWrap)

MonadTxpMem m => MonadTxpMem (PeerStateHolder m)

Methods

askTxpMem :: PeerStateHolder m TxpLocalData

MonadUSMem m => MonadUSMem (KademliaDHT m)

Methods

askUSMemVar :: KademliaDHT m MemVar

MonadUSMem m => MonadUSMem (PeerStateHolder m)

Methods

askUSMemVar :: PeerStateHolder m MemVar

MonadJL m => MonadJL (KademliaDHT m)

Methods

jlLog :: JLEvent -> KademliaDHT m ()

MonadJL m => MonadJL (PeerStateHolder m)

Methods

jlLog :: JLEvent -> PeerStateHolder m ()