Pos.WorkMode

Description

WorkMode constraint. It is widely used in almost every our code. Simple alias for bunch of useful constraints. This module also contains new monads to extend functional capabilities inside do-block.

Documentation

type WorkMode ssc m = (MinWorkMode m, MonadMask m, MonadSlots m, MonadDB m, MonadDBLimits m, MonadTxpMem m, MonadDhtMem m, MonadRelayMem m, MonadDelegation m, MonadSscMem ssc m, MonadReportingMem m, SscGStateClass ssc, SscLocalDataClass ssc, SscHelpersClass ssc, WithNodeContext ssc m, MonadStats m, MonadJL m, WithKademliaDHTInstance m, WithPeerState m, MonadUSMem m, MonadShutdownMem m)

Bunch of constraints to perform work for real world distributed system.

type MinWorkMode m = (WithLogger m, MonadMockable m, MonadDHT m, MonadIO m, WithPeerState m)

More relaxed version of WorkMode.

Actual modes

type ProductionMode ssc = NoStatsT (RawRealMode ssc)

ProductionMode is an instance of WorkMode which is used (unsurprisingly) in production.

type RawRealMode ssc = PeerStateHolder (KademliaDHT (USHolder (DelegationT (TxpHolder (SscHolder ssc (NtpSlotting (SlottingHolder (ContextHolder ssc (DBHolder (LoggerNameBox Production))))))))))

RawRealMode is a basis for WorkModes used to really run system.

type ServiceMode = PeerStateHolder (KademliaDHT (LoggerNameBox Production))

ServiceMode is the mode in which support nodes work.

type StatsMode ssc = StatsT (RawRealMode ssc)

StatsMode is used for remote benchmarking.

Orphan instances

MonadSscMem ssc m => MonadSscMem ssc (PeerStateHolder m)

Methods

askSscMem :: PeerStateHolder m (SscState ssc)

WithNodeContext ssc m => WithNodeContext ssc (SlottingHolder m)

Methods

getNodeContext :: SlottingHolder m (NodeContext ssc)

WithNodeContext ssc m => WithNodeContext ssc (NtpSlotting m)

Methods

getNodeContext :: NtpSlotting m (NodeContext ssc)

(Monad m, WithNodeContext ssc m) => WithNodeContext ssc (PeerStateHolder m)

Methods

getNodeContext :: PeerStateHolder m (NodeContext ssc)

(Monad m, WithNodeContext ssc m) => WithNodeContext ssc (KademliaDHT m)

Methods

getNodeContext :: KademliaDHT m (NodeContext ssc)

MonadDB m => MonadDB (PeerStateHolder m)

Methods

getNodeDBs :: PeerStateHolder m NodeDBs

usingReadOptions :: ReadOptions -> ASetter' NodeDBs DB -> PeerStateHolder m a -> PeerStateHolder m a

usingWriteOptions :: WriteOptions -> ASetter' NodeDBs DB -> PeerStateHolder m a -> PeerStateHolder m a

MonadDB m => MonadDB (KademliaDHT m)

Methods

getNodeDBs :: KademliaDHT m NodeDBs

usingReadOptions :: ReadOptions -> ASetter' NodeDBs DB -> KademliaDHT m a -> KademliaDHT m a

usingWriteOptions :: WriteOptions -> ASetter' NodeDBs DB -> KademliaDHT m a -> KademliaDHT m a

MonadDB m => MonadDB (NtpSlotting m)

Methods

getNodeDBs :: NtpSlotting m NodeDBs

usingReadOptions :: ReadOptions -> ASetter' NodeDBs DB -> NtpSlotting m a -> NtpSlotting m a

usingWriteOptions :: WriteOptions -> ASetter' NodeDBs DB -> NtpSlotting m a -> NtpSlotting m a

MonadDB m => MonadDB (SlottingHolder m)

Methods

getNodeDBs :: SlottingHolder m NodeDBs

usingReadOptions :: ReadOptions -> ASetter' NodeDBs DB -> SlottingHolder m a -> SlottingHolder m a

usingWriteOptions :: WriteOptions -> ASetter' NodeDBs DB -> SlottingHolder m a -> SlottingHolder m a

MonadDBLimits m => MonadDBLimits (PeerStateHolder m)

Methods

getMaxBlockSize :: PeerStateHolder m Byte

getMaxHeaderSize :: PeerStateHolder m Byte

getMaxTxSize :: PeerStateHolder m Byte

getMaxProposalSize :: PeerStateHolder m Byte

MonadDBLimits m => MonadDBLimits (KademliaDHT m)

Methods

getMaxBlockSize :: KademliaDHT m Byte

getMaxHeaderSize :: KademliaDHT m Byte

getMaxTxSize :: KademliaDHT m Byte

getMaxProposalSize :: KademliaDHT m Byte

MonadDBLimits m => MonadDBLimits (NtpSlotting m)

Methods

getMaxBlockSize :: NtpSlotting m Byte

getMaxHeaderSize :: NtpSlotting m Byte

getMaxTxSize :: NtpSlotting m Byte

getMaxProposalSize :: NtpSlotting m Byte

MonadDBLimits m => MonadDBLimits (SlottingHolder m)

Methods

getMaxBlockSize :: SlottingHolder m Byte

getMaxHeaderSize :: SlottingHolder m Byte

getMaxTxSize :: SlottingHolder m Byte

getMaxProposalSize :: SlottingHolder m Byte

MonadRelayMem m => MonadRelayMem (PeerStateHolder m)

Methods

askRelayMem :: PeerStateHolder m RelayContext

MonadRelayMem m => MonadRelayMem (KademliaDHT m)

Methods

askRelayMem :: KademliaDHT m RelayContext

MonadRelayMem m => MonadRelayMem (NtpSlotting m)

Methods

askRelayMem :: NtpSlotting m RelayContext

MonadRelayMem m => MonadRelayMem (SlottingHolder m)

Methods

askRelayMem :: SlottingHolder m RelayContext

MonadDhtMem m => MonadDhtMem (PeerStateHolder m)

Methods

askDhtMem :: PeerStateHolder m DhtContext

MonadDhtMem m => MonadDhtMem (KademliaDHT m)

Methods

askDhtMem :: KademliaDHT m DhtContext

MonadDhtMem m => MonadDhtMem (NtpSlotting m)

Methods

askDhtMem :: NtpSlotting m DhtContext

MonadDhtMem m => MonadDhtMem (SlottingHolder m)

Methods

askDhtMem :: SlottingHolder m DhtContext

MonadDHT m => MonadDHT (PeerStateHolder m)

Methods

joinNetwork :: [DHTNode] -> PeerStateHolder m ()

discoverPeers :: PeerStateHolder m [DHTNode]

getKnownPeers :: PeerStateHolder m [DHTNode]

currentNodeKey :: PeerStateHolder m DHTKey

dhtLoggerName :: Proxy (* -> *) (PeerStateHolder m) -> LoggerName

(Monad m, WithKademliaDHTInstance m) => WithKademliaDHTInstance (PeerStateHolder m)

Methods

getKademliaDHTInstance :: PeerStateHolder m KademliaDHTInstance

MonadReportingMem m => MonadReportingMem (PeerStateHolder m)

Methods

askReportingMem :: PeerStateHolder m ReportingContext

MonadReportingMem m => MonadReportingMem (KademliaDHT m)

Methods

askReportingMem :: KademliaDHT m ReportingContext

MonadReportingMem m => MonadReportingMem (NtpSlotting m)

Methods

askReportingMem :: NtpSlotting m ReportingContext

MonadReportingMem m => MonadReportingMem (SlottingHolder m)

Methods

askReportingMem :: SlottingHolder m ReportingContext

MonadShutdownMem m => MonadShutdownMem (PeerStateHolder m)

Methods

askShutdownMem :: PeerStateHolder m ShutdownContext

MonadShutdownMem m => MonadShutdownMem (KademliaDHT m)

Methods

askShutdownMem :: KademliaDHT m ShutdownContext

MonadShutdownMem m => MonadShutdownMem (NtpSlotting m)

Methods

askShutdownMem :: NtpSlotting m ShutdownContext

MonadShutdownMem m => MonadShutdownMem (SlottingHolder m)

Methods

askShutdownMem :: SlottingHolder m ShutdownContext

MonadSlots m => MonadSlots (PeerStateHolder m)

Methods

getCurrentSlot :: PeerStateHolder m (Maybe SlotId)

getCurrentSlotBlocking :: PeerStateHolder m SlotId

getCurrentSlotInaccurate :: PeerStateHolder m SlotId

currentTimeSlotting :: PeerStateHolder m Timestamp

slottingWorkers :: [PeerStateHolder m ()]

MonadSlots m => MonadSlots (KademliaDHT m)

Methods

getCurrentSlot :: KademliaDHT m (Maybe SlotId)

getCurrentSlotBlocking :: KademliaDHT m SlotId

getCurrentSlotInaccurate :: KademliaDHT m SlotId

currentTimeSlotting :: KademliaDHT m Timestamp

slottingWorkers :: [KademliaDHT m ()]

MonadSlotsData m => MonadSlotsData (PeerStateHolder m)

Methods

getSystemStart :: PeerStateHolder m Timestamp

getSlottingData :: PeerStateHolder m SlottingData

waitPenultEpochEquals :: EpochIndex -> PeerStateHolder m ()

putSlottingData :: SlottingData -> PeerStateHolder m ()

MonadSlotsData m => MonadSlotsData (KademliaDHT m)

Methods

getSystemStart :: KademliaDHT m Timestamp

getSlottingData :: KademliaDHT m SlottingData

waitPenultEpochEquals :: EpochIndex -> KademliaDHT m ()

putSlottingData :: SlottingData -> KademliaDHT m ()

MonadDelegation m => MonadDelegation (PeerStateHolder m)

Methods

askDelegationState :: PeerStateHolder m (TVar DelegationWrap)

MonadDelegation m => MonadDelegation (KademliaDHT m)

Methods

askDelegationState :: KademliaDHT m (TVar DelegationWrap)

MonadTxpMem m => MonadTxpMem (PeerStateHolder m)

Methods

askTxpMem :: PeerStateHolder m TxpLocalData

MonadUSMem m => MonadUSMem (PeerStateHolder m)

Methods

askUSMemVar :: PeerStateHolder m MemVar

MonadUSMem m => MonadUSMem (KademliaDHT m)

Methods

askUSMemVar :: KademliaDHT m MemVar

MonadJL m => MonadJL (PeerStateHolder m)

Methods

jlLog :: JLEvent -> PeerStateHolder m ()

MonadJL m => MonadJL (KademliaDHT m)

Methods

jlLog :: JLEvent -> KademliaDHT m ()

MonadJL m => MonadJL (NtpSlotting m)

Methods

jlLog :: JLEvent -> NtpSlotting m ()

MonadJL m => MonadJL (SlottingHolder m)

Methods

jlLog :: JLEvent -> SlottingHolder m ()