Pos.WorkMode

Description

WorkMode constraint. It is widely used in almost every our code. Simple alias for bunch of useful constraints. This module also contains new monads to extend functional capabilities inside do-block.

Documentation

type WorkMode ssc m = (MinWorkMode m, MonadMask m, MonadSlots m, MonadDB ssc m, MonadDBLimits m, MonadTxpLD ssc m, MonadDelegation m, MonadUtxo m, MonadSscMem ssc m, SscGStateClass ssc, SscLocalDataClass ssc, SscHelpersClass ssc, WithNodeContext ssc m, MonadStats m, MonadJL m, WithKademliaDHTInstance m, WithPeerState m, MonadUSMem m)

Bunch of constraints to perform work for real world distributed system.

type MinWorkMode m = (WithLogger m, MonadMockable m, MonadDHT m, MonadIO m, WithPeerState m)

More relaxed version of WorkMode.

Actual modes

type ProductionMode ssc = NoStatsT (RawRealMode ssc)

ProductionMode is an instance of WorkMode which is used (unsurprisingly) in production.

type RawRealMode ssc = PeerStateHolder (KademliaDHT (USHolder (DelegationT (TxpLDHolder ssc (SscHolder ssc (ContextHolder ssc (DBHolder ssc (LoggerNameBox Production))))))))

RawRealMode is a basis for WorkModes used to really run system.

type ServiceMode = PeerStateHolder (KademliaDHT (LoggerNameBox Production))

ServiceMode is the mode in which support nodes work.

type StatsMode ssc = StatsT (RawRealMode ssc)

StatsMode is used for remote benchmarking.

Orphan instances

MonadSscMem ssc m => MonadSscMem ssc (PeerStateHolder m)

Methods

askSscMem :: PeerStateHolder m (SscState ssc)

MonadDB ssc m => MonadDB ssc (PeerStateHolder m)

Methods

getNodeDBs :: PeerStateHolder m (NodeDBs ssc)

usingReadOptions :: ReadOptions -> ASetter' (NodeDBs ssc) (DB ssc) -> PeerStateHolder m a -> PeerStateHolder m a

usingWriteOptions :: WriteOptions -> ASetter' (NodeDBs ssc) (DB ssc) -> PeerStateHolder m a -> PeerStateHolder m a

MonadDB ssc m => MonadDB ssc (KademliaDHT m)

Methods

getNodeDBs :: KademliaDHT m (NodeDBs ssc)

usingReadOptions :: ReadOptions -> ASetter' (NodeDBs ssc) (DB ssc) -> KademliaDHT m a -> KademliaDHT m a

usingWriteOptions :: WriteOptions -> ASetter' (NodeDBs ssc) (DB ssc) -> KademliaDHT m a -> KademliaDHT m a

(Monad m, WithNodeContext ssc m) => WithNodeContext ssc (PeerStateHolder m)

Methods

getNodeContext :: PeerStateHolder m (NodeContext ssc)

(Monad m, WithNodeContext ssc m) => WithNodeContext ssc (KademliaDHT m)

Methods

getNodeContext :: KademliaDHT m (NodeContext ssc)

MonadTxpLD ssc m => MonadTxpLD ssc (PeerStateHolder m)

Methods

getTxpLDWrap :: PeerStateHolder m (TxpLDWrap ssc)

setUtxoView :: UtxoView ssc -> PeerStateHolder m ()

setMemPool :: MemPool -> PeerStateHolder m ()

modifyTxpLD :: (TxpLD ssc -> (a, TxpLD ssc)) -> PeerStateHolder m a

modifyTxpLD_ :: (TxpLD ssc -> TxpLD ssc) -> PeerStateHolder m ()

getTxpLD :: PeerStateHolder m (TxpLD ssc)

setTxpLD :: TxpLD ssc -> PeerStateHolder m ()

MonadDHT m => MonadDHT (PeerStateHolder m)

Methods

joinNetwork :: [DHTNode] -> PeerStateHolder m ()

discoverPeers :: PeerStateHolder m [DHTNode]

getKnownPeers :: PeerStateHolder m [DHTNode]

currentNodeKey :: PeerStateHolder m DHTKey

dhtLoggerName :: Proxy (* -> *) (PeerStateHolder m) -> LoggerName

(Monad m, WithKademliaDHTInstance m) => WithKademliaDHTInstance (PeerStateHolder m)

Methods

getKademliaDHTInstance :: PeerStateHolder m KademliaDHTInstance

MonadUtxo m => MonadUtxo (KademliaDHT m)

Methods

utxoPut :: TxIn -> TxOutAux -> KademliaDHT m ()

utxoDel :: TxIn -> KademliaDHT m ()

MonadUtxo m => MonadUtxo (PeerStateHolder m)

Methods

utxoPut :: TxIn -> TxOutAux -> PeerStateHolder m ()

utxoDel :: TxIn -> PeerStateHolder m ()

MonadUtxoRead m => MonadUtxoRead (KademliaDHT m)

Methods

utxoGet :: TxIn -> KademliaDHT m (Maybe TxOutAux)

MonadUtxoRead m => MonadUtxoRead (PeerStateHolder m)

Methods

utxoGet :: TxIn -> PeerStateHolder m (Maybe TxOutAux)

MonadDelegation m => MonadDelegation (KademliaDHT m)

Methods

askDelegationState :: KademliaDHT m (TVar DelegationWrap)

MonadDelegation m => MonadDelegation (PeerStateHolder m)

Methods

askDelegationState :: PeerStateHolder m (TVar DelegationWrap)

MonadSlots m => MonadSlots (PeerStateHolder m)

Methods

getSystemStartTime :: PeerStateHolder m Timestamp

getCurrentTime :: PeerStateHolder m Timestamp

getCurrentSlot :: PeerStateHolder m SlotId

getSlotDuration :: PeerStateHolder m Millisecond

MonadDBLimits m => MonadDBLimits (KademliaDHT m)

Methods

getMaxBlockSize :: KademliaDHT m Byte

MonadDBLimits m => MonadDBLimits (PeerStateHolder m)

Methods

getMaxBlockSize :: PeerStateHolder m Byte

MonadUSMem m => MonadUSMem (KademliaDHT m)

Methods

askUSMemVar :: KademliaDHT m MemVar

MonadUSMem m => MonadUSMem (PeerStateHolder m)

Methods

askUSMemVar :: PeerStateHolder m MemVar

MonadJL m => MonadJL (KademliaDHT m)

Methods

jlLog :: JLEvent -> KademliaDHT m ()

MonadJL m => MonadJL (PeerStateHolder m)

Methods

jlLog :: JLEvent -> PeerStateHolder m ()